Move remaining components to packages folder
