--- src/components/common/icons.js
@@ -1,578 +1,0 @@
-const icons = {};
-icons.userinfo = <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 216 216"
-width="20px" height="20px"
->
-<path
-  fill="#0073e6"
-  fillRule="nonzero"
-  d="M197.29 56.933H18.71c-1.92 0-3.5 1.58-3.5 3.5v95.134c0 1.92 1.58 3.5 3.5 3.5h178.58c1.92 0 3.5-1.58 3.5-3.5V60.413a3.517 3.517 0 00-3.5-3.48zm-3.5 95.154H22.21V63.933h171.58v88.154z"
-></path>
-<path
-  fill="#0073e6"
-  fillRule="nonzero"
-  d="M115.08 100.885h57c1.92 0 3.5-1.58 3.5-3.5s-1.58-3.5-3.5-3.5h-57c-1.92 0-3.5 1.58-3.5 3.5s1.58 3.5 3.5 3.5zM115.08 122.115h57c1.92 0 3.5-1.58 3.5-3.5s-1.58-3.5-3.5-3.5h-57c-1.92 0-3.5 1.58-3.5 3.5s1.58 3.5 3.5 3.5z"
-></path>
-<path
-  fill="#0073e6"
-  d="M64.703 73.164c19.226 0 34.835 15.61 34.835 34.836s-15.609 34.836-34.835 34.836c-19.227 0-34.836-15.61-34.836-34.836s15.609-34.836 34.836-34.836zm-.386 39.457l7.203-4.82c6.329 0 11.969 6.17 11.969 11.996v6.594s-9.665 2.992-19.172 2.992c-9.665 0-19.172-2.992-19.172-2.992v-6.594c0-5.958 5.138-11.863 11.837-11.863l7.335 4.687zm0-6.806c-6.011 0-7.229-9.109-7.229-9.109-.715-4.74 1.457-10.089 7.15-10.089 5.72 0 7.891 5.349 7.176 10.089 0 0-1.086 9.109-7.097 9.109z"
-></path>
-</svg>;
-icons.aligntop = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fill="currentColor" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M17.294,17.287l-14.588,0l0,-14.574l14.588,0c0,4.858 0,9.716 0,14.574Zm-13.738,-0.85l12.888,0l0,-12.874l-12.888,0c0,4.291 0,8.583 0,12.874Z" fillRule="nonzero" />
-	<rect x="4.489" y="4.545" width="11.022" height="2.512" />
-</svg>;
-icons.alignmiddle = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fill="currentColor" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M17.294,17.287l-14.588,0l0,-14.574l14.588,0c0,4.858 0,9.716 0,14.574Zm-13.738,-0.85l12.888,0l0,-12.874l-12.888,0c0,4.291 0,8.583 0,12.874Z" fillRule="nonzero" />
-	<rect x="4.489" y="8.744" width="11.022" height="2.512" />
-</svg>;
-icons.alignbottom = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fill="currentColor" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M17.294,17.287l-14.588,0l0,-14.574l14.588,0c0,4.858 0,9.716 0,14.574Zm-13.738,-0.85l12.888,0l0,-12.874l-12.888,0c0,4.291 0,8.583 0,12.874Z" fillRule="nonzero" />
-	<rect x="4.489" y="12.802" width="11.022" height="2.512" />
-</svg>;
-icons.outlinetop = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.714" y="5.492" width="1.048" height="9.017" fill="#555d66" />
-	<rect x="16.265" y="5.498" width="1.023" height="9.003" fill="#555d66" />
-	<rect x="5.518" y="2.186" width="8.964" height="2.482" fill="#272b2f" />
-	<rect x="5.487" y="16.261" width="9.026" height="1.037" fill="#555d66" />
-</svg>;
-icons.outlineright = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.714" y="5.492" width="1.046" height="9.017" fill="#555d66" />
-	<rect x="15.244" y="5.498" width="2.518" height="9.003" fill="#272b2f" />
-	<rect x="5.518" y="2.719" width="8.964" height="0.954" fill="#555d66" />
-	<rect x="5.487" y="16.308" width="9.026" height="0.99" fill="#555d66" />
-</svg>;
-icons.outlinebottom = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.714" y="5.492" width="1" height="9.017" fill="#555d66" />
-	<rect x="16.261" y="5.498" width="1.027" height="9.003" fill="#555d66" />
-	<rect x="5.518" y="2.719" width="8.964" height="0.968" fill="#555d66" />
-	<rect x="5.487" y="15.28" width="9.026" height="2.499" fill="#272b2f" />
-</svg>;
-icons.outlineleft = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.202" y="5.492" width="2.503" height="9.017" fill="#272b2f" />
-	<rect x="16.276" y="5.498" width="1.012" height="9.003" fill="#555d66" />
-	<rect x="5.518" y="2.719" width="8.964" height="0.966" fill="#555d66" />
-	<rect x="5.487" y="16.303" width="9.026" height="0.995" fill="#555d66" />
-</svg>;
-icons.linked = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd' clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-	<path d='M17.888,17.895l-15.776,0l0,-15.79l15.776,0l0,15.79Zm-13.276,-13.29l0,10.79l10.776,0l0,-10.79l-10.776,0Z' />
-</svg>;
-icons.individual = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd' clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-	<rect x='2.024' y='5.486' width='2.61' height='9.027' />
-	<rect x='15.379' y='5.497' width='2.56' height='9.007' />
-	<rect x='5.509' y='1.999' width='8.983' height='2.582' />
-	<rect x='5.51' y='15.36' width='8.979' height='2.563' />
-</svg>;
-icons.radiuslinked = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M14.044,2.12l0.101,0.003l0.101,0.007l0.099,0.009l0.099,0.011l0.098,0.014l0.098,0.016l0.096,0.018l0.096,0.021l0.095,0.023l0.094,0.026l0.093,0.028l0.092,0.03l0.091,0.032l0.09,0.034l0.089,0.036l0.088,0.039l0.086,0.04l0.086,0.043l0.084,0.044l0.083,0.047l0.082,0.048l0.081,0.051l0.079,0.052l0.078,0.054l0.077,0.056l0.075,0.057l0.074,0.06l0.072,0.061l0.071,0.063l0.069,0.064l0.068,0.066l0.066,0.068l0.064,0.069l0.063,0.071l0.061,0.072l0.059,0.074l0.058,0.075l0.056,0.076l0.054,0.078l0.052,0.08l0.05,0.08l0.049,0.082l0.046,0.083l0.045,0.085l0.042,0.085l0.041,0.087l0.038,0.088l0.037,0.089l0.034,0.09l0.032,0.091l0.03,0.092l0.028,0.093l0.025,0.094l0.023,0.094l0.021,0.096l0.019,0.097l0.016,0.097l0.014,0.098l0.011,0.099l0.009,0.1l0.006,0.1l0.004,0.101l0.001,0.102l0,7.881l-0.001,0.102l-0.004,0.101l-0.006,0.1l-0.009,0.1l-0.011,0.098l-0.014,0.099l-0.016,0.097l-0.019,0.097l-0.021,0.095l-0.023,0.095l-0.025,0.094l-0.028,0.093l-0.03,0.092l-0.032,0.091l-0.034,0.09l-0.037,0.089l-0.038,0.088l-0.041,0.087l-0.042,0.085l-0.045,0.085l-0.046,0.083l-0.049,0.082l-0.05,0.08l-0.052,0.079l-0.054,0.078l-0.056,0.077l-0.058,0.075l-0.059,0.074l-0.061,0.072l-0.063,0.071l-0.064,0.069l-0.066,0.068l-0.068,0.066l-0.069,0.064l-0.071,0.063l-0.072,0.061l-0.074,0.06l-0.075,0.057l-0.077,0.056l-0.078,0.054l-0.079,0.052l-0.081,0.051l-0.082,0.048l-0.083,0.047l-0.084,0.044l-0.086,0.043l-0.086,0.04l-0.088,0.039l-0.089,0.036l-0.09,0.034l-0.091,0.032l-0.092,0.03l-0.093,0.028l-0.094,0.025l-0.095,0.023l-0.096,0.021l-0.096,0.019l-0.098,0.016l-0.098,0.014l-0.099,0.011l-0.099,0.009l-0.101,0.006l-0.101,0.004l-0.101,0.001l-7.905,0l-0.101,-0.001l-0.101,-0.004l-0.101,-0.006l-0.099,-0.009l-0.099,-0.011l-0.098,-0.014l-0.098,-0.016l-0.096,-0.019l-0.096,-0.021l-0.095,-0.023l-0.094,-0.025l-0.093,-0.028l-0.092,-0.03l-0.091,-0.032l-0.09,-0.034l-0.089,-0.036l-0.088,-0.039l-0.086,-0.04l-0.086,-0.043l-0.084,-0.044l-0.083,-0.047l-0.082,-0.048l-0.081,-0.051l-0.079,-0.052l-0.078,-0.054l-0.077,-0.056l-0.075,-0.057l-0.074,-0.06l-0.072,-0.061l-0.071,-0.063l-0.069,-0.064l-0.067,-0.066l-0.067,-0.068l-0.064,-0.069l-0.063,-0.071l-0.061,-0.072l-0.059,-0.074l-0.058,-0.075l-0.056,-0.077l-0.054,-0.078l-0.052,-0.079l-0.05,-0.08l-0.049,-0.082l-0.046,-0.083l-0.045,-0.085l-0.042,-0.085l-0.041,-0.087l-0.038,-0.088l-0.037,-0.089l-0.034,-0.09l-0.032,-0.091l-0.03,-0.092l-0.028,-0.093l-0.025,-0.094l-0.023,-0.095l-0.021,-0.095l-0.019,-0.097l-0.016,-0.097l-0.014,-0.099l-0.011,-0.098l-0.009,-0.1l-0.006,-0.1l-0.004,-0.101l-0.001,-0.102l0,-7.881l0.001,-0.102l0.004,-0.101l0.006,-0.1l0.009,-0.1l0.011,-0.099l0.014,-0.098l0.016,-0.097l0.019,-0.097l0.021,-0.096l0.023,-0.094l0.025,-0.094l0.028,-0.093l0.03,-0.092l0.032,-0.091l0.034,-0.09l0.037,-0.089l0.038,-0.088l0.041,-0.087l0.042,-0.085l0.045,-0.085l0.046,-0.083l0.049,-0.082l0.05,-0.08l0.052,-0.08l0.054,-0.078l0.056,-0.076l0.058,-0.075l0.059,-0.074l0.061,-0.072l0.063,-0.071l0.064,-0.069l0.067,-0.068l0.067,-0.066l0.069,-0.064l0.071,-0.063l0.072,-0.061l0.074,-0.06l0.075,-0.057l0.077,-0.056l0.078,-0.054l0.079,-0.052l0.081,-0.051l0.082,-0.048l0.083,-0.047l0.084,-0.044l0.086,-0.043l0.086,-0.04l0.088,-0.039l0.089,-0.036l0.09,-0.034l0.091,-0.032l0.092,-0.03l0.093,-0.028l0.094,-0.026l0.095,-0.023l0.096,-0.021l0.096,-0.018l0.098,-0.016l0.098,-0.014l0.099,-0.011l0.099,-0.009l0.101,-0.007l0.101,-0.003l0.101,-0.002l7.905,0l0.101,0.002Zm-7.99,2.498l-0.054,0.001l-0.038,0.001l-0.037,0.003l-0.036,0.003l-0.037,0.004l-0.035,0.005l-0.036,0.006l-0.035,0.007l-0.034,0.007l-0.034,0.009l-0.034,0.009l-0.034,0.01l-0.033,0.011l-0.033,0.011l-0.032,0.013l-0.032,0.013l-0.032,0.014l-0.032,0.014l-0.031,0.016l-0.03,0.016l-0.031,0.017l-0.03,0.018l-0.029,0.018l-0.029,0.019l-0.029,0.02l-0.028,0.021l-0.028,0.021l-0.027,0.022l-0.027,0.022l-0.026,0.024l-0.026,0.023l-0.025,0.025l-0.024,0.025l-0.024,0.026l-0.023,0.026l-0.023,0.026l-0.022,0.028l-0.021,0.027l-0.02,0.029l-0.02,0.028l-0.019,0.029l-0.019,0.03l-0.018,0.03l-0.017,0.03l-0.016,0.031l-0.015,0.031l-0.015,0.031l-0.014,0.032l-0.013,0.032l-0.012,0.033l-0.012,0.033l-0.011,0.033l-0.01,0.033l-0.009,0.034l-0.008,0.034l-0.008,0.035l-0.006,0.035l-0.006,0.035l-0.005,0.036l-0.004,0.036l-0.004,0.037l-0.002,0.037l-0.002,0.038l0,0.054l0,7.849l0,0.054l0.002,0.038l0.002,0.037l0.004,0.037l0.004,0.036l0.005,0.036l0.006,0.035l0.006,0.035l0.008,0.034l0.008,0.035l0.009,0.034l0.01,0.033l0.011,0.033l0.012,0.033l0.012,0.032l0.013,0.033l0.014,0.031l0.015,0.032l0.015,0.031l0.016,0.031l0.017,0.03l0.018,0.03l0.019,0.029l0.019,0.03l0.02,0.028l0.02,0.028l0.021,0.028l0.022,0.027l0.023,0.027l0.023,0.026l0.024,0.026l0.024,0.025l0.025,0.024l0.026,0.024l0.026,0.023l0.027,0.023l0.027,0.022l0.028,0.021l0.028,0.021l0.029,0.02l0.029,0.019l0.029,0.018l0.03,0.018l0.031,0.017l0.03,0.016l0.031,0.016l0.032,0.014l0.032,0.014l0.032,0.013l0.032,0.013l0.033,0.011l0.033,0.011l0.034,0.01l0.034,0.009l0.034,0.008l0.034,0.008l0.035,0.007l0.036,0.006l0.035,0.005l0.037,0.004l0.036,0.003l0.037,0.002l0.038,0.002l0.054,0l7.873,0l0.054,0l0.038,-0.002l0.037,-0.002l0.036,-0.003l0.037,-0.004l0.035,-0.005l0.036,-0.006l0.035,-0.007l0.034,-0.008l0.034,-0.008l0.034,-0.009l0.034,-0.01l0.033,-0.011l0.033,-0.011l0.032,-0.013l0.032,-0.013l0.032,-0.014l0.032,-0.014l0.031,-0.016l0.03,-0.016l0.031,-0.017l0.03,-0.018l0.029,-0.018l0.029,-0.019l0.029,-0.02l0.028,-0.021l0.028,-0.021l0.027,-0.022l0.027,-0.023l0.026,-0.023l0.026,-0.024l0.025,-0.024l0.024,-0.025l0.024,-0.026l0.023,-0.026l0.023,-0.027l0.022,-0.027l0.021,-0.028l0.021,-0.028l0.019,-0.028l0.02,-0.03l0.018,-0.029l0.018,-0.03l0.017,-0.03l0.016,-0.031l0.015,-0.031l0.015,-0.032l0.014,-0.031l0.013,-0.033l0.012,-0.032l0.012,-0.033l0.011,-0.033l0.01,-0.033l0.009,-0.034l0.008,-0.035l0.008,-0.034l0.006,-0.035l0.006,-0.035l0.005,-0.036l0.004,-0.036l0.004,-0.037l0.002,-0.037l0.002,-0.038l0,-0.054l0,-7.849l0,-0.054l-0.002,-0.038l-0.002,-0.037l-0.004,-0.037l-0.004,-0.036l-0.005,-0.036l-0.006,-0.035l-0.006,-0.035l-0.008,-0.035l-0.008,-0.034l-0.009,-0.034l-0.01,-0.033l-0.011,-0.033l-0.012,-0.033l-0.012,-0.033l-0.013,-0.032l-0.014,-0.032l-0.015,-0.031l-0.015,-0.031l-0.016,-0.031l-0.017,-0.03l-0.018,-0.03l-0.018,-0.03l-0.02,-0.029l-0.019,-0.028l-0.021,-0.029l-0.021,-0.027l-0.022,-0.028l-0.023,-0.026l-0.023,-0.026l-0.024,-0.026l-0.024,-0.025l-0.025,-0.025l-0.026,-0.023l-0.026,-0.024l-0.027,-0.022l-0.027,-0.022l-0.028,-0.021l-0.028,-0.021l-0.029,-0.02l-0.029,-0.019l-0.029,-0.018l-0.03,-0.018l-0.031,-0.017l-0.03,-0.016l-0.031,-0.016l-0.032,-0.014l-0.032,-0.014l-0.032,-0.013l-0.032,-0.013l-0.033,-0.011l-0.033,-0.011l-0.034,-0.01l-0.033,-0.009l-0.035,-0.009l-0.034,-0.007l-0.035,-0.007l-0.036,-0.006l-0.035,-0.005l-0.037,-0.004l-0.036,-0.003l-0.037,-0.003l-0.038,-0.001l-0.054,-0.001l-7.873,0Z" fill="#555d66" />
-</svg>;
-icons.radiusindividual = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M14.044,2.12l0.101,0.003l0.101,0.007l0.099,0.009l0.099,0.011l0.098,0.014l0.098,0.016l0.096,0.018l0.096,0.021l0.095,0.023l0.094,0.026l0.093,0.028l0.092,0.03l0.091,0.032l0.09,0.034l0.089,0.036l0.088,0.039l0.086,0.04l0.086,0.043l0.084,0.044l0.083,0.047l0.082,0.048l0.081,0.051l0.079,0.052l0.078,0.054l0.077,0.056l0.075,0.057l0.074,0.06l0.072,0.061l0.071,0.063l0.069,0.064l0.068,0.066l0.066,0.068l0.064,0.069l0.063,0.071l0.061,0.072l0.059,0.074l0.058,0.075l0.056,0.076l0.054,0.078l0.052,0.08l0.05,0.08l0.049,0.082l0.046,0.083l0.045,0.085l0.042,0.085l0.041,0.087l0.038,0.088l0.037,0.089l0.034,0.09l0.032,0.091l0.03,0.092l0.028,0.093l0.025,0.094l0.023,0.094l0.021,0.096l0.019,0.097l0.016,0.097l0.014,0.098l0.011,0.099l0.009,0.1l0.006,0.1l0.004,0.101l0.001,0.102l0,1.273l-5.531,-5.214l1.591,0l0.101,0.002Zm-1.692,-0.002l5.531,5.214l-2.5,-0.012l0,-1.245l0,-0.054l-0.002,-0.038l-0.002,-0.037l-0.004,-0.037l-0.004,-0.036l-0.005,-0.036l-0.006,-0.035l-0.006,-0.035l-0.008,-0.035l-0.008,-0.034l-0.009,-0.034l-0.01,-0.033l-0.011,-0.033l-0.012,-0.033l-0.012,-0.033l-0.013,-0.032l-0.014,-0.032l-0.015,-0.031l-0.015,-0.031l-0.016,-0.031l-0.017,-0.03l-0.018,-0.03l-0.018,-0.03l-0.02,-0.029l-0.019,-0.028l-0.021,-0.029l-0.021,-0.027l-0.022,-0.028l-0.023,-0.026l-0.023,-0.026l-0.024,-0.026l-0.024,-0.025l-0.025,-0.025l-0.026,-0.023l-0.026,-0.024l-0.027,-0.022l-0.027,-0.022l-0.028,-0.021l-0.028,-0.021l-0.029,-0.02l-0.029,-0.019l-0.029,-0.018l-0.03,-0.018l-0.031,-0.017l-0.03,-0.016l-0.031,-0.016l-0.032,-0.014l-0.032,-0.014l-0.032,-0.013l-0.032,-0.013l-0.033,-0.011l-0.033,-0.011l-0.034,-0.01l-0.033,-0.009l-0.035,-0.009l-0.034,-0.007l-0.035,-0.007l-0.036,-0.006l-0.035,-0.005l-0.037,-0.004l-0.036,-0.003l-0.037,-0.003l-0.038,-0.001l-0.054,-0.001l-1.57,0l-0.005,-2.5Z" fill="#555d66" />
-	<path d="M14.044,17.873l0.101,-0.003l0.101,-0.007l0.099,-0.009l0.099,-0.011l0.098,-0.014l0.098,-0.016l0.096,-0.018l0.096,-0.021l0.095,-0.023l0.094,-0.026l0.093,-0.028l0.092,-0.03l0.091,-0.032l0.09,-0.034l0.089,-0.036l0.088,-0.039l0.086,-0.04l0.086,-0.043l0.084,-0.044l0.083,-0.047l0.082,-0.048l0.081,-0.051l0.079,-0.052l0.078,-0.054l0.077,-0.056l0.075,-0.057l0.074,-0.06l0.072,-0.061l0.071,-0.063l0.069,-0.064l0.068,-0.066l0.066,-0.068l0.064,-0.069l0.063,-0.071l0.061,-0.072l0.059,-0.074l0.058,-0.075l0.056,-0.076l0.054,-0.078l0.052,-0.08l0.05,-0.08l0.049,-0.082l0.046,-0.083l0.045,-0.085l0.042,-0.085l0.041,-0.087l0.038,-0.088l0.037,-0.089l0.034,-0.09l0.032,-0.091l0.03,-0.092l0.028,-0.093l0.025,-0.094l0.023,-0.094l0.021,-0.096l0.019,-0.097l0.016,-0.097l0.014,-0.098l0.011,-0.099l0.009,-0.1l0.006,-0.1l0.004,-0.101l0.001,-0.102l0,-1.272l-5.531,5.213l1.591,0l0.101,-0.002Zm-1.692,0.002l5.531,-5.213l-2.5,0.012l0,1.244l0,0.054l-0.002,0.038l-0.002,0.037l-0.004,0.037l-0.004,0.036l-0.005,0.036l-0.006,0.035l-0.006,0.035l-0.008,0.035l-0.008,0.034l-0.009,0.034l-0.01,0.033l-0.011,0.033l-0.012,0.033l-0.012,0.033l-0.013,0.032l-0.014,0.032l-0.015,0.031l-0.015,0.031l-0.016,0.031l-0.017,0.03l-0.018,0.03l-0.018,0.03l-0.02,0.029l-0.019,0.028l-0.021,0.029l-0.021,0.027l-0.022,0.028l-0.023,0.026l-0.023,0.026l-0.024,0.026l-0.024,0.025l-0.025,0.025l-0.026,0.023l-0.026,0.024l-0.027,0.022l-0.027,0.022l-0.028,0.021l-0.028,0.021l-0.029,0.02l-0.029,0.019l-0.029,0.018l-0.03,0.018l-0.031,0.017l-0.03,0.016l-0.031,0.016l-0.032,0.014l-0.032,0.014l-0.032,0.013l-0.032,0.013l-0.033,0.011l-0.033,0.011l-0.034,0.01l-0.033,0.009l-0.035,0.009l-0.034,0.007l-0.035,0.007l-0.036,0.006l-0.035,0.005l-0.037,0.004l-0.036,0.003l-0.037,0.003l-0.038,0.001l-0.054,0.001l-1.57,0l-0.005,2.5Z" fill="#555d66" />
-	<path d="M6.077,17.873l-0.101,-0.003l-0.101,-0.007l-0.099,-0.009l-0.099,-0.011l-0.098,-0.014l-0.098,-0.016l-0.096,-0.018l-0.096,-0.021l-0.095,-0.023l-0.094,-0.026l-0.093,-0.028l-0.092,-0.03l-0.091,-0.032l-0.09,-0.034l-0.089,-0.036l-0.088,-0.039l-0.086,-0.04l-0.086,-0.043l-0.084,-0.044l-0.083,-0.047l-0.082,-0.048l-0.081,-0.051l-0.079,-0.052l-0.078,-0.054l-0.077,-0.056l-0.075,-0.057l-0.074,-0.06l-0.072,-0.061l-0.071,-0.063l-0.069,-0.064l-0.068,-0.066l-0.066,-0.068l-0.064,-0.069l-0.063,-0.071l-0.061,-0.072l-0.059,-0.074l-0.058,-0.075l-0.056,-0.076l-0.054,-0.078l-0.052,-0.08l-0.05,-0.08l-0.049,-0.082l-0.046,-0.083l-0.045,-0.085l-0.042,-0.085l-0.041,-0.087l-0.038,-0.088l-0.037,-0.089l-0.034,-0.09l-0.032,-0.091l-0.03,-0.092l-0.028,-0.093l-0.025,-0.094l-0.023,-0.094l-0.021,-0.096l-0.019,-0.097l-0.016,-0.097l-0.014,-0.098l-0.011,-0.099l-0.009,-0.1l-0.006,-0.1l-0.004,-0.101l-0.001,-0.102l0,-1.272l5.531,5.213l-1.591,0l-0.101,-0.002Zm1.692,0.002l-5.531,-5.213l2.5,0.012l0,1.244l0,0.054l0.002,0.038l0.002,0.037l0.004,0.037l0.004,0.036l0.005,0.036l0.006,0.035l0.006,0.035l0.008,0.035l0.008,0.034l0.009,0.034l0.01,0.033l0.011,0.033l0.012,0.033l0.012,0.033l0.013,0.032l0.014,0.032l0.015,0.031l0.015,0.031l0.016,0.031l0.017,0.03l0.018,0.03l0.018,0.03l0.02,0.029l0.019,0.028l0.021,0.029l0.021,0.027l0.022,0.028l0.023,0.026l0.023,0.026l0.024,0.026l0.024,0.025l0.025,0.025l0.026,0.023l0.026,0.024l0.027,0.022l0.027,0.022l0.028,0.021l0.028,0.021l0.029,0.02l0.029,0.019l0.029,0.018l0.03,0.018l0.031,0.017l0.03,0.016l0.031,0.016l0.032,0.014l0.032,0.014l0.032,0.013l0.032,0.013l0.033,0.011l0.033,0.011l0.034,0.01l0.033,0.009l0.035,0.009l0.034,0.007l0.035,0.007l0.036,0.006l0.035,0.005l0.037,0.004l0.036,0.003l0.037,0.003l0.038,0.001l0.054,0.001l1.569,0l0.006,2.5Z" fill="#555d66" />
-	<path d="M6.077,2.124l-0.101,0.003l-0.101,0.007l-0.099,0.009l-0.099,0.011l-0.098,0.014l-0.098,0.016l-0.096,0.018l-0.096,0.021l-0.095,0.023l-0.094,0.026l-0.093,0.028l-0.092,0.03l-0.091,0.032l-0.09,0.034l-0.089,0.036l-0.088,0.039l-0.086,0.04l-0.086,0.043l-0.084,0.044l-0.083,0.047l-0.082,0.048l-0.081,0.051l-0.079,0.052l-0.078,0.054l-0.077,0.056l-0.075,0.057l-0.074,0.06l-0.072,0.061l-0.071,0.063l-0.069,0.064l-0.068,0.066l-0.066,0.068l-0.064,0.069l-0.063,0.071l-0.061,0.072l-0.059,0.074l-0.058,0.075l-0.056,0.076l-0.054,0.078l-0.052,0.08l-0.05,0.08l-0.049,0.082l-0.046,0.083l-0.045,0.085l-0.042,0.085l-0.041,0.087l-0.038,0.088l-0.037,0.089l-0.034,0.09l-0.032,0.091l-0.03,0.092l-0.028,0.093l-0.025,0.094l-0.023,0.094l-0.021,0.096l-0.019,0.097l-0.016,0.097l-0.014,0.098l-0.011,0.099l-0.009,0.1l-0.006,0.1l-0.004,0.101l-0.001,0.102l0,1.272l5.531,-5.213l-1.591,0l-0.101,0.002Zm1.692,-0.002l-5.531,5.213l2.5,-0.011l0,-1.245l0,-0.054l0.002,-0.038l0.002,-0.037l0.004,-0.037l0.004,-0.036l0.005,-0.036l0.006,-0.035l0.006,-0.035l0.008,-0.035l0.008,-0.034l0.009,-0.034l0.01,-0.033l0.011,-0.033l0.012,-0.033l0.012,-0.033l0.013,-0.032l0.014,-0.032l0.015,-0.031l0.015,-0.031l0.016,-0.031l0.017,-0.03l0.018,-0.03l0.018,-0.03l0.02,-0.029l0.019,-0.028l0.021,-0.029l0.021,-0.027l0.022,-0.028l0.023,-0.026l0.023,-0.026l0.024,-0.026l0.024,-0.025l0.025,-0.025l0.026,-0.023l0.026,-0.024l0.027,-0.022l0.027,-0.022l0.028,-0.021l0.028,-0.021l0.029,-0.02l0.029,-0.019l0.029,-0.018l0.03,-0.018l0.031,-0.017l0.03,-0.016l0.031,-0.016l0.032,-0.014l0.032,-0.014l0.032,-0.013l0.032,-0.013l0.033,-0.011l0.033,-0.011l0.034,-0.01l0.033,-0.009l0.035,-0.009l0.034,-0.007l0.035,-0.007l0.036,-0.006l0.035,-0.005l0.037,-0.004l0.036,-0.003l0.037,-0.003l0.038,-0.001l0.054,-0.001l1.569,0l0.006,-2.5Z" fill="#555d66" />
-</svg>;
-icons.topright = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M14.044,2.12l0.101,0.003l0.101,0.007l0.099,0.009l0.099,0.011l0.098,0.014l0.098,0.016l0.096,0.018l0.096,0.021l0.095,0.023l0.094,0.026l0.093,0.028l0.092,0.03l0.091,0.032l0.09,0.034l0.089,0.036l0.088,0.039l0.086,0.04l0.086,0.043l0.084,0.044l0.083,0.047l0.082,0.048l0.081,0.051l0.079,0.052l0.078,0.054l0.077,0.056l0.075,0.057l0.074,0.06l0.072,0.061l0.071,0.063l0.069,0.064l0.068,0.066l0.066,0.068l0.064,0.069l0.063,0.071l0.061,0.072l0.059,0.074l0.058,0.075l0.056,0.076l0.054,0.078l0.052,0.08l0.05,0.08l0.049,0.082l0.046,0.083l0.045,0.085l0.042,0.085l0.041,0.087l0.038,0.088l0.037,0.089l0.034,0.09l0.032,0.091l0.03,0.092l0.028,0.093l0.025,0.094l0.023,0.094l0.021,0.096l0.019,0.097l0.016,0.097l0.014,0.098l0.011,0.099l0.009,0.1l0.006,0.1l0.004,0.101l0.001,0.102l0,1.273l-5.531,-5.214l1.591,0l0.101,0.002Zm-1.692,-0.002l5.531,5.214l-2.5,-0.012l0,-1.245l0,-0.054l-0.002,-0.038l-0.002,-0.037l-0.004,-0.037l-0.004,-0.036l-0.005,-0.036l-0.006,-0.035l-0.006,-0.035l-0.008,-0.035l-0.008,-0.034l-0.009,-0.034l-0.01,-0.033l-0.011,-0.033l-0.012,-0.033l-0.012,-0.033l-0.013,-0.032l-0.014,-0.032l-0.015,-0.031l-0.015,-0.031l-0.016,-0.031l-0.017,-0.03l-0.018,-0.03l-0.018,-0.03l-0.02,-0.029l-0.019,-0.028l-0.021,-0.029l-0.021,-0.027l-0.022,-0.028l-0.023,-0.026l-0.023,-0.026l-0.024,-0.026l-0.024,-0.025l-0.025,-0.025l-0.026,-0.023l-0.026,-0.024l-0.027,-0.022l-0.027,-0.022l-0.028,-0.021l-0.028,-0.021l-0.029,-0.02l-0.029,-0.019l-0.029,-0.018l-0.03,-0.018l-0.031,-0.017l-0.03,-0.016l-0.031,-0.016l-0.032,-0.014l-0.032,-0.014l-0.032,-0.013l-0.032,-0.013l-0.033,-0.011l-0.033,-0.011l-0.034,-0.01l-0.033,-0.009l-0.035,-0.009l-0.034,-0.007l-0.035,-0.007l-0.036,-0.006l-0.035,-0.005l-0.037,-0.004l-0.036,-0.003l-0.037,-0.003l-0.038,-0.001l-0.054,-0.001l-1.57,0l-0.005,-2.5Z" fill="#555d66" />
-	<path d="M6.136,2.516l-0.096,0.003l-0.096,0.006l-0.094,0.009l-0.094,0.01l-0.093,0.014l-0.093,0.015l-0.091,0.017l-0.091,0.02l-0.09,0.022l-0.09,0.024l-0.088,0.027l-0.087,0.029l-0.087,0.03l-0.085,0.032l-0.085,0.034l-0.084,0.037l-0.081,0.038l-0.082,0.041l-0.08,0.042l-0.079,0.045l-0.077,0.045l-0.077,0.049l-0.075,0.049l-0.075,0.051l-0.073,0.054l-0.071,0.054l-0.07,0.057l-0.069,0.058l-0.067,0.06l-0.066,0.06l-0.064,0.063l-0.063,0.065l-0.061,0.065l-0.059,0.068l-0.058,0.068l-0.056,0.07l-0.056,0.072l-0.053,0.072l-0.051,0.074l-0.049,0.076l-0.048,0.076l-0.046,0.078l-0.044,0.079l-0.043,0.08l-0.04,0.081l-0.039,0.083l-0.036,0.083l-0.035,0.085l-0.032,0.085l-0.031,0.087l-0.028,0.087l-0.027,0.088l-0.024,0.09l-0.021,0.089l-0.02,0.091l-0.018,0.092l-0.016,0.093l-0.013,0.093l-0.01,0.094l-0.009,0.095l-0.006,0.095l-0.003,0.096l-0.001,0.096l0,1.043l5.124,-4.786l-1.382,0l-0.096,0.002Zm1.478,-0.002l-5.124,4.786l0.95,-0.004l0,-1.033l0.001,-0.078l0.002,-0.072l0.004,-0.071l0.007,-0.071l0.008,-0.07l0.01,-0.07l0.011,-0.068l0.014,-0.069l0.014,-0.067l0.016,-0.068l0.018,-0.066l0.02,-0.066l0.021,-0.065l0.023,-0.064l0.024,-0.064l0.026,-0.063l0.027,-0.062l0.028,-0.062l0.03,-0.06l0.032,-0.059l0.033,-0.059l0.034,-0.058l0.036,-0.057l0.037,-0.057l0.038,-0.055l0.04,-0.054l0.041,-0.054l0.042,-0.052l0.044,-0.051l0.044,-0.051l0.046,-0.049l0.047,-0.048l0.048,-0.047l0.049,-0.045l0.051,-0.045l0.051,-0.044l0.052,-0.042l0.053,-0.041l0.054,-0.039l0.057,-0.039l0.056,-0.037l0.057,-0.035l0.057,-0.035l0.059,-0.033l0.06,-0.031l0.061,-0.031l0.061,-0.028l0.063,-0.028l0.062,-0.025l0.064,-0.024l0.064,-0.023l0.065,-0.021l0.066,-0.02l0.066,-0.018l0.068,-0.016l0.067,-0.015l0.068,-0.013l0.07,-0.012l0.069,-0.009l0.07,-0.008l0.071,-0.007l0.071,-0.004l0.072,-0.003l0.079,-0.001l1.392,0l-0.017,-0.95Z" fill="#555d66" />
-	<path d="M13.856,17.518l0.096,-0.003l0.096,-0.006l0.094,-0.009l0.094,-0.01l0.093,-0.014l0.094,-0.015l0.091,-0.017l0.091,-0.02l0.09,-0.022l0.089,-0.024l0.089,-0.027l0.087,-0.029l0.087,-0.03l0.085,-0.032l0.085,-0.034l0.083,-0.038l0.082,-0.038l0.082,-0.04l0.079,-0.042l0.079,-0.045l0.078,-0.045l0.077,-0.049l0.075,-0.049l0.074,-0.052l0.073,-0.053l0.072,-0.054l0.07,-0.057l0.068,-0.058l0.068,-0.06l0.065,-0.06l0.065,-0.063l0.063,-0.065l0.06,-0.065l0.06,-0.068l0.058,-0.068l0.056,-0.07l0.055,-0.072l0.053,-0.072l0.052,-0.074l0.049,-0.076l0.048,-0.076l0.046,-0.078l0.044,-0.079l0.043,-0.08l0.039,-0.081l0.039,-0.083l0.037,-0.083l0.035,-0.085l0.032,-0.085l0.03,-0.087l0.029,-0.087l0.026,-0.089l0.024,-0.089l0.022,-0.089l0.02,-0.091l0.018,-0.092l0.015,-0.093l0.014,-0.093l0.01,-0.094l0.009,-0.095l0.005,-0.095l0.004,-0.096l0.001,-0.096l0,-1.043l-5.125,4.786l1.382,0l0.096,-0.002Zm-1.478,0.002l5.125,-4.786l-0.95,0.003l0,1.034l-0.001,0.078l-0.003,0.072l-0.004,0.071l-0.006,0.071l-0.009,0.07l-0.009,0.07l-0.012,0.068l-0.013,0.069l-0.014,0.067l-0.016,0.068l-0.018,0.066l-0.02,0.066l-0.021,0.065l-0.023,0.064l-0.025,0.064l-0.025,0.063l-0.027,0.062l-0.028,0.062l-0.031,0.06l-0.031,0.059l-0.033,0.059l-0.035,0.058l-0.036,0.057l-0.037,0.057l-0.038,0.055l-0.04,0.054l-0.04,0.054l-0.042,0.052l-0.044,0.051l-0.045,0.05l-0.045,0.05l-0.048,0.048l-0.047,0.047l-0.05,0.045l-0.05,0.045l-0.051,0.044l-0.052,0.042l-0.054,0.041l-0.054,0.039l-0.056,0.039l-0.056,0.037l-0.057,0.035l-0.058,0.035l-0.059,0.033l-0.06,0.031l-0.06,0.031l-0.061,0.028l-0.063,0.028l-0.063,0.025l-0.063,0.024l-0.065,0.023l-0.064,0.021l-0.066,0.02l-0.066,0.018l-0.068,0.016l-0.067,0.015l-0.069,0.013l-0.069,0.012l-0.069,0.009l-0.071,0.008l-0.07,0.007l-0.071,0.004l-0.072,0.003l-0.079,0.001l-1.393,0l0.017,0.95Z" fill="#555d66" />
-	<path d="M6.191,17.518l-0.096,-0.003l-0.096,-0.006l-0.094,-0.009l-0.094,-0.01l-0.093,-0.014l-0.093,-0.015l-0.091,-0.017l-0.092,-0.02l-0.09,-0.022l-0.089,-0.024l-0.088,-0.027l-0.088,-0.029l-0.086,-0.03l-0.086,-0.032l-0.084,-0.034l-0.084,-0.038l-0.082,-0.038l-0.081,-0.04l-0.08,-0.042l-0.079,-0.045l-0.078,-0.045l-0.077,-0.049l-0.075,-0.049l-0.074,-0.052l-0.073,-0.053l-0.071,-0.054l-0.071,-0.057l-0.068,-0.058l-0.067,-0.06l-0.066,-0.06l-0.064,-0.063l-0.063,-0.065l-0.061,-0.065l-0.06,-0.068l-0.058,-0.068l-0.056,-0.07l-0.055,-0.072l-0.053,-0.072l-0.051,-0.074l-0.05,-0.076l-0.047,-0.076l-0.047,-0.078l-0.044,-0.079l-0.042,-0.08l-0.04,-0.081l-0.039,-0.083l-0.036,-0.083l-0.035,-0.085l-0.033,-0.085l-0.03,-0.087l-0.029,-0.087l-0.026,-0.089l-0.024,-0.089l-0.022,-0.089l-0.02,-0.091l-0.018,-0.092l-0.015,-0.093l-0.013,-0.093l-0.011,-0.094l-0.008,-0.095l-0.006,-0.095l-0.004,-0.096l-0.001,-0.096l0,-1.043l5.125,4.786l-1.382,0l-0.096,-0.002Zm1.478,0.002l-5.125,-4.786l0.95,0.003l0,1.034l0.001,0.078l0.003,0.072l0.004,0.071l0.006,0.071l0.009,0.07l0.01,0.07l0.011,0.068l0.013,0.069l0.014,0.067l0.017,0.068l0.018,0.066l0.02,0.066l0.021,0.065l0.022,0.064l0.025,0.064l0.026,0.063l0.026,0.062l0.029,0.062l0.03,0.06l0.031,0.059l0.034,0.059l0.034,0.058l0.036,0.057l0.037,0.057l0.038,0.055l0.04,0.054l0.041,0.054l0.042,0.052l0.043,0.051l0.045,0.05l0.046,0.05l0.047,0.048l0.048,0.047l0.049,0.045l0.05,0.045l0.052,0.044l0.052,0.042l0.053,0.041l0.054,0.039l0.056,0.039l0.056,0.037l0.057,0.035l0.058,0.035l0.059,0.033l0.06,0.031l0.061,0.031l0.06,0.028l0.063,0.028l0.063,0.025l0.063,0.024l0.065,0.023l0.065,0.021l0.065,0.02l0.067,0.018l0.067,0.016l0.068,0.015l0.068,0.013l0.069,0.012l0.07,0.009l0.07,0.008l0.07,0.007l0.071,0.004l0.073,0.003l0.079,0.001l1.392,0l-0.017,0.95Z" fill="#555d66" />
-</svg>;
-icons.topleft = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M6.077,2.124l-0.101,0.003l-0.101,0.007l-0.099,0.009l-0.099,0.011l-0.098,0.014l-0.098,0.016l-0.096,0.018l-0.096,0.021l-0.095,0.023l-0.094,0.026l-0.093,0.028l-0.092,0.03l-0.091,0.032l-0.09,0.034l-0.089,0.036l-0.088,0.039l-0.086,0.04l-0.086,0.043l-0.084,0.044l-0.083,0.047l-0.082,0.048l-0.081,0.051l-0.079,0.052l-0.078,0.054l-0.077,0.056l-0.075,0.057l-0.074,0.06l-0.072,0.061l-0.071,0.063l-0.069,0.064l-0.068,0.066l-0.066,0.068l-0.064,0.069l-0.063,0.071l-0.061,0.072l-0.059,0.074l-0.058,0.075l-0.056,0.076l-0.054,0.078l-0.052,0.08l-0.05,0.08l-0.049,0.082l-0.046,0.083l-0.045,0.085l-0.042,0.085l-0.041,0.087l-0.038,0.088l-0.037,0.089l-0.034,0.09l-0.032,0.091l-0.03,0.092l-0.028,0.093l-0.025,0.094l-0.023,0.094l-0.021,0.096l-0.019,0.097l-0.016,0.097l-0.014,0.098l-0.011,0.099l-0.009,0.1l-0.006,0.1l-0.004,0.101l-0.001,0.102l0,1.272l5.531,-5.213l-1.591,0l-0.101,0.002Zm1.692,-0.002l-5.531,5.213l2.5,-0.011l0,-1.245l0,-0.054l0.002,-0.038l0.002,-0.037l0.004,-0.037l0.004,-0.036l0.005,-0.036l0.006,-0.035l0.006,-0.035l0.008,-0.035l0.008,-0.034l0.009,-0.034l0.01,-0.033l0.011,-0.033l0.012,-0.033l0.012,-0.033l0.013,-0.032l0.014,-0.032l0.015,-0.031l0.015,-0.031l0.016,-0.031l0.017,-0.03l0.018,-0.03l0.018,-0.03l0.02,-0.029l0.019,-0.028l0.021,-0.029l0.021,-0.027l0.022,-0.028l0.023,-0.026l0.023,-0.026l0.024,-0.026l0.024,-0.025l0.025,-0.025l0.026,-0.023l0.026,-0.024l0.027,-0.022l0.027,-0.022l0.028,-0.021l0.028,-0.021l0.029,-0.02l0.029,-0.019l0.029,-0.018l0.03,-0.018l0.031,-0.017l0.03,-0.016l0.031,-0.016l0.032,-0.014l0.032,-0.014l0.032,-0.013l0.032,-0.013l0.033,-0.011l0.033,-0.011l0.034,-0.01l0.033,-0.009l0.035,-0.009l0.034,-0.007l0.035,-0.007l0.036,-0.006l0.035,-0.005l0.037,-0.004l0.036,-0.003l0.037,-0.003l0.038,-0.001l0.054,-0.001l1.569,0l0.006,-2.5Z"
-    fill="#555d66" />
-    <path d="M13.85,2.516l0.096,0.003l0.096,0.006l0.094,0.009l0.094,0.01l0.093,0.014l0.093,0.015l0.092,0.017l0.091,0.02l0.09,0.022l0.089,0.024l0.089,0.027l0.087,0.029l0.087,0.03l0.085,0.032l0.085,0.034l0.083,0.037l0.082,0.038l0.082,0.041l0.079,0.042l0.079,0.045l0.078,0.045l0.077,0.049l0.075,0.049l0.074,0.051l0.073,0.054l0.072,0.054l0.07,0.057l0.068,0.058l0.068,0.06l0.065,0.06l0.065,0.063l0.062,0.065l0.061,0.065l0.06,0.068l0.058,0.068l0.056,0.07l0.055,0.072l0.053,0.072l0.052,0.074l0.049,0.076l0.048,0.076l0.046,0.078l0.044,0.079l0.043,0.08l0.039,0.081l0.039,0.083l0.036,0.083l0.036,0.085l0.032,0.085l0.03,0.087l0.029,0.087l0.026,0.088l0.024,0.09l0.022,0.089l0.02,0.091l0.018,0.092l0.015,0.093l0.013,0.093l0.011,0.094l0.008,0.095l0.006,0.095l0.004,0.096l0.001,0.096l0,1.043l-5.125,-4.786l1.382,0l0.096,0.002Zm-1.478,-0.002l5.125,4.786l-0.95,-0.004l0,-1.033l-0.001,-0.078l-0.003,-0.072l-0.004,-0.071l-0.006,-0.071l-0.009,-0.07l-0.009,-0.07l-0.012,-0.068l-0.013,-0.069l-0.014,-0.067l-0.016,-0.068l-0.018,-0.066l-0.02,-0.066l-0.021,-0.065l-0.023,-0.064l-0.025,-0.064l-0.025,-0.063l-0.027,-0.062l-0.028,-0.062l-0.031,-0.06l-0.031,-0.059l-0.033,-0.059l-0.035,-0.058l-0.036,-0.057l-0.037,-0.057l-0.038,-0.055l-0.04,-0.054l-0.041,-0.054l-0.041,-0.052l-0.044,-0.051l-0.045,-0.051l-0.045,-0.049l-0.048,-0.048l-0.047,-0.047l-0.05,-0.045l-0.05,-0.045l-0.051,-0.044l-0.053,-0.042l-0.053,-0.041l-0.054,-0.039l-0.056,-0.039l-0.056,-0.037l-0.057,-0.035l-0.058,-0.035l-0.059,-0.033l-0.06,-0.031l-0.06,-0.031l-0.061,-0.028l-0.063,-0.028l-0.063,-0.025l-0.063,-0.024l-0.065,-0.023l-0.064,-0.021l-0.066,-0.02l-0.066,-0.018l-0.068,-0.016l-0.067,-0.015l-0.069,-0.013l-0.069,-0.012l-0.069,-0.009l-0.071,-0.008l-0.07,-0.007l-0.071,-0.004l-0.072,-0.003l-0.079,-0.001l-1.393,0l0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M13.856,17.518l0.096,-0.003l0.096,-0.006l0.094,-0.009l0.094,-0.01l0.093,-0.014l0.094,-0.015l0.091,-0.017l0.091,-0.02l0.09,-0.022l0.089,-0.024l0.089,-0.027l0.087,-0.029l0.087,-0.03l0.085,-0.032l0.085,-0.034l0.083,-0.038l0.082,-0.038l0.082,-0.04l0.079,-0.042l0.079,-0.045l0.078,-0.045l0.077,-0.049l0.075,-0.049l0.074,-0.052l0.073,-0.053l0.072,-0.054l0.07,-0.057l0.068,-0.058l0.068,-0.06l0.065,-0.06l0.065,-0.063l0.063,-0.065l0.06,-0.065l0.06,-0.068l0.058,-0.068l0.056,-0.07l0.055,-0.072l0.053,-0.072l0.052,-0.074l0.049,-0.076l0.048,-0.076l0.046,-0.078l0.044,-0.079l0.043,-0.08l0.039,-0.081l0.039,-0.083l0.037,-0.083l0.035,-0.085l0.032,-0.085l0.03,-0.087l0.029,-0.087l0.026,-0.089l0.024,-0.089l0.022,-0.089l0.02,-0.091l0.018,-0.092l0.015,-0.093l0.014,-0.093l0.01,-0.094l0.009,-0.095l0.005,-0.095l0.004,-0.096l0.001,-0.096l0,-1.043l-5.125,4.786l1.382,0l0.096,-0.002Zm-1.478,0.002l5.125,-4.786l-0.95,0.003l0,1.034l-0.001,0.078l-0.003,0.072l-0.004,0.071l-0.006,0.071l-0.009,0.07l-0.009,0.07l-0.012,0.068l-0.013,0.069l-0.014,0.067l-0.016,0.068l-0.018,0.066l-0.02,0.066l-0.021,0.065l-0.023,0.064l-0.025,0.064l-0.025,0.063l-0.027,0.062l-0.028,0.062l-0.031,0.06l-0.031,0.059l-0.033,0.059l-0.035,0.058l-0.036,0.057l-0.037,0.057l-0.038,0.055l-0.04,0.054l-0.04,0.054l-0.042,0.052l-0.044,0.051l-0.045,0.05l-0.045,0.05l-0.048,0.048l-0.047,0.047l-0.05,0.045l-0.05,0.045l-0.051,0.044l-0.052,0.042l-0.054,0.041l-0.054,0.039l-0.056,0.039l-0.056,0.037l-0.057,0.035l-0.058,0.035l-0.059,0.033l-0.06,0.031l-0.06,0.031l-0.061,0.028l-0.063,0.028l-0.063,0.025l-0.063,0.024l-0.065,0.023l-0.064,0.021l-0.066,0.02l-0.066,0.018l-0.068,0.016l-0.067,0.015l-0.069,0.013l-0.069,0.012l-0.069,0.009l-0.071,0.008l-0.07,0.007l-0.071,0.004l-0.072,0.003l-0.079,0.001l-1.393,0l0.017,0.95Z"
-    fill="#555d66" />
-    <path d="M6.191,17.518l-0.096,-0.003l-0.096,-0.006l-0.094,-0.009l-0.094,-0.01l-0.093,-0.014l-0.093,-0.015l-0.091,-0.017l-0.092,-0.02l-0.09,-0.022l-0.089,-0.024l-0.088,-0.027l-0.088,-0.029l-0.086,-0.03l-0.086,-0.032l-0.084,-0.034l-0.084,-0.038l-0.082,-0.038l-0.081,-0.04l-0.08,-0.042l-0.079,-0.045l-0.078,-0.045l-0.077,-0.049l-0.075,-0.049l-0.074,-0.052l-0.073,-0.053l-0.071,-0.054l-0.071,-0.057l-0.068,-0.058l-0.067,-0.06l-0.066,-0.06l-0.064,-0.063l-0.063,-0.065l-0.061,-0.065l-0.06,-0.068l-0.058,-0.068l-0.056,-0.07l-0.055,-0.072l-0.053,-0.072l-0.051,-0.074l-0.05,-0.076l-0.047,-0.076l-0.047,-0.078l-0.044,-0.079l-0.042,-0.08l-0.04,-0.081l-0.039,-0.083l-0.036,-0.083l-0.035,-0.085l-0.033,-0.085l-0.03,-0.087l-0.029,-0.087l-0.026,-0.089l-0.024,-0.089l-0.022,-0.089l-0.02,-0.091l-0.018,-0.092l-0.015,-0.093l-0.013,-0.093l-0.011,-0.094l-0.008,-0.095l-0.006,-0.095l-0.004,-0.096l-0.001,-0.096l0,-1.043l5.125,4.786l-1.382,0l-0.096,-0.002Zm1.478,0.002l-5.125,-4.786l0.95,0.003l0,1.034l0.001,0.078l0.003,0.072l0.004,0.071l0.006,0.071l0.009,0.07l0.01,0.07l0.011,0.068l0.013,0.069l0.014,0.067l0.017,0.068l0.018,0.066l0.02,0.066l0.021,0.065l0.022,0.064l0.025,0.064l0.026,0.063l0.026,0.062l0.029,0.062l0.03,0.06l0.031,0.059l0.034,0.059l0.034,0.058l0.036,0.057l0.037,0.057l0.038,0.055l0.04,0.054l0.041,0.054l0.042,0.052l0.043,0.051l0.045,0.05l0.046,0.05l0.047,0.048l0.048,0.047l0.049,0.045l0.05,0.045l0.052,0.044l0.052,0.042l0.053,0.041l0.054,0.039l0.056,0.039l0.056,0.037l0.057,0.035l0.058,0.035l0.059,0.033l0.06,0.031l0.061,0.031l0.06,0.028l0.063,0.028l0.063,0.025l0.063,0.024l0.065,0.023l0.065,0.021l0.065,0.02l0.067,0.018l0.067,0.016l0.068,0.015l0.068,0.013l0.069,0.012l0.07,0.009l0.07,0.008l0.07,0.007l0.071,0.004l0.073,0.003l0.079,0.001l1.392,0l-0.017,0.95Z"
-    fill="#555d66" />
-</svg>;
-icons.bottomleft= <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M6.077,17.873l-0.101,-0.003l-0.101,-0.007l-0.099,-0.009l-0.099,-0.011l-0.098,-0.014l-0.098,-0.016l-0.096,-0.018l-0.096,-0.021l-0.095,-0.023l-0.094,-0.026l-0.093,-0.028l-0.092,-0.03l-0.091,-0.032l-0.09,-0.034l-0.089,-0.036l-0.088,-0.039l-0.086,-0.04l-0.086,-0.043l-0.084,-0.044l-0.083,-0.047l-0.082,-0.048l-0.081,-0.051l-0.079,-0.052l-0.078,-0.054l-0.077,-0.056l-0.075,-0.057l-0.074,-0.06l-0.072,-0.061l-0.071,-0.063l-0.069,-0.064l-0.068,-0.066l-0.066,-0.068l-0.064,-0.069l-0.063,-0.071l-0.061,-0.072l-0.059,-0.074l-0.058,-0.075l-0.056,-0.076l-0.054,-0.078l-0.052,-0.08l-0.05,-0.08l-0.049,-0.082l-0.046,-0.083l-0.045,-0.085l-0.042,-0.085l-0.041,-0.087l-0.038,-0.088l-0.037,-0.089l-0.034,-0.09l-0.032,-0.091l-0.03,-0.092l-0.028,-0.093l-0.025,-0.094l-0.023,-0.094l-0.021,-0.096l-0.019,-0.097l-0.016,-0.097l-0.014,-0.098l-0.011,-0.099l-0.009,-0.1l-0.006,-0.1l-0.004,-0.101l-0.001,-0.102l0,-1.272l5.531,5.213l-1.591,0l-0.101,-0.002Zm1.692,0.002l-5.531,-5.213l2.5,0.012l0,1.244l0,0.054l0.002,0.038l0.002,0.037l0.004,0.037l0.004,0.036l0.005,0.036l0.006,0.035l0.006,0.035l0.008,0.035l0.008,0.034l0.009,0.034l0.01,0.033l0.011,0.033l0.012,0.033l0.012,0.033l0.013,0.032l0.014,0.032l0.015,0.031l0.015,0.031l0.016,0.031l0.017,0.03l0.018,0.03l0.018,0.03l0.02,0.029l0.019,0.028l0.021,0.029l0.021,0.027l0.022,0.028l0.023,0.026l0.023,0.026l0.024,0.026l0.024,0.025l0.025,0.025l0.026,0.023l0.026,0.024l0.027,0.022l0.027,0.022l0.028,0.021l0.028,0.021l0.029,0.02l0.029,0.019l0.029,0.018l0.03,0.018l0.031,0.017l0.03,0.016l0.031,0.016l0.032,0.014l0.032,0.014l0.032,0.013l0.032,0.013l0.033,0.011l0.033,0.011l0.034,0.01l0.033,0.009l0.035,0.009l0.034,0.007l0.035,0.007l0.036,0.006l0.035,0.005l0.037,0.004l0.036,0.003l0.037,0.003l0.038,0.001l0.054,0.001l1.569,0l0.006,2.5Z"
-    fill="#555d66" />
-    <path d="M13.85,2.516l0.096,0.003l0.096,0.006l0.094,0.009l0.094,0.01l0.093,0.014l0.093,0.015l0.092,0.017l0.091,0.02l0.09,0.022l0.089,0.024l0.089,0.027l0.087,0.029l0.087,0.03l0.085,0.032l0.085,0.034l0.083,0.037l0.082,0.038l0.082,0.041l0.079,0.042l0.079,0.045l0.078,0.045l0.077,0.049l0.075,0.049l0.074,0.051l0.073,0.054l0.072,0.054l0.07,0.057l0.068,0.058l0.068,0.06l0.065,0.06l0.065,0.063l0.062,0.065l0.061,0.065l0.06,0.068l0.058,0.068l0.056,0.07l0.055,0.072l0.053,0.072l0.052,0.074l0.049,0.076l0.048,0.076l0.046,0.078l0.044,0.079l0.043,0.08l0.039,0.081l0.039,0.083l0.036,0.083l0.036,0.085l0.032,0.085l0.03,0.087l0.029,0.087l0.026,0.088l0.024,0.09l0.022,0.089l0.02,0.091l0.018,0.092l0.015,0.093l0.013,0.093l0.011,0.094l0.008,0.095l0.006,0.095l0.004,0.096l0.001,0.096l0,1.043l-5.125,-4.786l1.382,0l0.096,0.002Zm-1.478,-0.002l5.125,4.786l-0.95,-0.004l0,-1.033l-0.001,-0.078l-0.003,-0.072l-0.004,-0.071l-0.006,-0.071l-0.009,-0.07l-0.009,-0.07l-0.012,-0.068l-0.013,-0.069l-0.014,-0.067l-0.016,-0.068l-0.018,-0.066l-0.02,-0.066l-0.021,-0.065l-0.023,-0.064l-0.025,-0.064l-0.025,-0.063l-0.027,-0.062l-0.028,-0.062l-0.031,-0.06l-0.031,-0.059l-0.033,-0.059l-0.035,-0.058l-0.036,-0.057l-0.037,-0.057l-0.038,-0.055l-0.04,-0.054l-0.041,-0.054l-0.041,-0.052l-0.044,-0.051l-0.045,-0.051l-0.045,-0.049l-0.048,-0.048l-0.047,-0.047l-0.05,-0.045l-0.05,-0.045l-0.051,-0.044l-0.053,-0.042l-0.053,-0.041l-0.054,-0.039l-0.056,-0.039l-0.056,-0.037l-0.057,-0.035l-0.058,-0.035l-0.059,-0.033l-0.06,-0.031l-0.06,-0.031l-0.061,-0.028l-0.063,-0.028l-0.063,-0.025l-0.063,-0.024l-0.065,-0.023l-0.064,-0.021l-0.066,-0.02l-0.066,-0.018l-0.068,-0.016l-0.067,-0.015l-0.069,-0.013l-0.069,-0.012l-0.069,-0.009l-0.071,-0.008l-0.07,-0.007l-0.071,-0.004l-0.072,-0.003l-0.079,-0.001l-1.393,0l0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M6.136,2.516l-0.096,0.003l-0.096,0.006l-0.094,0.009l-0.094,0.01l-0.093,0.014l-0.093,0.015l-0.091,0.017l-0.091,0.02l-0.09,0.022l-0.09,0.024l-0.088,0.027l-0.087,0.029l-0.087,0.03l-0.085,0.032l-0.085,0.034l-0.084,0.037l-0.081,0.038l-0.082,0.041l-0.08,0.042l-0.079,0.045l-0.077,0.045l-0.077,0.049l-0.075,0.049l-0.075,0.051l-0.073,0.054l-0.071,0.054l-0.07,0.057l-0.069,0.058l-0.067,0.06l-0.066,0.06l-0.064,0.063l-0.063,0.065l-0.061,0.065l-0.059,0.068l-0.058,0.068l-0.056,0.07l-0.056,0.072l-0.053,0.072l-0.051,0.074l-0.049,0.076l-0.048,0.076l-0.046,0.078l-0.044,0.079l-0.043,0.08l-0.04,0.081l-0.039,0.083l-0.036,0.083l-0.035,0.085l-0.032,0.085l-0.031,0.087l-0.028,0.087l-0.027,0.088l-0.024,0.09l-0.021,0.089l-0.02,0.091l-0.018,0.092l-0.016,0.093l-0.013,0.093l-0.01,0.094l-0.009,0.095l-0.006,0.095l-0.003,0.096l-0.001,0.096l0,1.043l5.124,-4.786l-1.382,0l-0.096,0.002Zm1.478,-0.002l-5.124,4.786l0.95,-0.004l0,-1.033l0.001,-0.078l0.002,-0.072l0.004,-0.071l0.007,-0.071l0.008,-0.07l0.01,-0.07l0.011,-0.068l0.014,-0.069l0.014,-0.067l0.016,-0.068l0.018,-0.066l0.02,-0.066l0.021,-0.065l0.023,-0.064l0.024,-0.064l0.026,-0.063l0.027,-0.062l0.028,-0.062l0.03,-0.06l0.032,-0.059l0.033,-0.059l0.034,-0.058l0.036,-0.057l0.037,-0.057l0.038,-0.055l0.04,-0.054l0.041,-0.054l0.042,-0.052l0.044,-0.051l0.044,-0.051l0.046,-0.049l0.047,-0.048l0.048,-0.047l0.049,-0.045l0.051,-0.045l0.051,-0.044l0.052,-0.042l0.053,-0.041l0.054,-0.039l0.057,-0.039l0.056,-0.037l0.057,-0.035l0.057,-0.035l0.059,-0.033l0.06,-0.031l0.061,-0.031l0.061,-0.028l0.063,-0.028l0.062,-0.025l0.064,-0.024l0.064,-0.023l0.065,-0.021l0.066,-0.02l0.066,-0.018l0.068,-0.016l0.067,-0.015l0.068,-0.013l0.07,-0.012l0.069,-0.009l0.07,-0.008l0.071,-0.007l0.071,-0.004l0.072,-0.003l0.079,-0.001l1.392,0l-0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M13.856,17.518l0.096,-0.003l0.096,-0.006l0.094,-0.009l0.094,-0.01l0.093,-0.014l0.094,-0.015l0.091,-0.017l0.091,-0.02l0.09,-0.022l0.089,-0.024l0.089,-0.027l0.087,-0.029l0.087,-0.03l0.085,-0.032l0.085,-0.034l0.083,-0.038l0.082,-0.038l0.082,-0.04l0.079,-0.042l0.079,-0.045l0.078,-0.045l0.077,-0.049l0.075,-0.049l0.074,-0.052l0.073,-0.053l0.072,-0.054l0.07,-0.057l0.068,-0.058l0.068,-0.06l0.065,-0.06l0.065,-0.063l0.063,-0.065l0.06,-0.065l0.06,-0.068l0.058,-0.068l0.056,-0.07l0.055,-0.072l0.053,-0.072l0.052,-0.074l0.049,-0.076l0.048,-0.076l0.046,-0.078l0.044,-0.079l0.043,-0.08l0.039,-0.081l0.039,-0.083l0.037,-0.083l0.035,-0.085l0.032,-0.085l0.03,-0.087l0.029,-0.087l0.026,-0.089l0.024,-0.089l0.022,-0.089l0.02,-0.091l0.018,-0.092l0.015,-0.093l0.014,-0.093l0.01,-0.094l0.009,-0.095l0.005,-0.095l0.004,-0.096l0.001,-0.096l0,-1.043l-5.125,4.786l1.382,0l0.096,-0.002Zm-1.478,0.002l5.125,-4.786l-0.95,0.003l0,1.034l-0.001,0.078l-0.003,0.072l-0.004,0.071l-0.006,0.071l-0.009,0.07l-0.009,0.07l-0.012,0.068l-0.013,0.069l-0.014,0.067l-0.016,0.068l-0.018,0.066l-0.02,0.066l-0.021,0.065l-0.023,0.064l-0.025,0.064l-0.025,0.063l-0.027,0.062l-0.028,0.062l-0.031,0.06l-0.031,0.059l-0.033,0.059l-0.035,0.058l-0.036,0.057l-0.037,0.057l-0.038,0.055l-0.04,0.054l-0.04,0.054l-0.042,0.052l-0.044,0.051l-0.045,0.05l-0.045,0.05l-0.048,0.048l-0.047,0.047l-0.05,0.045l-0.05,0.045l-0.051,0.044l-0.052,0.042l-0.054,0.041l-0.054,0.039l-0.056,0.039l-0.056,0.037l-0.057,0.035l-0.058,0.035l-0.059,0.033l-0.06,0.031l-0.06,0.031l-0.061,0.028l-0.063,0.028l-0.063,0.025l-0.063,0.024l-0.065,0.023l-0.064,0.021l-0.066,0.02l-0.066,0.018l-0.068,0.016l-0.067,0.015l-0.069,0.013l-0.069,0.012l-0.069,0.009l-0.071,0.008l-0.07,0.007l-0.071,0.004l-0.072,0.003l-0.079,0.001l-1.393,0l0.017,0.95Z"
-    fill="#555d66" />
-</svg>;
-icons.bottomright = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M14.044,17.873l0.101,-0.003l0.101,-0.007l0.099,-0.009l0.099,-0.011l0.098,-0.014l0.098,-0.016l0.096,-0.018l0.096,-0.021l0.095,-0.023l0.094,-0.026l0.093,-0.028l0.092,-0.03l0.091,-0.032l0.09,-0.034l0.089,-0.036l0.088,-0.039l0.086,-0.04l0.086,-0.043l0.084,-0.044l0.083,-0.047l0.082,-0.048l0.081,-0.051l0.079,-0.052l0.078,-0.054l0.077,-0.056l0.075,-0.057l0.074,-0.06l0.072,-0.061l0.071,-0.063l0.069,-0.064l0.068,-0.066l0.066,-0.068l0.064,-0.069l0.063,-0.071l0.061,-0.072l0.059,-0.074l0.058,-0.075l0.056,-0.076l0.054,-0.078l0.052,-0.08l0.05,-0.08l0.049,-0.082l0.046,-0.083l0.045,-0.085l0.042,-0.085l0.041,-0.087l0.038,-0.088l0.037,-0.089l0.034,-0.09l0.032,-0.091l0.03,-0.092l0.028,-0.093l0.025,-0.094l0.023,-0.094l0.021,-0.096l0.019,-0.097l0.016,-0.097l0.014,-0.098l0.011,-0.099l0.009,-0.1l0.006,-0.1l0.004,-0.101l0.001,-0.102l0,-1.272l-5.531,5.213l1.591,0l0.101,-0.002Zm-1.692,0.002l5.531,-5.213l-2.5,0.012l0,1.244l0,0.054l-0.002,0.038l-0.002,0.037l-0.004,0.037l-0.004,0.036l-0.005,0.036l-0.006,0.035l-0.006,0.035l-0.008,0.035l-0.008,0.034l-0.009,0.034l-0.01,0.033l-0.011,0.033l-0.012,0.033l-0.012,0.033l-0.013,0.032l-0.014,0.032l-0.015,0.031l-0.015,0.031l-0.016,0.031l-0.017,0.03l-0.018,0.03l-0.018,0.03l-0.02,0.029l-0.019,0.028l-0.021,0.029l-0.021,0.027l-0.022,0.028l-0.023,0.026l-0.023,0.026l-0.024,0.026l-0.024,0.025l-0.025,0.025l-0.026,0.023l-0.026,0.024l-0.027,0.022l-0.027,0.022l-0.028,0.021l-0.028,0.021l-0.029,0.02l-0.029,0.019l-0.029,0.018l-0.03,0.018l-0.031,0.017l-0.03,0.016l-0.031,0.016l-0.032,0.014l-0.032,0.014l-0.032,0.013l-0.032,0.013l-0.033,0.011l-0.033,0.011l-0.034,0.01l-0.033,0.009l-0.035,0.009l-0.034,0.007l-0.035,0.007l-0.036,0.006l-0.035,0.005l-0.037,0.004l-0.036,0.003l-0.037,0.003l-0.038,0.001l-0.054,0.001l-1.57,0l-0.005,2.5Z"
-    fill="#555d66" />
-    <path d="M13.85,2.516l0.096,0.003l0.096,0.006l0.094,0.009l0.094,0.01l0.093,0.014l0.093,0.015l0.092,0.017l0.091,0.02l0.09,0.022l0.089,0.024l0.089,0.027l0.087,0.029l0.087,0.03l0.085,0.032l0.085,0.034l0.083,0.037l0.082,0.038l0.082,0.041l0.079,0.042l0.079,0.045l0.078,0.045l0.077,0.049l0.075,0.049l0.074,0.051l0.073,0.054l0.072,0.054l0.07,0.057l0.068,0.058l0.068,0.06l0.065,0.06l0.065,0.063l0.062,0.065l0.061,0.065l0.06,0.068l0.058,0.068l0.056,0.07l0.055,0.072l0.053,0.072l0.052,0.074l0.049,0.076l0.048,0.076l0.046,0.078l0.044,0.079l0.043,0.08l0.039,0.081l0.039,0.083l0.036,0.083l0.036,0.085l0.032,0.085l0.03,0.087l0.029,0.087l0.026,0.088l0.024,0.09l0.022,0.089l0.02,0.091l0.018,0.092l0.015,0.093l0.013,0.093l0.011,0.094l0.008,0.095l0.006,0.095l0.004,0.096l0.001,0.096l0,1.043l-5.125,-4.786l1.382,0l0.096,0.002Zm-1.478,-0.002l5.125,4.786l-0.95,-0.004l0,-1.033l-0.001,-0.078l-0.003,-0.072l-0.004,-0.071l-0.006,-0.071l-0.009,-0.07l-0.009,-0.07l-0.012,-0.068l-0.013,-0.069l-0.014,-0.067l-0.016,-0.068l-0.018,-0.066l-0.02,-0.066l-0.021,-0.065l-0.023,-0.064l-0.025,-0.064l-0.025,-0.063l-0.027,-0.062l-0.028,-0.062l-0.031,-0.06l-0.031,-0.059l-0.033,-0.059l-0.035,-0.058l-0.036,-0.057l-0.037,-0.057l-0.038,-0.055l-0.04,-0.054l-0.041,-0.054l-0.041,-0.052l-0.044,-0.051l-0.045,-0.051l-0.045,-0.049l-0.048,-0.048l-0.047,-0.047l-0.05,-0.045l-0.05,-0.045l-0.051,-0.044l-0.053,-0.042l-0.053,-0.041l-0.054,-0.039l-0.056,-0.039l-0.056,-0.037l-0.057,-0.035l-0.058,-0.035l-0.059,-0.033l-0.06,-0.031l-0.06,-0.031l-0.061,-0.028l-0.063,-0.028l-0.063,-0.025l-0.063,-0.024l-0.065,-0.023l-0.064,-0.021l-0.066,-0.02l-0.066,-0.018l-0.068,-0.016l-0.067,-0.015l-0.069,-0.013l-0.069,-0.012l-0.069,-0.009l-0.071,-0.008l-0.07,-0.007l-0.071,-0.004l-0.072,-0.003l-0.079,-0.001l-1.393,0l0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M6.136,2.516l-0.096,0.003l-0.096,0.006l-0.094,0.009l-0.094,0.01l-0.093,0.014l-0.093,0.015l-0.091,0.017l-0.091,0.02l-0.09,0.022l-0.09,0.024l-0.088,0.027l-0.087,0.029l-0.087,0.03l-0.085,0.032l-0.085,0.034l-0.084,0.037l-0.081,0.038l-0.082,0.041l-0.08,0.042l-0.079,0.045l-0.077,0.045l-0.077,0.049l-0.075,0.049l-0.075,0.051l-0.073,0.054l-0.071,0.054l-0.07,0.057l-0.069,0.058l-0.067,0.06l-0.066,0.06l-0.064,0.063l-0.063,0.065l-0.061,0.065l-0.059,0.068l-0.058,0.068l-0.056,0.07l-0.056,0.072l-0.053,0.072l-0.051,0.074l-0.049,0.076l-0.048,0.076l-0.046,0.078l-0.044,0.079l-0.043,0.08l-0.04,0.081l-0.039,0.083l-0.036,0.083l-0.035,0.085l-0.032,0.085l-0.031,0.087l-0.028,0.087l-0.027,0.088l-0.024,0.09l-0.021,0.089l-0.02,0.091l-0.018,0.092l-0.016,0.093l-0.013,0.093l-0.01,0.094l-0.009,0.095l-0.006,0.095l-0.003,0.096l-0.001,0.096l0,1.043l5.124,-4.786l-1.382,0l-0.096,0.002Zm1.478,-0.002l-5.124,4.786l0.95,-0.004l0,-1.033l0.001,-0.078l0.002,-0.072l0.004,-0.071l0.007,-0.071l0.008,-0.07l0.01,-0.07l0.011,-0.068l0.014,-0.069l0.014,-0.067l0.016,-0.068l0.018,-0.066l0.02,-0.066l0.021,-0.065l0.023,-0.064l0.024,-0.064l0.026,-0.063l0.027,-0.062l0.028,-0.062l0.03,-0.06l0.032,-0.059l0.033,-0.059l0.034,-0.058l0.036,-0.057l0.037,-0.057l0.038,-0.055l0.04,-0.054l0.041,-0.054l0.042,-0.052l0.044,-0.051l0.044,-0.051l0.046,-0.049l0.047,-0.048l0.048,-0.047l0.049,-0.045l0.051,-0.045l0.051,-0.044l0.052,-0.042l0.053,-0.041l0.054,-0.039l0.057,-0.039l0.056,-0.037l0.057,-0.035l0.057,-0.035l0.059,-0.033l0.06,-0.031l0.061,-0.031l0.061,-0.028l0.063,-0.028l0.062,-0.025l0.064,-0.024l0.064,-0.023l0.065,-0.021l0.066,-0.02l0.066,-0.018l0.068,-0.016l0.067,-0.015l0.068,-0.013l0.07,-0.012l0.069,-0.009l0.07,-0.008l0.071,-0.007l0.071,-0.004l0.072,-0.003l0.079,-0.001l1.392,0l-0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M6.191,17.518l-0.096,-0.003l-0.096,-0.006l-0.094,-0.009l-0.094,-0.01l-0.093,-0.014l-0.093,-0.015l-0.091,-0.017l-0.092,-0.02l-0.09,-0.022l-0.089,-0.024l-0.088,-0.027l-0.088,-0.029l-0.086,-0.03l-0.086,-0.032l-0.084,-0.034l-0.084,-0.038l-0.082,-0.038l-0.081,-0.04l-0.08,-0.042l-0.079,-0.045l-0.078,-0.045l-0.077,-0.049l-0.075,-0.049l-0.074,-0.052l-0.073,-0.053l-0.071,-0.054l-0.071,-0.057l-0.068,-0.058l-0.067,-0.06l-0.066,-0.06l-0.064,-0.063l-0.063,-0.065l-0.061,-0.065l-0.06,-0.068l-0.058,-0.068l-0.056,-0.07l-0.055,-0.072l-0.053,-0.072l-0.051,-0.074l-0.05,-0.076l-0.047,-0.076l-0.047,-0.078l-0.044,-0.079l-0.042,-0.08l-0.04,-0.081l-0.039,-0.083l-0.036,-0.083l-0.035,-0.085l-0.033,-0.085l-0.03,-0.087l-0.029,-0.087l-0.026,-0.089l-0.024,-0.089l-0.022,-0.089l-0.02,-0.091l-0.018,-0.092l-0.015,-0.093l-0.013,-0.093l-0.011,-0.094l-0.008,-0.095l-0.006,-0.095l-0.004,-0.096l-0.001,-0.096l0,-1.043l5.125,4.786l-1.382,0l-0.096,-0.002Zm1.478,0.002l-5.125,-4.786l0.95,0.003l0,1.034l0.001,0.078l0.003,0.072l0.004,0.071l0.006,0.071l0.009,0.07l0.01,0.07l0.011,0.068l0.013,0.069l0.014,0.067l0.017,0.068l0.018,0.066l0.02,0.066l0.021,0.065l0.022,0.064l0.025,0.064l0.026,0.063l0.026,0.062l0.029,0.062l0.03,0.06l0.031,0.059l0.034,0.059l0.034,0.058l0.036,0.057l0.037,0.057l0.038,0.055l0.04,0.054l0.041,0.054l0.042,0.052l0.043,0.051l0.045,0.05l0.046,0.05l0.047,0.048l0.048,0.047l0.049,0.045l0.05,0.045l0.052,0.044l0.052,0.042l0.053,0.041l0.054,0.039l0.056,0.039l0.056,0.037l0.057,0.035l0.058,0.035l0.059,0.033l0.06,0.031l0.061,0.031l0.06,0.028l0.063,0.028l0.063,0.025l0.063,0.024l0.065,0.023l0.065,0.021l0.065,0.02l0.067,0.018l0.067,0.016l0.068,0.015l0.068,0.013l0.069,0.012l0.07,0.009l0.07,0.008l0.07,0.007l0.071,0.004l0.073,0.003l0.079,0.001l1.392,0l-0.017,0.95Z"
-    fill="#555d66" />
-</svg>;
-icons.h1 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M12.993,9.914c0.232,-0.096 0.474,-0.206 0.726,-0.33c0.252,-0.124 0.498,-0.26 0.738,-0.408c0.24,-0.148 0.47,-0.304 0.69,-0.468c0.22,-0.164 0.418,-0.338 0.594,-0.522l1.248,0l0,8.316l-1.788,0l0,-5.988c-0.24,0.16 -0.508,0.306 -0.804,0.438c-0.296,0.132 -0.584,0.246 -0.864,0.342l-0.54,-1.38Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h2 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-	/>
-	<path d="M18.019,10.342c0,0.304 -0.06,0.596 -0.18,0.876c-0.12,0.28 -0.276,0.55 -0.468,0.81c-0.192,0.26 -0.408,0.51 -0.648,0.75c-0.24,0.24 -0.476,0.468 -0.708,0.684c-0.12,0.112 -0.25,0.238 -0.39,0.378c-0.14,0.14 -0.274,0.282 -0.402,0.426c-0.128,0.144 -0.242,0.278 -0.342,0.402c-0.1,0.124 -0.162,0.226 -0.186,0.306l3.576,0l0,1.5l-5.544,0c-0.016,-0.088 -0.024,-0.2 -0.024,-0.336c0,-0.136 0,-0.232 0,-0.288c0,-0.384 0.062,-0.736 0.186,-1.056c0.124,-0.32 0.286,-0.618 0.486,-0.894c0.2,-0.276 0.426,-0.534 0.678,-0.774c0.252,-0.24 0.502,-0.48 0.75,-0.72c0.192,-0.184 0.372,-0.358 0.54,-0.522c0.168,-0.164 0.316,-0.324 0.444,-0.48c0.128,-0.156 0.228,-0.312 0.3,-0.468c0.072,-0.156 0.108,-0.314 0.108,-0.474c0,-0.352 -0.1,-0.6 -0.3,-0.744c-0.2,-0.144 -0.448,-0.216 -0.744,-0.216c-0.216,0 -0.418,0.034 -0.606,0.102c-0.188,0.068 -0.36,0.148 -0.516,0.24c-0.156,0.092 -0.29,0.184 -0.402,0.276c-0.112,0.092 -0.196,0.166 -0.252,0.222l-0.888,-1.248c0.352,-0.328 0.762,-0.598 1.23,-0.81c0.468,-0.212 0.97,-0.318 1.506,-0.318c0.488,0 0.908,0.056 1.26,0.168c0.352,0.112 0.642,0.27 0.87,0.474c0.228,0.204 0.396,0.452 0.504,0.744c0.108,0.292 0.162,0.622 0.162,0.99Z"
-	/>
-</svg>;
-icons.h3 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M14.884,16.485c-0.216,0 -0.444,-0.014 -0.684,-0.042c-0.24,-0.028 -0.472,-0.066 -0.696,-0.114c-0.224,-0.048 -0.428,-0.1 -0.612,-0.156c-0.184,-0.056 -0.328,-0.108 -0.432,-0.156l0.348,-1.488c0.208,0.088 0.474,0.182 0.798,0.282c0.324,0.1 0.726,0.15 1.206,0.15c0.552,0 0.956,-0.104 1.212,-0.312c0.256,-0.208 0.384,-0.488 0.384,-0.84c0,-0.216 -0.046,-0.398 -0.138,-0.546c-0.092,-0.148 -0.218,-0.268 -0.378,-0.36c-0.16,-0.092 -0.35,-0.156 -0.57,-0.192c-0.22,-0.036 -0.454,-0.054 -0.702,-0.054l-0.696,0l0,-1.44l0.792,0c0.176,0 0.346,-0.016 0.51,-0.048c0.164,-0.032 0.31,-0.086 0.438,-0.162c0.128,-0.076 0.23,-0.18 0.306,-0.312c0.076,-0.132 0.114,-0.298 0.114,-0.498c0,-0.152 -0.032,-0.284 -0.096,-0.396c-0.064,-0.112 -0.146,-0.204 -0.246,-0.276c-0.1,-0.072 -0.216,-0.126 -0.348,-0.162c-0.132,-0.036 -0.266,-0.054 -0.402,-0.054c-0.344,0 -0.662,0.052 -0.954,0.156c-0.292,0.104 -0.558,0.232 -0.798,0.384l-0.636,-1.308c0.128,-0.08 0.278,-0.164 0.45,-0.252c0.172,-0.088 0.362,-0.168 0.57,-0.24c0.208,-0.072 0.43,-0.132 0.666,-0.18c0.236,-0.048 0.486,-0.072 0.75,-0.072c0.488,0 0.91,0.058 1.266,0.174c0.356,0.116 0.65,0.28 0.882,0.492c0.232,0.212 0.404,0.46 0.516,0.744c0.112,0.284 0.168,0.594 0.168,0.93c0,0.328 -0.092,0.646 -0.276,0.954c-0.184,0.308 -0.432,0.542 -0.744,0.702c0.432,0.176 0.766,0.438 1.002,0.786c0.236,0.348 0.354,0.766 0.354,1.254c0,0.384 -0.064,0.738 -0.192,1.062c-0.128,0.324 -0.328,0.604 -0.6,0.84c-0.272,0.236 -0.618,0.42 -1.038,0.552c-0.42,0.132 -0.918,0.198 -1.494,0.198Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h4 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M15.966,10.329c-0.312,0.408 -0.644,0.852 -0.996,1.332c-0.352,0.48 -0.648,0.972 -0.888,1.476l1.884,0l0,-2.808Zm1.752,-2.184l0,4.992l0.888,0l0,1.464l-0.888,0l0,1.86l-1.752,0l0,-1.86l-3.588,0l0,-1.308c0.176,-0.352 0.396,-0.75 0.66,-1.194c0.264,-0.444 0.558,-0.9 0.882,-1.368c0.324,-0.468 0.666,-0.926 1.026,-1.374c0.36,-0.448 0.72,-0.852 1.08,-1.212l1.692,0Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h5 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M16.468,13.857c0,-0.232 -0.052,-0.438 -0.156,-0.618c-0.104,-0.18 -0.284,-0.33 -0.54,-0.45c-0.256,-0.12 -0.6,-0.212 -1.032,-0.276c-0.432,-0.064 -0.976,-0.096 -1.632,-0.096c0.088,-0.744 0.16,-1.5 0.216,-2.268c0.056,-0.768 0.1,-1.488 0.132,-2.16l4.5,0l0,1.5l-3,0c-0.016,0.288 -0.034,0.566 -0.054,0.834c-0.02,0.268 -0.042,0.502 -0.066,0.702c1.168,0.08 2.03,0.354 2.586,0.822c0.556,0.468 0.834,1.118 0.834,1.95c0,0.384 -0.068,0.74 -0.204,1.068c-0.136,0.328 -0.34,0.612 -0.612,0.852c-0.272,0.24 -0.614,0.428 -1.026,0.564c-0.412,0.136 -0.894,0.204 -1.446,0.204c-0.216,0 -0.442,-0.014 -0.678,-0.042c-0.236,-0.028 -0.464,-0.064 -0.684,-0.108c-0.22,-0.044 -0.418,-0.09 -0.594,-0.138c-0.176,-0.048 -0.312,-0.096 -0.408,-0.144l0.324,-1.476c0.2,0.088 0.464,0.174 0.792,0.258c0.328,0.084 0.72,0.126 1.176,0.126c0.552,0 0.952,-0.11 1.2,-0.33c0.248,-0.22 0.372,-0.478 0.372,-0.774Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h6 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M15.436,12.393c-0.176,0 -0.364,0.016 -0.564,0.048c-0.2,0.032 -0.356,0.076 -0.468,0.132c0,0.032 -0.002,0.08 -0.006,0.144c-0.004,0.064 -0.006,0.12 -0.006,0.168c0,0.288 0.02,0.56 0.06,0.816c0.04,0.256 0.108,0.48 0.204,0.672c0.096,0.192 0.224,0.342 0.384,0.45c0.16,0.108 0.36,0.162 0.6,0.162c0.2,0 0.37,-0.042 0.51,-0.126c0.14,-0.084 0.256,-0.19 0.348,-0.318c0.092,-0.128 0.16,-0.27 0.204,-0.426c0.044,-0.156 0.066,-0.306 0.066,-0.45c0,-0.408 -0.102,-0.722 -0.306,-0.942c-0.204,-0.22 -0.546,-0.33 -1.026,-0.33Zm0.228,-1.392c0.52,0 0.962,0.07 1.326,0.21c0.364,0.14 0.662,0.33 0.894,0.57c0.232,0.24 0.4,0.52 0.504,0.84c0.104,0.32 0.156,0.66 0.156,1.02c0,0.32 -0.058,0.648 -0.174,0.984c-0.116,0.336 -0.292,0.64 -0.528,0.912c-0.236,0.272 -0.536,0.496 -0.9,0.672c-0.364,0.176 -0.794,0.264 -1.29,0.264c-1,0 -1.76,-0.316 -2.28,-0.948c-0.52,-0.632 -0.78,-1.504 -0.78,-2.616c0,-0.8 0.124,-1.508 0.372,-2.124c0.248,-0.616 0.602,-1.134 1.062,-1.554c0.46,-0.42 1.016,-0.74 1.668,-0.96c0.652,-0.22 1.386,-0.334 2.202,-0.342c0.016,0.248 0.032,0.49 0.048,0.726c0.016,0.236 0.032,0.482 0.048,0.738c-0.408,0.008 -0.79,0.046 -1.146,0.114c-0.356,0.068 -0.68,0.174 -0.972,0.318c-0.292,0.144 -0.546,0.33 -0.762,0.558c-0.216,0.228 -0.384,0.506 -0.504,0.834c0.176,-0.08 0.356,-0.136 0.54,-0.168c0.184,-0.032 0.356,-0.048 0.516,-0.048Z"
-    fillRule="nonzero" />
-</svg>;
-icons.fontfamily = <svg width="16px" height="16px" viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M39.939,7.124l0,-3.751l-11.251,0l0,3.751l3.75,0l0,33.752l-3.75,0l0,3.751l11.251,0l0,-3.751l-3.751,0l0,-33.752l3.751,0Zm-11.251,4.391c0,-0.195 -0.366,-0.429 -0.6,-0.596c-1.033,-0.771 -2.175,-1.273 -3.443,-1.506c-1.268,-0.234 -2.747,-0.35 -4.45,-0.35c-1.234,0 -2.501,0.195 -3.797,0.592c-1.297,0.396 -2.4,0.876 -3.311,1.446c-1.056,0.667 -1.911,1.429 -2.569,2.278c-0.658,0.852 -0.988,1.744 -0.988,2.676c0,0.893 0.238,1.697 0.714,2.417c0.476,0.72 1.221,1.078 2.233,1.078c1.135,0 2.022,-0.292 2.661,-0.883c0.639,-0.585 0.958,-1.234 0.958,-1.944c0,-0.668 -0.097,-1.397 -0.289,-2.189c-0.195,-0.789 -0.288,-1.387 -0.288,-1.792c0.324,-0.345 0.887,-0.674 1.687,-0.989c0.799,-0.315 1.667,-0.47 2.599,-0.47c1.358,0 2.406,0.277 3.147,0.834c0.74,0.559 1.312,1.212 1.717,1.962c0.364,0.669 0.269,2.186 0.269,2.186l0,4.472c0,0.527 -2.545,1.257 -5.333,2.189c-2.789,0.931 -4.484,1.569 -5.397,1.914c-0.728,0.285 -1.429,0.666 -2.261,1.14c-0.831,0.478 -1.463,1.018 -1.971,1.628c-0.649,0.729 -1.116,1.549 -1.44,2.46c-0.325,0.911 -0.475,1.944 -0.475,3.101c0,2.23 0.73,4.034 2.179,5.412c1.448,1.376 3.3,2.068 5.549,2.068c2.128,0 3.947,-0.834 5.456,-1.603c1.514,-0.774 2.928,-1.92 4.244,-3.795l0.184,0c0.264,1.875 0.851,2.946 1.792,3.686l1.223,0.064l0,-27.486Zm-3.259,22.286c-0.75,0.955 -1.673,1.78 -2.766,2.483c-1.095,0.699 -2.363,1.052 -3.801,1.052c-1.359,0 -2.477,-0.396 -3.358,-1.191c-0.883,-0.791 -1.322,-2.029 -1.322,-3.712c0,-1.3 0.292,-2.425 0.883,-3.368c0.585,-0.947 1.388,-1.763 2.398,-2.453c1.118,-0.733 2.322,-1.341 3.619,-1.83c1.296,-0.488 2.622,-0.934 4.347,-1.438l0,10.457Z"
-    fill="#0e9cd1" fillRule="nonzero" />
-</svg>;
-icons.infoBackground = <svg viewBox="0 0 400 400" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <rect x="49.173" y="68.261" width="301.653" height="263.478" fill="#0a6689"
-    />
-    <g fill="#fff">
-        <path d="M289.818,251.349c0,-4.195 -3.406,-7.601 -7.6,-7.601l-164.435,0c-4.194,0 -7.6,3.406 -7.6,7.601l0,3.8c0,4.195 3.406,7.601 7.6,7.601l164.435,0c4.194,0 7.6,-3.406 7.6,-7.601l0,-3.8Z"
-        />
-        <path d="M238.631,288.345c0,-3.12 -2.533,-5.652 -5.652,-5.652l-65.958,0c-3.119,0 -5.652,2.532 -5.652,5.652l0,2.825c0,3.119 2.533,5.652 5.652,5.652l65.958,0c3.119,0 5.652,-2.533 5.652,-5.652l0,-2.825Z"
-        />
-    </g>
-    <path d="M199.486,97.833l-13.692,41.077l-41.077,0l34.231,27.385l-13.693,41.077l34.231,-27.384l34.231,27.384l-13.692,-41.077l34.231,-27.385l-41.077,0l-13.693,-41.077Z"
-    fill="#fff" fillRule="nonzero" />
-</svg>;
-icons.opacity = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <g fillRule='nonzero'>
-        <path d='M14.043,7.612c2.23,2.178 2.23,5.726 0,7.904c-1.081,1.055 -2.515,1.638 -4.042,1.638c-1.527,0 -2.964,-0.583 -4.042,-1.638c-1.08,-1.056 -1.674,-2.46 -1.674,-3.952c0,-1.492 0.594,-2.896 1.676,-3.956l4.04,-4.026l4.038,4.026c0.002,0.002 0.004,0.002 0.004,0.004Z'
-        fill='url(#_Linear1)' />
-        <path d='M15.595,6.066c3.082,3.013 3.082,7.917 -0.002,10.931c-1.493,1.461 -3.479,2.265 -5.591,2.265c-2.112,0 -4.097,-0.804 -5.593,-2.265c-1.493,-1.46 -2.315,-3.402 -2.315,-5.467c0,-2.065 0.822,-4.007 2.315,-5.466l5.176,-5.162c0.221,-0.219 0.614,-0.219 0.833,0l5.177,5.164Zm-0.826,10.124c2.629,-2.568 2.629,-6.751 0,-9.32c0,-0.002 -0.003,-0.002 -0.005,-0.004l-4.762,-4.749l-4.764,4.749c-1.276,1.249 -1.976,2.905 -1.976,4.664c0,1.76 0.7,3.416 1.974,4.66c1.272,1.245 2.966,1.932 4.766,1.932c1.801,0 3.493,-0.687 4.767,-1.932Z'
-        />
-    </g>
-    <defs>
-        <linearGradient id='_Linear1' x2='1' gradientUnits='userSpaceOnUse' gradientTransform='matrix(11.4305 0 0 13.5721 4.285 10.368)'>
-            <stop offset='0' />
-            <stop offset='0.5' stopOpacity='0.749' />
-            <stop offset='1' stopOpacity='0' />
-        </linearGradient>
-    </defs>
-</svg>;
-icons.infoStart = <svg
-  xmlns="http://www.w3.org/2000/svg"
-  fillRule="evenodd"
-  strokeLinejoin="round"
-  strokeMiterlimit="2"
-  clipRule="evenodd"
-  viewBox="0 0 400 400"
->
-  <path fill="#eee" d="M46.942 44.98H353.057V355.019H46.942z"></path>
-  <path
-	fill="#565656"
-	d="M289.817 181.49a7.59 7.59 0 00-7.585-7.586H117.768a7.59 7.59 0 00-7.585 7.586v3.834a7.59 7.59 0 007.585 7.586h164.464a7.59 7.59 0 007.585-7.586v-3.834z"
-  ></path>
-  <path
-	fill="#888"
-	d="M245.908 310.489a5.593 5.593 0 00-5.59-5.59h-80.636a5.593 5.593 0 00-5.59 5.59v2.813a5.593 5.593 0 005.59 5.59h80.636a5.593 5.593 0 005.59-5.59v-2.813zM306.311 260.695a5.686 5.686 0 00-5.684-5.683H99.373a5.686 5.686 0 00-5.684 5.683v2.674a5.687 5.687 0 005.684 5.684h201.254a5.687 5.687 0 005.684-5.684v-2.674zM326.619 237.933a5.686 5.686 0 00-5.683-5.683H72.341a5.686 5.686 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.675zM326.619 215.172a5.687 5.687 0 00-5.683-5.684H72.341a5.687 5.687 0 00-5.683 5.684v2.674a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.674z"
-  ></path>
-  <path
-	fill="#888"
-	fillRule="nonzero"
-	d="M176.481 80.65h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719h-46.122a3.72 3.72 0 01-3.719-3.719V84.37a3.721 3.721 0 013.719-3.72zm43.89 47.609V86.601h-41.658v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976h-29.756a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-  ></path>
-</svg>;
-  icons.infoBasic = <svg
-  xmlns="http://www.w3.org/2000/svg"
-  fillRule="evenodd"
-  strokeLinejoin="round"
-  strokeMiterlimit="2"
-  clipRule="evenodd"
-  viewBox="0 0 400 400"
->
-  <path
-	fill="#fff"
-	d="M350.827 47.314c0-11.515-9.348-20.863-20.863-20.863H70.036c-11.515 0-20.863 9.348-20.863 20.863v305.372c0 11.515 9.348 20.863 20.863 20.863h259.928c11.515 0 20.863-9.348 20.863-20.863V47.314z"
-  ></path>
-  <path
-	fill="#eee"
-	d="M330.137 20.452l.173.001.172.003.172.004.172.005.172.006.172.007.171.008.171.009.342.022.17.012.34.028.17.016.169.017.338.036.168.02.168.021.168.022.168.023.167.024.167.025.167.026.166.027.166.028.165.029.166.03.165.031.164.032.165.033.163.035.164.035.163.036.163.037.163.038.162.039.162.04.161.041.161.042.16.043.161.044.159.045.16.046.159.047.158.047.159.049.157.05.158.05.157.052.156.052.312.108.155.055.155.056.154.057.154.058.153.059.153.06.153.061.152.061.302.126.151.065.15.065.149.066.15.067.149.068.148.068.147.07.148.07.146.071.292.146.145.073.145.075.144.076.143.076.143.077.143.078.142.079.282.16.141.082.139.082.14.083.138.083.276.17.137.086.137.087.136.088.27.178.134.09.134.091.132.092.133.092.132.093.131.094.131.095.13.096.129.096.129.097.136.103.247.192.252.201.25.203.247.207.245.209.242.213.239.215.237.218.235.22.231.224.229.226.226.229.224.232.22.234.218.236.215.24.213.242.209.245.207.247.203.25.201.252.192.248.103.135.097.129.096.129.096.13.095.131.094.131.093.132.092.133.092.132.091.134.09.134.089.135.089.136.088.135.087.137.086.137.17.276.083.138.083.14.082.139.082.141.16.282.079.142.078.142.077.144.076.143.076.144.075.145.073.145.073.146.072.146.072.147.14.294.068.149.068.148.067.149.066.15.065.15.065.151.063.151.062.151.062.152.061.153.059.152.06.154.058.154.057.154.056.154.055.156.108.312.052.156.052.157.1.316.049.158.047.158.047.159.046.159.045.16.044.161.043.16.042.161.041.162.04.161.039.162.038.163.037.163.036.163.035.163.035.165.033.164.032.164.031.166.03.164.029.166.028.166.027.166.026.167.025.167.024.167.023.168.022.168.021.168.02.168.018.168.018.17.017.17.015.169.015.17.014.17.012.17.022.342.009.17.008.172.007.172.006.172.005.172.004.171.003.173.001.173.001.163v305.392l-.001.163-.001.173-.003.173-.004.172-.005.171-.006.172-.007.172-.008.172-.009.17-.011.171-.011.171-.012.17-.014.17-.015.17-.015.169-.017.17-.018.17-.018.167-.02.169-.021.168-.022.168-.023.168-.024.167-.025.167-.026.167-.027.166-.028.166-.029.166-.03.164-.031.166-.032.164-.033.164-.035.165-.035.163-.036.163-.037.163-.038.163-.039.162-.04.161-.041.162-.042.161-.043.16-.044.161-.045.16-.046.159-.047.159-.047.158-.049.158-.05.158-.05.158-.052.157-.052.156-.054.156-.054.156-.055.155-.056.155-.057.154-.058.154-.059.154-.06.152-.061.153-.061.152-.063.151-.063.151-.065.151-.065.15-.066.15-.067.149-.068.148-.068.149-.07.147-.07.148-.071.146-.073.146-.073.146-.073.145-.075.145-.076.144-.076.143-.077.144-.078.142-.079.142-.079.141-.081.142-.082.14-.082.139-.083.14-.083.138-.085.138-.085.138-.086.137-.087.137-.088.136-.089.135-.089.135-.09.134-.091.134-.092.132-.092.133-.093.132-.094.131-.095.131-.096.13-.096.129-.097.129-.103.136-.192.247-.201.252-.203.25-.207.247-.209.245-.213.242-.215.24-.217.236-.221.235-.224.231-.226.229-.229.226-.232.224-.234.22-.236.218-.24.215-.242.213-.245.209-.247.207-.25.203-.252.201-.248.192-.135.103-.129.097-.129.096-.13.096-.131.095-.131.094-.132.093-.133.093-.132.091-.134.091-.134.09-.135.089-.135.089-.136.088-.137.087-.137.086-.138.085-.138.085-.138.083-.14.083-.139.082-.14.082-.142.081-.141.079-.142.079-.142.078-.144.077-.143.076-.144.076-.145.075-.145.074-.146.072-.146.073-.146.071-.148.07-.147.07-.149.068-.148.068-.15.067-.149.066-.149.065-.152.065-.151.063-.151.063-.152.061-.153.061-.152.06-.154.059-.154.058-.154.057-.155.056-.155.055-.156.054-.156.054-.156.052-.157.052-.158.05-.158.05-.158.049-.158.047-.159.047-.159.046-.16.045-.161.044-.16.043-.161.042-.162.041-.161.04-.162.039-.163.038-.163.037-.163.036-.163.035-.165.035-.164.033-.164.032-.166.031-.164.03-.166.029-.167.028-.165.027-.167.026-.167.025-.167.024-.168.023-.168.022-.168.021-.169.02-.167.018-.17.018-.17.017-.169.015-.17.015-.17.014-.17.012-.171.011-.171.011-.17.009-.172.008-.172.007-.172.006-.172.005-.171.004-.173.003-.173.001-.163.001H70.026l-.163-.001-.173-.001-.173-.003-.171-.004-.172-.005-.172-.006-.172-.007-.172-.008-.17-.009-.171-.011-.171-.011-.17-.012-.17-.014-.17-.015-.169-.015-.17-.017-.17-.018-.167-.018-.169-.02-.168-.021-.168-.022-.168-.023-.167-.024-.167-.025-.167-.026-.165-.027-.167-.028-.166-.029-.164-.03-.166-.031-.164-.032-.164-.033-.165-.035-.163-.035-.163-.036-.163-.037-.163-.038-.162-.039-.161-.04-.162-.041-.161-.042-.16-.043-.161-.044-.159-.045-.16-.046-.159-.047-.158-.047-.158-.049-.158-.05-.158-.05-.157-.052-.156-.052-.156-.054-.156-.054-.155-.055-.155-.056-.154-.057-.154-.058-.154-.06-.152-.059-.153-.061-.152-.061-.151-.063-.151-.063-.151-.065-.15-.065-.149-.066-.15-.067-.148-.067-.149-.069-.147-.07-.147-.07-.147-.071-.146-.073-.146-.072-.145-.074-.145-.075-.144-.076-.143-.076-.144-.077-.142-.078-.142-.079-.141-.079-.141-.081-.141-.082-.139-.082-.14-.083-.138-.083-.138-.085-.138-.085-.137-.086-.137-.087-.135-.088-.136-.089-.135-.089-.134-.09-.134-.091-.132-.091-.133-.093-.132-.093-.131-.094-.131-.095-.13-.096-.129-.096-.129-.096-.135-.104-.248-.192-.252-.201-.25-.203-.247-.207-.245-.209-.242-.213-.24-.215-.236-.218-.234-.22-.232-.224-.229-.226-.226-.229-.224-.231-.221-.235-.217-.237-.215-.239-.213-.242-.209-.245-.207-.247-.203-.25-.201-.252-.192-.247-.103-.136-.097-.129-.097-.129-.095-.13-.095-.131-.094-.131-.093-.132-.092-.132-.092-.133-.091-.134-.09-.134-.089-.135-.089-.135-.088-.136-.086-.137-.087-.137-.085-.138-.085-.138-.083-.138-.083-.14-.082-.139-.082-.141-.08-.141-.08-.141-.079-.142-.078-.143-.077-.143-.076-.143-.076-.144-.075-.145-.073-.145-.073-.146-.073-.146-.071-.146-.07-.148-.07-.147-.068-.148-.068-.149-.067-.15-.066-.149-.065-.15-.065-.151-.063-.151-.063-.151-.061-.152-.061-.153-.06-.153-.059-.153-.057-.154-.058-.154-.056-.155-.055-.155-.054-.156-.054-.156-.052-.156-.052-.157-.05-.158-.05-.157-.049-.158-.047-.159-.047-.159-.046-.16-.045-.159-.044-.161-.043-.16-.042-.161-.041-.161-.04-.162-.039-.162-.038-.163-.037-.163-.036-.163-.035-.163-.035-.164-.033-.165-.032-.164-.031-.165-.03-.166-.029-.165-.028-.166-.027-.166-.026-.167-.025-.167-.024-.167-.023-.168-.022-.168-.021-.168-.02-.168-.018-.169-.018-.169-.017-.169-.015-.17-.015-.17-.014-.17-.012-.17-.011-.171-.011-.171-.009-.171-.008-.171-.007-.172-.006-.172-.005-.171-.004-.173-.003-.172-.001-.173-.001-.163V47.304l.001-.163.001-.173.003-.172.004-.172.005-.172.006-.172.007-.172.008-.171.009-.171.011-.171.011-.171.012-.17.014-.17.015-.17.015-.17.017-.169.018-.169.018-.169.02-.168.021-.168.022-.168.023-.168.024-.167.025-.167.026-.167.027-.166.028-.166.029-.165.03-.166.031-.165.032-.164.033-.165.035-.164.035-.163.036-.163.037-.163.038-.163.039-.162.04-.162.041-.161.042-.161.043-.16.044-.161.045-.159.046-.16.047-.159.047-.158.049-.159.05-.157.05-.158.052-.157.052-.156.054-.156.054-.156.055-.155.056-.155.057-.154.058-.154.059-.153.06-.153.061-.153.061-.152.063-.151.063-.151.065-.151.065-.15.066-.15.067-.149.068-.148.068-.149.07-.147.07-.147.072-.147.072-.146.073-.146.073-.145.075-.145.076-.144.076-.143.077-.143.078-.143.079-.142.08-.141.08-.141.082-.141.082-.139.083-.14.083-.138.085-.139.085-.137.086-.137.087-.137.088-.136.089-.135.089-.135.09-.134.091-.133.092-.134.092-.132.093-.132.094-.131.095-.131.095-.13.097-.129.097-.129.103-.136.192-.247.201-.252.203-.25.207-.247.209-.245.213-.242.215-.239.218-.237.22-.235.224-.231.226-.229.229-.226.231-.224.235-.22.237-.218.239-.215.242-.213.245-.209.247-.207.25-.203.252-.201.248-.192.135-.103.129-.097.129-.097.13-.095.131-.095.131-.094.132-.093.132-.092.134-.092.133-.091.134-.09.135-.089.135-.089.136-.088.137-.087.137-.086.137-.085.139-.085.138-.083.14-.083.139-.083.141-.081.141-.08.141-.08.142-.079.143-.078.143-.077.143-.076.144-.076.145-.075.145-.073.146-.073.146-.073.147-.071.147-.07.147-.07.149-.068.148-.068.149-.067.15-.066.15-.065.151-.065.151-.063.151-.063.152-.061.153-.061.153-.06.153-.059.154-.058.154-.057.155-.056.155-.055.156-.054.156-.054.156-.052.157-.052.158-.05.157-.05.159-.049.158-.047.159-.047.16-.046.159-.045.161-.044.16-.043.161-.042.162-.041.161-.04.162-.039.163-.038.163-.037.163-.036.163-.035.164-.035.165-.033.164-.032.165-.031.166-.03.165-.029.166-.028.167-.027.166-.026.167-.025.167-.024.168-.023.168-.022.168-.021.168-.02.169-.018.169-.018.169-.017.17-.016.17-.014.17-.014.17-.012.171-.011.171-.011.171-.009.171-.008.172-.007.172-.006.172-.005.172-.004.172-.003.173-.001.163-.001h259.948l.163.001zM70.036 26.451l-.135.001-.135.001-.134.002-.134.003-.134.004-.134.005-.134.005-.133.006-.133.008-.133.008-.133.008-.132.01-.133.011-.132.011-.131.012-.132.013-.131.014-.131.014-.131.016-.131.016-.13.017-.13.018-.13.018-.13.02-.129.02-.129.021-.129.022-.128.022-.128.024-.128.024-.128.025-.127.025-.128.027-.126.027-.127.028-.126.029-.126.029-.126.031-.125.031-.125.032-.125.032-.125.033-.124.034-.124.035-.123.036-.124.036-.123.037-.122.038-.122.038-.122.04-.122.039-.121.041-.121.042-.121.042-.12.042-.12.044-.12.044-.119.045-.119.046-.119.046-.118.047-.118.048-.117.049-.118.049-.116.05-.117.05-.116.051-.116.052-.115.053-.115.053-.114.054-.115.055-.113.055-.114.056-.113.057-.112.057-.113.058-.112.059-.111.059-.111.06-.111.06-.11.062-.11.061-.109.063-.109.063-.109.064-.108.064-.108.065-.107.066-.107.066-.106.067-.106.068-.106.068-.105.069-.105.069-.104.07-.104.071-.103.071-.103.072-.103.072-.102.073-.101.074-.102.074-.1.075-.1.076-.1.076-.198.153-.196.156-.194.159-.193.16-.19.163-.188.165-.187.168-.184.169-.182.172-.18.174-.178.176-.176.178-.174.18-.172.182-.169.184-.168.187-.165.188-.163.19-.16.192-.159.195-.156.196-.153.198-.076.1-.076.1-.075.1-.074.101-.073.102-.074.102-.072.103-.072.103-.071.103-.071.104-.07.104-.069.105-.069.105-.068.106-.068.106-.067.106-.066.107-.066.107-.065.108-.064.108-.064.109-.063.109-.063.109-.061.11-.062.11-.06.111-.06.111-.059.111-.059.112-.058.112-.057.113-.057.113-.056.114-.055.113-.055.115-.054.114-.053.115-.053.115-.052.116-.051.116-.05.117-.05.116-.049.118-.049.117-.048.118-.047.118-.046.119-.046.119-.045.119-.044.12-.043.12-.043.12-.042.121-.042.121-.04.121-.04.122-.04.121-.038.123-.038.122-.037.123-.036.124-.036.123-.035.124-.034.124-.033.125-.032.125-.032.125-.031.125-.031.126-.029.126-.029.126-.028.127-.027.126-.027.128-.025.127-.025.128-.024.128-.024.128-.022.128-.022.129-.021.129-.02.129-.02.13-.018.13-.018.13-.017.13-.016.131-.016.131-.014.131-.014.131-.013.132-.012.131-.011.132-.011.133-.01.132-.008.133-.008.133-.008.133-.006.133-.005.134-.005.134-.004.134-.003.134-.002.134-.001.135-.001.135v305.372l.002.27.002.134.003.134.004.134.01.268.014.266.016.266.021.265.023.263.027.263.03.262.033.261.036.26.04.259.043.258.046.256.049.256.052.255.055.253.058.252.062.251.064.25.067.249.071.247.073.247.076.245.08.243.082.242.085.241.087.24.091.238.093.237.097.235.099.234.101.233.105.231.107.229.11.228.113.227.115.225.118.223.12.222.123.22.126.218.128.217.131.215.133.213.136.212.138.21.141.208.143.206.146.205.147.203.151.2.076.1.153.198.156.196.159.195.16.192.163.19.165.188.168.187.169.184.172.182.174.18.176.178.178.176.18.174.182.172.184.169.187.168.188.165.19.163.193.16.194.159.196.156.198.153.2.152.202.149.203.147.206.144.207.142.209.139.211.137.212.135.214.132.216.129.218.127.219.124.221.122.222.119.225.117.225.114.227.111.229.109.23.106.232.103.233.1.235.098.236.095.238.092.239.089.24.086.242.084.243.08.244.078.245.075.247.072.248.069.25.065.25.063.252.06.253.057.254.054.255.05.256.048.257.044.258.041.26.038.26.035.262.032.262.028.263.025.265.022.265.018.266.016.267.011.134.005.134.004.134.003.134.002.27.002h259.928l.135-.001.135-.001.134-.002.134-.003.134-.004.134-.005.134-.005.133-.006.133-.008.133-.008.133-.008.132-.01.133-.011.132-.011.131-.012.132-.013.131-.014.131-.014.131-.016.131-.016.13-.017.13-.018.13-.018.13-.02.129-.02.129-.021.129-.022.128-.022.128-.024.128-.024.128-.025.127-.025.128-.027.126-.027.127-.028.126-.029.126-.029.126-.031.125-.031.125-.032.125-.032.125-.033.124-.034.124-.035.123-.036.124-.036.123-.037.122-.038.122-.038.122-.04.122-.039.121-.041.121-.042.121-.042.12-.042.12-.044.12-.044.119-.045.119-.046.119-.046.118-.047.118-.048.117-.049.118-.049.116-.05.117-.05.116-.051.116-.052.115-.053.115-.053.114-.054.115-.055.113-.055.114-.056.113-.057.112-.057.113-.058.112-.059.111-.059.111-.06.111-.06.11-.062.11-.061.109-.063.109-.063.109-.064.108-.064.108-.065.107-.066.107-.066.106-.067.106-.068.106-.068.105-.069.105-.069.104-.07.104-.071.103-.071.103-.072.103-.072.102-.073.101-.074.102-.074.1-.075.1-.076.1-.076.198-.153.196-.156.194-.159.193-.16.19-.163.188-.165.187-.168.184-.169.182-.172.18-.174.178-.176.176-.178.174-.18.172-.182.169-.184.168-.187.165-.188.163-.19.16-.192.159-.195.156-.196.153-.198.076-.1.076-.1.075-.1.074-.101.073-.102.074-.102.072-.103.072-.103.071-.103.071-.104.07-.104.069-.105.069-.105.068-.106.068-.106.067-.106.066-.107.066-.107.065-.108.064-.108.064-.109.063-.109.063-.109.061-.11.062-.11.06-.111.06-.111.059-.111.059-.112.058-.112.057-.113.057-.113.056-.114.055-.113.055-.115.054-.114.053-.115.053-.115.052-.116.051-.116.05-.117.05-.116.049-.118.049-.117.048-.118.047-.118.046-.119.046-.119.045-.119.044-.12.043-.12.043-.12.042-.121.042-.121.04-.121.04-.122.04-.121.038-.123.038-.122.037-.123.036-.124.036-.123.035-.124.034-.124.033-.125.032-.125.032-.125.031-.125.031-.126.029-.126.029-.126.028-.127.027-.126.027-.128.025-.127.025-.128.024-.128.024-.128.022-.128.022-.129.021-.129.02-.129.02-.13.018-.13.018-.13.017-.13.016-.131.016-.131.014-.131.014-.131.013-.132.012-.131.011-.132.011-.133.01-.132.008-.133.008-.133.008-.133.006-.133.005-.134.005-.134.004-.134.003-.134.002-.134.001-.135.001-.135V47.314l-.001-.135-.001-.135-.002-.134-.003-.134-.004-.134-.005-.134-.005-.134-.006-.133-.008-.133-.008-.133-.008-.133-.01-.132-.011-.133-.011-.132-.012-.131-.013-.132-.014-.131-.014-.131-.016-.131-.016-.131-.017-.13-.018-.13-.018-.13-.02-.13-.02-.129-.021-.129-.022-.129-.022-.128-.024-.128-.024-.128-.025-.128-.025-.127-.027-.128-.027-.126-.028-.127-.029-.126-.029-.126-.031-.126-.031-.125-.032-.125-.032-.125-.033-.125-.034-.124-.035-.124-.036-.123-.036-.124-.037-.123-.038-.122-.038-.123-.04-.121-.04-.122-.04-.121-.042-.121-.042-.121-.043-.12-.043-.12-.044-.12-.045-.119-.046-.119-.046-.119-.047-.118-.048-.118-.049-.117-.049-.118-.05-.116-.05-.117-.051-.116-.052-.116-.053-.115-.053-.115-.054-.114-.055-.115-.055-.113-.056-.114-.057-.113-.057-.113-.058-.112-.059-.112-.059-.111-.06-.111-.06-.111-.062-.11-.061-.11-.063-.109-.063-.109-.064-.109-.064-.108-.065-.108-.066-.107-.066-.107-.067-.106-.068-.106-.068-.106-.069-.105-.069-.105-.07-.104-.071-.104-.071-.103-.072-.103-.072-.103-.074-.102-.073-.102-.074-.101-.075-.1-.076-.1-.076-.1-.153-.198-.156-.196-.159-.195-.16-.192-.163-.19-.165-.188-.168-.187-.169-.184-.172-.182-.174-.18-.176-.178-.178-.176-.18-.174-.182-.172-.184-.169-.187-.168-.188-.165-.19-.163-.193-.16-.194-.159-.196-.156-.198-.153-.1-.076-.1-.076-.1-.075-.102-.074-.101-.074-.102-.073-.103-.072-.103-.072-.103-.071-.104-.071-.104-.07-.105-.069-.105-.069-.106-.068-.106-.068-.106-.067-.107-.066-.107-.066-.108-.065-.108-.064-.109-.064-.109-.063-.109-.063-.11-.061-.11-.062-.111-.06-.111-.06-.111-.059-.112-.059-.113-.058-.112-.057-.113-.057-.114-.056-.113-.055-.115-.055-.114-.054-.115-.053-.115-.053-.116-.052-.116-.051-.117-.05-.116-.05-.118-.049-.117-.049-.118-.048-.118-.047-.119-.046-.119-.046-.119-.045-.12-.044-.12-.044-.12-.042-.121-.042-.121-.042-.121-.041-.122-.039-.122-.04-.122-.038-.122-.038-.123-.037-.124-.036-.123-.036-.124-.035-.124-.034-.125-.033-.125-.032-.125-.032-.125-.031-.126-.031-.126-.029-.126-.029-.127-.028-.126-.027-.128-.027-.127-.025-.128-.025-.128-.024-.128-.024-.128-.022-.129-.022-.129-.021-.129-.02-.13-.02-.13-.018-.13-.018-.13-.017-.131-.016-.131-.016-.131-.014-.131-.014-.132-.013-.131-.012-.132-.011-.133-.011-.132-.01-.133-.008-.133-.008-.133-.008-.133-.006-.134-.005-.134-.005-.134-.004-.134-.003-.134-.002-.135-.001-.135-.001H70.036z"
-  ></path>
-  <circle cx="200" cy="119.069" r="70.019" fill="#eee"></circle>
-  <path
-	fill="#565656"
-	d="M289.817 213.579a7.59 7.59 0 00-7.585-7.586H117.768a7.59 7.59 0 00-7.585 7.586v3.834a7.59 7.59 0 007.585 7.586h164.464a7.59 7.59 0 007.585-7.586v-3.834z"
-  ></path>
-  <path
-	fill="#888"
-	d="M245.908 342.578a5.593 5.593 0 00-5.59-5.59h-80.636a5.593 5.593 0 00-5.59 5.59v2.813a5.593 5.593 0 005.59 5.59h80.636a5.593 5.593 0 005.59-5.59v-2.813zM306.311 292.784a5.686 5.686 0 00-5.684-5.683H99.373a5.686 5.686 0 00-5.684 5.683v2.674a5.687 5.687 0 005.684 5.684h201.254a5.687 5.687 0 005.684-5.684v-2.674zM326.619 270.022a5.686 5.686 0 00-5.683-5.683H72.341a5.686 5.686 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.675zM326.619 247.261a5.687 5.687 0 00-5.683-5.684H72.341a5.687 5.687 0 00-5.683 5.684v2.674a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.674z"
-  ></path>
-  <path
-	fill="#888"
-	fillRule="nonzero"
-	d="M176.939 94.182h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719h-46.122a3.72 3.72 0 01-3.719-3.719V97.902a3.721 3.721 0 013.719-3.72zm43.89 47.609v-41.658h-41.658v41.658h41.658zM200 109.06c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976h-29.756a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-  ></path>
-</svg>;
-icons.infoLeftAbove = <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fillRule="evenodd"
-      strokeLinejoin="round"
-      strokeMiterlimit="2"
-      clipRule="evenodd"
-      viewBox="0 0 400 400"
-    >
-      <path
-        fill="#fff"
-        d="M7.03 60.607H392.61699999999996V345.77200000000005H7.03z"
-      ></path>
-      <path
-        fill="#eee"
-        d="M392.617 345.772H7.03V60.607h385.587v285.165zM12.03 65.607v275.165h375.587V65.607H12.03z"
-      ></path>
-      <path
-        fill="#eee"
-        d="M29.352 81.149H151.446V196.28300000000002H29.352z"
-      ></path>
-      <path
-        fill="#888"
-        fillRule="nonzero"
-        d="M67.338 111.82h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719H67.338a3.72 3.72 0 01-3.719-3.719V115.54a3.721 3.721 0 013.719-3.72zm43.89 47.609v-41.658H69.57v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976H75.521a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-      ></path>
-      <path
-        fill="#565656"
-        d="M299.746 222.29a7.59 7.59 0 00-7.585-7.586H35.651a7.589 7.589 0 00-7.585 7.586v3.834a7.588 7.588 0 007.585 7.585h256.51a7.588 7.588 0 007.585-7.585v-3.834z"
-      ></path>
-      <path
-        fill="#888"
-        d="M372.406 285.764a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h333.112a5.687 5.687 0 005.683-5.684v-2.674zM266.41 313.012a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h227.116a5.686 5.686 0 005.683-5.683v-2.675zM370.188 257.889a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h330.894a5.687 5.687 0 005.683-5.684v-2.674z"
-      ></path>
-	</svg>;
-icons.infoLeft = <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 400 400"
->
-<path
-  fill="#fff"
-  stroke="#eee"
-  strokeWidth="5"
-  d="M390.993 128.492c0-6.336-5.144-11.479-11.479-11.479H20.486c-6.335 0-11.479 5.143-11.479 11.479v143.016c0 6.336 5.144 11.479 11.479 11.479h359.028c6.335 0 11.479-5.143 11.479-11.479V128.492z"
-></path>
-<circle cx="78.187" cy="200" r="59.757" fill="#eee"></circle>
-<path
-  fill="#888"
-  fillRule="nonzero"
-  d="M55.126 173.22h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719H55.126a3.72 3.72 0 01-3.719-3.719V176.94a3.721 3.721 0 013.719-3.72zm43.89 47.609v-41.658H57.358v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976H63.309a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-></path>
-<path
-  fill="#565656"
-  d="M335.948 154.253a7.59 7.59 0 00-7.585-7.586H163.899a7.589 7.589 0 00-7.585 7.586v3.834a7.588 7.588 0 007.585 7.585h164.464a7.588 7.588 0 007.585-7.585v-3.834z"
-></path>
-<path
-  fill="#888"
-  d="M368.797 217.727a5.686 5.686 0 00-5.683-5.683H161.859a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h201.255a5.687 5.687 0 005.683-5.684v-2.674zM303.84 244.975a5.686 5.686 0 00-5.683-5.683H161.859a5.685 5.685 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h136.298a5.686 5.686 0 005.683-5.683v-2.675zM368.797 189.852a5.686 5.686 0 00-5.683-5.683H161.859a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h201.255a5.687 5.687 0 005.683-5.684v-2.674z"
-></path>
-</svg>;
-icons.infoTopOverlay = <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 400 400"
->
-<path
-  fill="#fff"
-  d="M350.827 132.614c0-11.515-9.348-20.863-20.863-20.863H70.036c-11.515 0-20.863 9.348-20.863 20.863v209.179c0 11.515 9.348 20.863 20.863 20.863h259.928c11.515 0 20.863-9.348 20.863-20.863V132.614z"
-></path>
-<path
-  fill="#eee"
-  d="M330.137 106.007l.173.001.172.003.172.004.172.005.172.006.172.007.171.008.171.009.342.022.17.012.34.028.17.016.169.017.338.036.168.02.168.021.168.022.168.023.167.024.167.025.167.026.166.027.166.028.165.029.166.03.165.031.164.032.165.033.163.035.164.035.163.036.163.037.163.038.162.039.162.04.161.041.161.042.16.043.161.044.159.045.16.046.159.047.158.047.159.049.157.05.158.05.157.052.156.052.312.108.155.055.155.056.154.057.154.058.153.059.153.06.153.061.152.061.302.126.151.065.15.065.149.066.15.067.149.068.148.068.147.07.148.07.146.071.292.146.145.073.145.075.144.076.143.076.143.077.143.078.142.079.282.16.141.082.139.082.14.083.138.083.276.17.137.086.137.087.136.088.27.178.134.09.134.091.132.092.133.092.132.093.131.094.131.095.13.096.129.096.129.097.136.103.247.192.252.201.25.203.247.207.245.209.242.213.239.215.237.218.235.22.231.224.229.226.226.229.224.232.22.234.218.236.215.24.213.242.209.245.207.247.203.25.201.252.192.248.103.135.097.129.096.129.096.13.095.131.094.131.093.132.092.133.092.132.091.134.09.134.089.135.089.136.088.135.087.137.086.137.17.276.083.138.083.14.082.139.082.141.16.282.079.142.078.142.077.144.076.143.076.144.075.145.073.145.073.146.072.146.072.147.14.294.068.149.068.148.067.149.066.15.065.15.065.151.063.151.062.151.062.152.061.153.059.152.06.154.058.154.057.154.056.154.055.156.108.312.052.156.052.157.1.316.049.158.047.158.047.159.046.159.045.16.044.161.043.16.042.161.041.162.04.161.039.162.038.163.037.163.036.163.035.163.035.165.033.164.032.164.031.166.03.164.029.166.028.166.027.166.026.167.025.167.024.167.023.168.022.168.021.168.02.168.018.168.018.17.017.17.015.169.015.17.014.17.012.17.022.342.009.17.008.172.007.172.006.172.005.172.004.171.003.173.001.173.001.163v208.944l-.001.163-.001.173-.003.173-.004.172-.005.171-.006.172-.007.172-.008.172-.009.17-.011.171-.011.171-.012.17-.014.17-.015.17-.015.169-.017.17-.018.17-.018.167-.02.169-.021.168-.022.168-.023.168-.024.167-.025.167-.026.167-.027.166-.028.166-.029.166-.03.164-.031.166-.032.164-.033.164-.035.165-.035.163-.036.163-.037.163-.038.163-.039.162-.04.161-.041.162-.042.161-.043.16-.044.161-.045.16-.046.159-.047.159-.047.158-.049.158-.05.158-.05.158-.052.157-.052.156-.054.156-.054.156-.055.155-.056.155-.057.154-.058.154-.059.154-.06.152-.061.153-.061.152-.063.151-.063.151-.065.151-.065.15-.066.15-.067.149-.068.148-.068.149-.07.147-.07.148-.071.146-.073.146-.073.146-.073.145-.075.145-.076.144-.076.143-.077.144-.078.142-.079.142-.079.141-.081.142-.082.14-.082.139-.083.14-.083.138-.085.138-.085.138-.086.137-.087.137-.088.136-.089.135-.089.135-.09.134-.091.134-.092.132-.092.133-.093.132-.094.131-.095.131-.096.13-.096.129-.097.129-.103.136-.192.247-.201.252-.203.25-.207.247-.209.245-.213.242-.215.24-.217.236-.221.235-.224.231-.226.229-.229.226-.232.224-.234.22-.236.218-.24.215-.242.213-.245.209-.247.207-.25.203-.252.201-.248.192-.135.103-.129.097-.129.096-.13.096-.131.095-.131.094-.132.093-.133.093-.132.091-.134.091-.134.09-.135.089-.135.089-.136.088-.137.087-.137.086-.138.085-.138.085-.138.083-.14.083-.139.082-.14.082-.142.081-.141.079-.142.079-.142.078-.144.077-.143.076-.144.076-.145.075-.145.074-.146.072-.146.073-.146.071-.148.07-.147.07-.149.068-.148.068-.15.067-.149.066-.149.065-.152.065-.151.063-.151.063-.152.061-.153.061-.152.06-.154.059-.154.058-.154.057-.155.056-.155.055-.156.054-.156.054-.156.052-.157.052-.158.05-.158.05-.158.049-.158.047-.159.047-.159.046-.16.045-.161.044-.16.043-.161.042-.162.041-.161.04-.162.039-.163.038-.163.037-.163.036-.163.035-.165.035-.164.033-.164.032-.166.031-.164.03-.166.029-.167.028-.165.027-.167.026-.167.025-.167.024-.168.023-.168.022-.168.021-.169.02-.167.018-.17.018-.17.017-.169.015-.17.015-.17.014-.17.012-.171.011-.171.011-.17.009-.172.008-.172.007-.172.006-.172.005-.171.004-.173.003-.173.001-.163.001H70.026l-.163-.001-.173-.001-.173-.003-.171-.004-.172-.005-.172-.006-.172-.007-.172-.008-.17-.009-.171-.011-.171-.011-.17-.012-.17-.014-.17-.015-.169-.015-.17-.017-.17-.018-.167-.018-.169-.02-.168-.021-.168-.022-.168-.023-.167-.024-.167-.025-.167-.026-.165-.027-.167-.028-.166-.029-.164-.03-.166-.031-.164-.032-.164-.033-.165-.035-.163-.035-.163-.036-.163-.037-.163-.038-.162-.039-.161-.04-.162-.041-.161-.042-.16-.043-.161-.044-.159-.045-.16-.046-.159-.047-.158-.047-.158-.049-.158-.05-.158-.05-.157-.052-.156-.052-.156-.054-.156-.054-.155-.055-.155-.056-.154-.057-.154-.058-.154-.06-.152-.059-.153-.061-.152-.061-.151-.063-.151-.063-.151-.065-.15-.065-.149-.066-.15-.067-.148-.067-.149-.069-.147-.07-.147-.07-.147-.071-.146-.073-.146-.072-.145-.074-.145-.075-.144-.076-.143-.076-.144-.077-.142-.078-.142-.079-.141-.079-.141-.081-.141-.082-.139-.082-.14-.083-.138-.083-.138-.085-.138-.085-.137-.086-.137-.087-.135-.088-.136-.089-.135-.089-.134-.09-.134-.091-.132-.091-.133-.093-.132-.093-.131-.094-.131-.095-.13-.096-.129-.096-.129-.096-.135-.104-.248-.192-.252-.201-.25-.203-.247-.207-.245-.209-.242-.213-.24-.215-.236-.218-.234-.22-.232-.224-.229-.226-.226-.229-.224-.231-.221-.235-.217-.237-.215-.239-.213-.242-.209-.245-.207-.247-.203-.25-.201-.252-.192-.247-.103-.136-.097-.129-.097-.129-.095-.13-.095-.131-.094-.131-.093-.132-.092-.132-.092-.133-.091-.134-.09-.134-.089-.135-.089-.135-.088-.136-.086-.137-.087-.137-.085-.138-.085-.138-.083-.138-.083-.14-.082-.139-.082-.141-.08-.141-.08-.141-.079-.142-.078-.143-.077-.143-.076-.143-.076-.144-.075-.145-.073-.145-.073-.146-.073-.146-.071-.146-.07-.148-.07-.147-.068-.148-.068-.149-.067-.15-.066-.149-.065-.15-.065-.151-.063-.151-.063-.151-.061-.152-.061-.153-.06-.153-.059-.153-.057-.154-.058-.154-.056-.155-.055-.155-.054-.156-.054-.156-.052-.156-.052-.157-.05-.158-.05-.157-.049-.158-.047-.159-.047-.159-.046-.16-.045-.159-.044-.161-.043-.16-.042-.161-.041-.161-.04-.162-.039-.162-.038-.163-.037-.163-.036-.163-.035-.163-.035-.164-.033-.165-.032-.164-.031-.165-.03-.166-.029-.165-.028-.166-.027-.166-.026-.167-.025-.167-.024-.167-.023-.168-.022-.168-.021-.168-.02-.168-.018-.169-.018-.169-.017-.169-.015-.17-.015-.17-.014-.17-.012-.17-.011-.171-.011-.171-.009-.171-.008-.171-.007-.172-.006-.172-.005-.171-.004-.173-.003-.172-.001-.173-.001-.163V132.859l.001-.163.001-.173.003-.172.004-.172.005-.172.006-.172.007-.172.008-.171.009-.171.011-.171.011-.171.012-.17.014-.17.015-.17.015-.17.017-.169.018-.169.018-.169.02-.168.021-.168.022-.168.023-.168.024-.167.025-.167.026-.167.027-.166.028-.166.029-.165.03-.166.031-.165.032-.164.033-.165.035-.164.035-.163.036-.163.037-.163.038-.163.039-.162.04-.162.041-.161.042-.161.043-.16.044-.161.045-.159.046-.16.047-.159.047-.158.049-.159.05-.157.05-.158.052-.157.052-.156.054-.156.054-.156.055-.155.056-.155.057-.154.058-.154.059-.153.06-.153.061-.153.061-.152.063-.151.063-.151.065-.151.065-.15.066-.15.067-.149.068-.148.068-.149.07-.147.07-.147.072-.147.072-.146.073-.146.073-.145.075-.145.076-.144.076-.143.077-.143.078-.143.079-.142.08-.141.08-.141.082-.141.082-.139.083-.14.083-.138.085-.139.085-.137.086-.137.087-.137.088-.136.089-.135.089-.135.09-.134.091-.133.092-.134.092-.132.093-.132.094-.131.095-.131.095-.13.097-.129.097-.129.103-.136.192-.247.201-.252.203-.25.207-.247.209-.245.213-.242.215-.239.218-.237.22-.235.224-.231.226-.229.229-.226.231-.224.235-.22.237-.218.239-.215.242-.213.245-.209.247-.207.25-.203.252-.201.248-.192.135-.103.129-.097.129-.097.13-.095.131-.095.131-.094.132-.093.132-.092.134-.092.133-.091.134-.09.135-.089.135-.089.136-.088.137-.087.137-.086.137-.085.139-.085.138-.083.14-.083.139-.083.141-.081.141-.08.141-.08.142-.079.143-.078.143-.077.143-.076.144-.076.145-.075.145-.073.146-.073.146-.073.147-.071.147-.07.147-.07.149-.068.148-.068.149-.067.15-.066.15-.065.151-.065.151-.063.151-.063.152-.061.153-.061.153-.06.153-.059.154-.058.154-.057.155-.056.155-.055.156-.054.156-.054.156-.052.157-.052.158-.05.157-.05.159-.049.158-.047.159-.047.16-.046.159-.045.161-.044.16-.043.161-.042.162-.041.161-.04.162-.039.163-.038.163-.037.163-.036.163-.035.164-.035.165-.033.164-.032.165-.031.166-.03.165-.029.166-.028.167-.027.166-.026.167-.025.167-.024.168-.023.168-.022.168-.021.168-.02.169-.018.169-.018.169-.017.17-.016.17-.014.17-.014.17-.012.171-.011.171-.011.171-.009.171-.008.172-.007.172-.006.172-.005.172-.004.172-.003.173-.001.163-.001h259.948l.163.001zm-260.101 5.999l-.135.001-.135.001-.134.002-.134.003-.134.004-.134.005-.134.005-.133.006-.133.008-.133.008-.133.008-.132.01-.133.011-.132.011-.131.012-.132.013-.131.014-.131.014-.131.016-.131.016-.13.017-.13.018-.13.018-.13.02-.129.02-.129.021-.129.022-.128.022-.128.024-.128.024-.128.025-.127.025-.128.027-.126.027-.127.028-.126.029-.126.029-.126.031-.125.031-.125.032-.125.032-.125.033-.124.034-.124.035-.123.036-.124.036-.123.037-.122.038-.122.038-.122.04-.122.039-.121.041-.121.042-.121.042-.12.042-.12.044-.12.044-.119.045-.119.046-.119.046-.118.047-.118.048-.117.049-.118.049-.116.05-.117.05-.116.051-.116.052-.115.053-.115.053-.114.054-.115.055-.113.055-.114.056-.113.057-.112.057-.113.058-.112.059-.111.059-.111.06-.111.06-.11.062-.11.061-.109.063-.109.063-.109.064-.108.064-.108.065-.107.066-.107.066-.106.067-.106.068-.106.068-.105.069-.105.069-.104.07-.104.071-.103.071-.103.072-.103.072-.102.073-.101.074-.102.074-.1.075-.1.076-.1.076-.198.153-.196.156-.194.159-.193.16-.19.163-.188.165-.187.168-.184.169-.182.172-.18.174-.178.176-.176.178-.174.18-.172.182-.169.184-.168.187-.165.188-.163.19-.16.192-.159.195-.156.196-.153.198-.076.1-.076.1-.075.1-.074.101-.073.102-.074.102-.072.103-.072.103-.071.103-.071.104-.07.104-.069.105-.069.105-.068.106-.068.106-.067.106-.066.107-.066.107-.065.108-.064.108-.064.109-.063.109-.063.109-.061.11-.062.11-.06.111-.06.111-.059.111-.059.112-.058.112-.057.113-.057.113-.056.114-.055.113-.055.115-.054.114-.053.115-.053.115-.052.116-.051.116-.05.117-.05.116-.049.118-.049.117-.048.118-.047.118-.046.119-.046.119-.045.119-.044.12-.043.12-.043.12-.042.121-.042.121-.04.121-.04.122-.04.121-.038.123-.038.122-.037.123-.036.124-.036.123-.035.124-.034.124-.033.125-.032.125-.032.125-.031.125-.031.126-.029.126-.029.126-.028.127-.027.126-.027.128-.025.127-.025.128-.024.128-.024.128-.022.128-.022.129-.021.129-.02.129-.02.13-.018.13-.018.13-.017.13-.016.131-.016.131-.014.131-.014.131-.013.132-.012.131-.011.132-.011.133-.01.132-.008.133-.008.133-.008.133-.006.133-.005.134-.005.134-.004.134-.003.134-.002.134-.001.135-.001.135v208.924l.002.27.002.134.003.134.004.134.01.268.014.266.016.266.021.265.023.263.027.263.03.262.033.261.036.26.04.259.043.258.046.256.049.256.052.255.055.253.058.252.062.251.064.25.067.249.071.247.073.247.076.245.08.243.082.242.085.241.087.24.091.238.093.237.097.235.099.234.101.233.105.231.107.229.11.228.113.227.115.225.118.223.12.222.123.22.126.218.128.217.131.215.133.213.136.212.138.21.141.208.143.206.146.205.147.203.151.2.076.1.153.198.156.196.159.195.16.192.163.19.165.188.168.187.169.184.172.182.174.18.176.178.178.176.18.174.182.172.184.169.187.168.188.165.19.163.193.16.194.159.196.156.198.153.2.152.202.149.203.147.206.144.207.142.209.139.211.137.212.135.214.132.216.129.218.127.219.124.221.122.222.119.225.117.225.114.227.111.229.109.23.106.232.103.233.1.235.098.236.095.238.092.239.089.24.086.242.084.243.08.244.078.245.075.247.072.248.069.25.065.25.063.252.06.253.057.254.054.255.05.256.048.257.044.258.041.26.038.26.035.262.032.262.028.263.025.265.022.265.018.266.016.267.011.134.005.134.004.134.003.134.002.27.002h259.928l.135-.001.135-.001.134-.002.134-.003.134-.004.134-.005.134-.005.133-.006.133-.008.133-.008.133-.008.132-.01.133-.011.132-.011.131-.012.132-.013.131-.014.131-.014.131-.016.131-.016.13-.017.13-.018.13-.018.13-.02.129-.02.129-.021.129-.022.128-.022.128-.024.128-.024.128-.025.127-.025.128-.027.126-.027.127-.028.126-.029.126-.029.126-.031.125-.031.125-.032.125-.032.125-.033.124-.034.124-.035.123-.036.124-.036.123-.037.122-.038.122-.038.122-.04.122-.039.121-.041.121-.042.121-.042.12-.042.12-.044.12-.044.119-.045.119-.046.119-.046.118-.047.118-.048.117-.049.118-.049.116-.05.117-.05.116-.051.116-.052.115-.053.115-.053.114-.054.115-.055.113-.055.114-.056.113-.057.112-.057.113-.058.112-.059.111-.059.111-.06.111-.06.11-.062.11-.061.109-.063.109-.063.109-.064.108-.064.108-.065.107-.066.107-.066.106-.067.106-.068.106-.068.105-.069.105-.069.104-.07.104-.071.103-.071.103-.072.103-.072.102-.073.101-.074.102-.074.1-.075.1-.076.1-.076.198-.153.196-.156.194-.159.193-.16.19-.163.188-.165.187-.168.184-.169.182-.172.18-.174.178-.176.176-.178.174-.18.172-.182.169-.184.168-.187.165-.188.163-.19.16-.192.159-.195.156-.196.153-.198.076-.1.076-.1.075-.1.074-.101.073-.102.074-.102.072-.103.072-.103.071-.103.071-.104.07-.104.069-.105.069-.105.068-.106.068-.106.067-.106.066-.107.066-.107.065-.108.064-.108.064-.109.063-.109.063-.109.061-.11.062-.11.06-.111.06-.111.059-.111.059-.112.058-.112.057-.113.057-.113.056-.114.055-.113.055-.115.054-.114.053-.115.053-.115.052-.116.051-.116.05-.117.05-.116.049-.118.049-.117.048-.118.047-.118.046-.119.046-.119.045-.119.044-.12.043-.12.043-.12.042-.121.042-.121.04-.121.04-.122.04-.121.038-.123.038-.122.037-.123.036-.124.036-.123.035-.124.034-.124.033-.125.032-.125.032-.125.031-.125.031-.126.029-.126.029-.126.028-.127.027-.126.027-.128.025-.127.025-.128.024-.128.024-.128.022-.128.022-.129.021-.129.02-.129.02-.13.018-.13.018-.13.017-.13.016-.131.016-.131.014-.131.014-.131.013-.132.012-.131.011-.132.011-.133.01-.132.008-.133.008-.133.008-.133.006-.133.005-.134.005-.134.004-.134.003-.134.002-.134.001-.135.001-.135V132.869l-.001-.135-.001-.135-.002-.134-.003-.134-.004-.134-.005-.134-.005-.134-.006-.133-.008-.133-.008-.133-.008-.133-.01-.132-.011-.133-.011-.132-.012-.131-.013-.132-.014-.131-.014-.131-.016-.131-.016-.131-.017-.13-.018-.13-.018-.13-.02-.13-.02-.129-.021-.129-.022-.129-.022-.128-.024-.128-.024-.128-.025-.128-.025-.127-.027-.128-.027-.126-.028-.127-.029-.126-.029-.126-.031-.126-.031-.125-.032-.125-.032-.125-.033-.125-.034-.124-.035-.124-.036-.123-.036-.124-.037-.123-.038-.122-.038-.123-.04-.121-.04-.122-.04-.121-.042-.121-.042-.121-.043-.12-.043-.12-.044-.12-.045-.119-.046-.119-.046-.119-.047-.118-.048-.118-.049-.117-.049-.118-.05-.116-.05-.117-.051-.116-.052-.116-.053-.115-.053-.115-.054-.114-.055-.115-.055-.113-.056-.114-.057-.113-.057-.113-.058-.112-.059-.112-.059-.111-.06-.111-.06-.111-.062-.11-.061-.11-.063-.109-.063-.109-.064-.109-.064-.108-.065-.108-.066-.107-.066-.107-.067-.106-.068-.106-.068-.106-.069-.105-.069-.105-.07-.104-.071-.104-.071-.103-.072-.103-.072-.103-.074-.102-.073-.102-.074-.101-.075-.1-.076-.1-.076-.1-.153-.198-.156-.196-.159-.195-.16-.192-.163-.19-.165-.188-.168-.187-.169-.184-.172-.182-.174-.18-.176-.178-.178-.176-.18-.174-.182-.172-.184-.169-.187-.168-.188-.165-.19-.163-.193-.16-.194-.159-.196-.156-.198-.153-.1-.076-.1-.076-.1-.075-.102-.074-.101-.074-.102-.073-.103-.072-.103-.072-.103-.071-.104-.071-.104-.07-.105-.069-.105-.069-.106-.068-.106-.068-.106-.067-.107-.066-.107-.066-.108-.065-.108-.064-.109-.064-.109-.063-.109-.063-.11-.061-.11-.062-.111-.06-.111-.06-.111-.059-.112-.059-.113-.058-.112-.057-.113-.057-.114-.056-.113-.055-.115-.055-.114-.054-.115-.053-.115-.053-.116-.052-.116-.051-.117-.05-.116-.05-.118-.049-.117-.049-.118-.048-.118-.047-.119-.046-.119-.046-.119-.045-.12-.044-.12-.044-.12-.042-.121-.042-.121-.042-.121-.041-.122-.039-.122-.04-.122-.038-.122-.038-.123-.037-.124-.036-.123-.036-.124-.035-.124-.034-.125-.033-.125-.032-.125-.032-.125-.031-.126-.031-.126-.029-.126-.029-.127-.028-.126-.027-.128-.027-.127-.025-.128-.025-.128-.024-.128-.024-.128-.022-.129-.022-.129-.021-.129-.02-.13-.02-.13-.018-.13-.018-.13-.017-.131-.016-.131-.016-.131-.014-.131-.014-.132-.013-.131-.012-.132-.011-.133-.011-.132-.01-.133-.008-.133-.008-.133-.008-.133-.006-.134-.005-.134-.005-.134-.004-.134-.003-.134-.002-.135-.001-.135-.001H70.036z"
-></path>
-<circle cx="200" cy="108.176" r="70.019" fill="#fff"></circle>
-<path
-  fill="#eee"
-  d="M200.453 38.158l.451.005.452.007.45.01.45.013.449.015.448.019.448.021.447.024.446.027.445.03.445.032.444.036.443.038.442.04.442.044.44.046.44.049.439.052.439.054.437.057.437.06.435.062.435.065.434.068.433.071.432.073.432.075.43.079.43.081.428.083.428.087.427.088.425.092.425.094.424.096.423.099.421.102.421.104.42.107.419.109.418.112.417.114.415.117.415.119.414.122.412.124.412.127.41.129.41.132.408.134.407.137.406.138.405.142.404.144.403.146.401.148.4.151.4.153.398.156.396.158.396.16.394.163.394.165.392.167.39.17.39.172.388.174.387.177.386.179.385.181.383.183.382.186.38.188.38.19.378.192.377.194.375.197.374.199.373.201.371.203.37.206.369.208.367.209.366.212.364.214.363.217.362.218.36.22.359.223.357.225.356.226.354.229.353.231.352.233.35.235.348.237.347.239.346.241.344.243.342.246.341.247.34.249.337.251.337.253.334.255.333.257.332.259.33.261.328.263.327.265.325.266.323.269.322.27.32.273.319.274.317.276.315.278.313.28.312.282.31.283.308.286.307.287.305.289.303.291.302.292.299.295.298.296.296.298.295.299.292.302.291.303.289.305.287.307.286.308.283.31.282.312.28.313.278.315.276.317.274.319.273.32.27.322.269.323.266.325.265.327.263.328.261.33.259.332.257.333.255.334.253.337.251.337.249.34.248.341.245.342.243.344.241.346.239.347.237.348.235.35.233.352.231.353.229.354.226.356.225.357.223.359.22.36.218.362.217.363.214.364.212.366.209.367.208.369.206.37.203.371.201.373.199.374.197.375.194.377.192.378.191.38.187.38.186.382.183.383.181.385.179.386.177.387.174.388.172.39.17.39.167.392.165.394.163.394.16.396.158.396.156.398.153.4.151.4.148.402.147.402.143.404.142.405.139.406.136.407.134.408.132.41.129.41.127.412.124.412.122.414.119.415.117.415.114.417.112.418.109.419.107.42.104.421.102.421.099.423.096.424.094.425.092.426.088.426.087.428.083.428.081.43.079.43.075.432.073.432.071.433.068.434.065.435.062.435.06.437.057.437.054.439.052.439.049.44.046.44.044.442.04.442.038.443.036.444.032.445.03.445.027.446.024.447.021.448.019.448.015.449.013.45.01.45.007.452.005.451.001.453-.001.453-.005.451-.007.452-.01.45-.013.45-.015.449-.019.448-.021.448-.024.447-.027.446-.03.445-.032.445-.036.444-.038.443-.04.442-.044.442-.046.44-.049.44-.052.439-.054.439-.057.437-.06.437-.062.435-.065.435-.068.434-.071.433-.073.433-.075.431-.079.43-.081.43-.083.428-.087.428-.088.427-.092.425-.094.425-.096.424-.099.423-.102.422-.104.42-.107.42-.109.419-.112.418-.114.417-.117.415-.119.415-.122.414-.124.412-.127.412-.129.41-.132.41-.134.408-.136.407-.139.406-.142.405-.143.404-.147.403-.148.401-.151.4-.153.4-.156.398-.158.396-.16.396-.163.394-.165.394-.167.392-.17.39-.172.39-.174.388-.177.387-.179.386-.181.385-.183.383-.186.382-.187.38-.191.38-.192.378-.194.377-.197.375-.199.374-.201.373-.203.371-.206.37-.208.369-.209.367-.212.366-.214.364-.217.363-.218.362-.22.36-.223.359-.225.357-.226.356-.229.355-.231.352-.233.352-.235.35-.237.348-.239.347-.241.346-.243.344-.245.342-.248.341-.249.34-.251.337-.253.337-.255.334-.257.333-.259.332-.261.33-.263.328-.265.327-.266.325-.269.323-.27.322-.273.32-.274.319-.276.317-.278.315-.28.313-.282.312-.283.31-.286.308-.287.307-.289.305-.291.303-.292.302-.295.299-.296.298-.298.296-.299.295-.302.292-.303.291-.305.289-.307.287-.308.286-.31.283-.312.282-.313.28-.315.278-.317.276-.319.274-.32.273-.322.27-.323.269-.325.266-.327.265-.328.263-.33.261-.332.259-.333.257-.334.255-.337.253-.337.251-.34.249-.341.248-.342.245-.344.243-.346.241-.347.239-.348.237-.35.235-.352.233-.353.231-.354.229-.356.226-.357.225-.359.223-.36.22-.362.218-.363.217-.364.214-.366.212-.367.209-.369.208-.37.206-.371.203-.373.201-.374.199-.375.197-.377.194-.378.192-.38.191-.38.187-.382.186-.383.183-.385.181-.386.179-.387.177-.388.174-.39.172-.39.17-.392.167-.394.165-.394.163-.396.16-.396.158-.398.156-.4.153-.4.151-.401.148-.403.147-.404.143-.405.142-.406.139-.407.136-.408.134-.41.132-.41.129-.412.127-.412.124-.414.122-.415.119-.415.117-.417.114-.418.112-.419.109-.42.107-.421.104-.421.102-.423.099-.424.096-.425.094-.425.092-.427.088-.428.087-.428.083-.43.081-.43.079-.432.075-.432.073-.433.071-.434.068-.435.065-.435.062-.437.06-.437.057-.439.054-.439.052-.44.049-.44.046-.442.044-.442.04-.443.038-.444.036-.445.032-.445.03-.446.027-.447.024-.448.021-.448.019-.449.015-.45.013-.45.01-.452.007-.451.005-.453.001-.453-.001-.451-.005-.452-.007-.45-.01-.45-.013-.449-.015-.448-.019-.448-.021-.447-.024-.446-.027-.445-.03-.445-.032-.444-.036-.443-.038-.442-.04-.442-.044-.44-.046-.44-.049-.439-.052-.439-.054-.437-.057-.437-.06-.435-.062-.435-.065-.434-.068-.433-.071-.432-.073-.432-.075-.43-.079-.43-.081-.428-.083-.428-.087-.427-.088-.425-.092-.425-.094-.424-.096-.423-.099-.421-.102-.421-.104-.42-.107-.419-.109-.418-.112-.417-.114-.415-.117-.415-.119-.414-.122-.412-.124-.412-.127-.41-.129-.41-.132-.408-.134-.407-.136-.406-.139-.405-.142-.404-.143-.403-.147-.401-.148-.4-.151-.4-.153-.398-.156-.396-.158-.396-.16-.394-.163-.394-.165-.392-.167-.39-.17-.39-.172-.388-.174-.387-.177-.386-.179-.385-.181-.383-.183-.382-.186-.38-.187-.38-.191-.378-.192-.377-.194-.375-.197-.374-.199-.373-.201-.371-.203-.37-.206-.369-.208-.367-.209-.366-.212-.364-.214-.363-.217-.362-.218-.36-.22-.359-.223-.357-.225-.356-.226-.354-.229-.353-.231-.352-.233-.35-.235-.348-.237-.347-.239-.346-.241-.344-.243-.342-.245-.341-.248-.34-.249-.337-.251-.337-.253-.334-.255-.333-.257-.332-.259-.33-.261-.328-.263-.327-.265-.325-.266-.323-.269-.322-.27-.32-.273-.319-.274-.317-.276-.315-.278-.313-.28-.312-.282-.31-.283-.308-.286-.307-.287-.305-.289-.303-.291-.302-.292-.299-.295-.298-.296-.296-.298-.295-.299-.292-.302-.291-.303-.289-.305-.287-.307-.286-.308-.283-.31-.282-.312-.28-.313-.278-.315-.276-.317-.274-.319-.273-.32-.27-.322-.269-.323-.266-.325-.265-.327-.263-.328-.261-.33-.259-.332-.257-.333-.255-.334-.253-.337-.251-.337-.249-.34-.248-.341-.245-.342-.243-.344-.241-.346-.239-.347-.237-.348-.235-.35-.233-.352-.231-.352-.229-.355-.226-.356-.225-.357-.223-.359-.22-.36-.218-.362-.217-.363-.214-.364-.212-.366-.209-.367-.208-.369-.206-.37-.203-.371-.201-.373-.199-.374-.197-.375-.194-.377-.192-.378-.191-.38-.187-.38-.186-.382-.183-.383-.181-.385-.179-.386-.177-.387-.174-.388-.172-.39-.17-.39-.167-.392-.165-.394-.163-.394-.16-.396-.158-.396-.156-.398-.153-.4-.151-.4-.148-.401-.147-.403-.143-.404-.142-.405-.139-.406-.136-.407-.134-.408-.132-.41-.129-.41-.127-.412-.124-.412-.122-.414-.119-.415-.117-.415-.114-.417-.112-.418-.109-.419-.107-.42-.104-.42-.102-.422-.099-.423-.096-.424-.094-.425-.092-.425-.088-.427-.087-.428-.083-.428-.081-.43-.079-.43-.075-.431-.073-.433-.071-.433-.068-.434-.065-.435-.062-.435-.06-.437-.057-.437-.054-.439-.052-.439-.049-.44-.046-.44-.044-.442-.04-.442-.038-.443-.036-.444-.032-.445-.03-.445-.027-.446-.024-.447-.021-.448-.019-.448-.015-.449-.013-.45-.01-.45-.007-.452-.005-.451-.001-.453.001-.453.005-.451.007-.452.01-.45.013-.45.015-.449.019-.448.021-.448.024-.447.027-.446.03-.445.032-.445.036-.444.038-.443.04-.442.044-.442.046-.44.049-.44.052-.439.054-.439.057-.437.06-.437.062-.435.065-.435.068-.434.071-.433.073-.432.075-.432.079-.43.081-.43.083-.428.087-.428.088-.426.092-.426.094-.425.096-.424.099-.423.102-.421.104-.421.107-.42.109-.419.112-.418.114-.417.117-.415.119-.415.122-.414.124-.412.127-.412.129-.41.132-.41.134-.408.136-.407.139-.406.142-.405.143-.404.147-.402.148-.402.151-.4.153-.4.156-.398.158-.396.16-.396.163-.394.165-.394.167-.392.17-.39.172-.39.174-.388.177-.387.179-.386.181-.385.183-.383.186-.382.187-.38.191-.38.192-.378.194-.377.197-.375.199-.374.201-.373.203-.371.206-.37.208-.369.209-.367.212-.366.214-.364.217-.363.218-.362.22-.36.223-.359.225-.357.226-.356.229-.354.231-.353.233-.352.235-.35.237-.348.239-.347.241-.346.243-.344.245-.342.248-.341.249-.34.251-.337.253-.337.255-.334.257-.333.259-.332.261-.33.263-.328.265-.327.266-.325.269-.323.27-.322.273-.32.274-.319.276-.317.278-.315.28-.313.282-.312.283-.31.286-.308.287-.307.289-.305.291-.303.292-.302.295-.299.296-.298.298-.296.299-.295.302-.292.303-.291.305-.289.307-.287.308-.286.31-.283.312-.282.313-.28.315-.278.317-.276.319-.274.32-.273.322-.27.323-.269.325-.266.327-.265.328-.263.33-.261.332-.259.333-.257.334-.255.337-.253.337-.251.34-.249.341-.247.342-.246.344-.243.346-.241.347-.239.348-.237.35-.235.352-.233.353-.231.354-.229.356-.226.357-.225.359-.223.36-.22.362-.218.363-.217.364-.214.366-.212.367-.209.369-.208.37-.206.371-.203.373-.201.374-.199.375-.197.377-.194.378-.192.38-.19.38-.188.382-.186.383-.183.385-.181.386-.179.387-.177.388-.174.39-.172.39-.17.392-.167.394-.165.394-.163.396-.16.396-.158.398-.156.4-.153.4-.151.401-.148.403-.146.404-.144.405-.142.406-.138.407-.137.408-.134.41-.132.41-.129.412-.127.412-.124.414-.122.415-.119.415-.117.417-.114.418-.112.419-.109.42-.107.421-.104.421-.102.423-.099.424-.096.425-.094.425-.092.427-.088.428-.087.428-.083.43-.081.43-.079.432-.075.432-.073.433-.071.434-.068.435-.065.435-.062.437-.06.437-.057.439-.054.439-.052.44-.049.44-.046.442-.044.442-.04.443-.038.444-.036.445-.032.445-.03.446-.027.447-.024.448-.021.448-.019.449-.015.45-.013.45-.01.452-.007.451-.005.453-.001.453.001zm-.864 6.5l-.411.004-.41.007-.409.009-.409.012-.408.014-.407.017-.406.019-.406.022-.405.024-.405.027-.403.03-.403.032-.402.034-.402.037-.4.039-.4.042-.399.045-.399.047-.397.049-.397.052-.396.054-.395.057-.394.059-.394.061-.393.064-.391.066-.392.069-.39.071-.389.073-.389.076-.387.078-.387.081-.386.083-.385.085-.384.087-.384.09-.382.092-.381.095-.381.097-.38.099-.378.101-.378.104-.377.105-.376.109-.375.11-.374.113-.373.114-.372.117-.371.12-.37.121-.369.124-.368.126-.367.128-.366.13-.365.133-.364.134-.363.137-.362.139-.361.141-.359.144-.359.145-.358.147-.356.15-.355.152-.355.154-.353.155-.352.158-.351.16-.35.163-.348.164-.348.166-.346.168-.345.171-.344.172-.343.174-.342.177-.34.178-.34.181-.338.182-.337.185-.335.186-.334.188-.334.191-.332.192-.33.194-.329.196-.329.198-.326.2-.326.202-.324.204-.323.206-.321.207-.321.21-.319.211-.317.213-.316.216-.315.217-.314.218-.312.221-.311.222-.309.225-.308.226-.307.228-.305.23-.304.231-.302.233-.301.235-.299.237-.298.239-.297.24-.295.242-.294.244-.292.246-.29.247-.29.249-.287.251-.286.252-.285.254-.283.256-.281.257-.28.259-.279.261-.277.262-.275.264-.273.266-.272.267-.271.269-.269.271-.267.272-.266.273-.264.275-.262.277-.261.279-.259.28-.257.281-.256.283-.254.285-.252.286-.251.287-.249.29-.247.29-.246.292-.244.294-.242.295-.24.297-.239.298-.237.299-.235.301-.233.302-.232.304-.229.305-.228.307-.226.308-.225.309-.222.311-.221.312-.219.314-.216.314-.216.317-.213.317-.211.319-.21.321-.207.321-.206.323-.204.324-.202.326-.2.327-.198.328-.196.329-.194.331-.192.331-.191.334-.188.334-.186.335-.185.337-.182.338-.181.34-.178.34-.177.342-.174.343-.172.344-.171.345-.168.346-.166.348-.164.348-.163.35-.16.351-.158.352-.156.353-.153.355-.152.355-.15.357-.147.357-.145.359-.144.36-.141.36-.139.362-.137.363-.134.364-.133.365-.13.366-.128.367-.126.368-.124.369-.121.37-.12.371-.117.372-.114.373-.113.374-.11.375-.109.376-.105.377-.104.378-.101.378-.099.38-.097.381-.095.381-.092.382-.09.384-.087.384-.085.385-.083.386-.081.387-.078.387-.076.389-.073.389-.071.39-.069.391-.066.392-.064.393-.061.394-.059.394-.057.395-.054.396-.052.397-.049.397-.047.399-.045.399-.042.4-.039.4-.037.402-.034.402-.032.403-.03.403-.027.405-.024.405-.022.406-.019.406-.017.407-.015.408-.011.409-.009.409-.007.41-.004.411-.001.411.001.411.004.411.007.41.009.409.011.409.015.408.017.407.019.406.022.406.024.405.027.405.03.403.032.403.034.402.037.402.039.4.042.4.045.399.047.399.049.397.052.397.054.396.057.395.059.394.061.394.064.393.066.392.069.391.071.39.073.389.076.389.078.387.081.387.083.386.085.385.087.384.09.384.092.382.095.382.097.38.099.38.101.378.104.378.105.377.109.376.11.375.113.374.114.373.117.372.12.371.121.37.124.369.126.368.128.367.13.366.133.365.134.364.137.363.139.362.141.361.144.359.145.359.147.358.15.356.152.355.153.355.156.353.158.352.16.351.163.35.164.348.166.348.168.346.171.346.172.343.174.343.177.342.178.34.181.34.182.338.185.337.186.335.188.334.191.334.192.332.194.33.196.329.198.329.2.326.202.326.204.324.206.323.207.322.21.32.211.319.213.317.216.317.216.314.219.314.221.312.222.311.225.309.226.308.228.307.23.305.231.304.233.302.235.301.237.299.239.299.24.296.242.295.244.294.246.292.247.29.249.29.251.287.252.286.254.285.256.283.257.281.259.28.261.279.262.277.264.275.266.273.267.273.269.27.271.269.272.267.273.266.275.264.277.262.279.261.28.259.281.257.283.256.285.254.286.253.287.25.29.249.29.247.292.246.294.244.295.242.297.24.298.239.299.237.301.235.302.233.304.232.305.229.307.228.308.226.309.225.311.222.312.221.314.219.314.216.317.216.317.213.319.211.32.21.322.207.323.206.324.204.326.202.326.2.329.198.329.196.33.194.332.193.334.19.334.188.335.186.337.185.338.182.34.181.34.178.342.177.343.174.343.172.346.171.346.168.348.166.348.164.35.163.351.16.352.158.353.156.355.153.355.152.356.15.358.147.359.145.36.144.36.141.362.139.363.137.364.134.365.133.366.13.367.128.368.126.369.124.37.121.371.12.372.117.373.114.374.113.375.11.376.109.377.105.378.104.378.101.38.099.38.097.382.095.382.092.384.09.384.087.385.085.386.083.387.081.387.078.389.076.389.073.39.071.391.069.392.066.393.064.394.061.394.059.395.057.396.054.397.052.397.049.399.047.399.045.4.042.4.039.402.037.402.034.403.032.403.03.405.027.405.024.406.022.406.02.407.016.408.015.409.011.409.009.41.007.411.004.411.001.411-.001.411-.004.41-.007.409-.009.409-.011.408-.015.407-.016.406-.02.406-.022.405-.024.405-.027.403-.03.403-.032.402-.034.402-.037.4-.039.4-.042.399-.045.399-.047.397-.049.397-.052.396-.054.395-.057.394-.059.394-.061.393-.064.392-.066.391-.069.39-.071.389-.073.389-.076.387-.078.387-.081.386-.083.385-.085.384-.087.384-.09.382-.092.382-.095.38-.097.38-.099.378-.101.378-.104.377-.105.376-.109.375-.11.374-.113.373-.114.372-.117.371-.12.37-.121.369-.124.368-.126.367-.128.366-.13.365-.133.364-.134.363-.137.362-.139.36-.141.36-.144.359-.145.357-.147.357-.15.355-.152.355-.153.353-.156.352-.158.351-.16.35-.163.348-.164.348-.166.346-.168.345-.171.344-.172.343-.174.342-.177.34-.178.34-.181.338-.182.337-.184.335-.187.334-.188.334-.191.332-.192.33-.194.329-.196.328-.198.327-.2.326-.202.324-.204.323-.206.322-.207.32-.21.319-.211.317-.213.317-.216.314-.216.314-.219.312-.221.311-.222.309-.225.308-.226.307-.228.305-.229.304-.232.302-.233.301-.235.299-.237.299-.239.296-.24.295-.242.294-.244.292-.246.29-.247.29-.249.287-.25.286-.253.285-.254.283-.256.281-.257.28-.259.279-.261.277-.262.275-.264.273-.266.272-.267.271-.269.269-.271.267-.272.266-.273.264-.276.262-.276.261-.279.259-.28.257-.281.256-.283.254-.285.252-.286.251-.288.249-.289.247-.29.246-.292.244-.294.242-.295.24-.297.239-.297.237-.3.235-.301.233-.302.231-.304.23-.305.228-.307.226-.308.224-.309.223-.311.221-.312.218-.314.218-.315.215-.316.213-.317.211-.319.21-.32.207-.322.206-.323.204-.324.202-.326.2-.326.198-.329.196-.329.194-.33.192-.332.191-.334.188-.334.186-.335.185-.337.182-.338.181-.34.178-.34.176-.342.175-.343.172-.344.171-.345.168-.346.166-.348.164-.348.163-.35.16-.351.158-.352.155-.353.154-.355.152-.355.15-.357.147-.357.145-.359.144-.359.141-.361.139-.362.137-.363.134-.364.133-.365.13-.366.128-.367.126-.368.124-.369.121-.37.12-.371.117-.372.114-.373.113-.374.11-.375.109-.376.105-.377.104-.378.101-.378.099-.38.097-.38.095-.382.092-.382.09-.383.087-.385.085-.385.083-.386.081-.387.078-.387.076-.389.073-.389.071-.39.069-.391.066-.392.064-.393.061-.394.059-.394.057-.395.054-.396.052-.397.049-.397.047-.399.045-.399.042-.4.039-.4.037-.402.034-.402.032-.403.03-.403.027-.405.024-.405.022-.406.019-.406.017-.407.015-.408.011-.409.009-.409.007-.41.004-.411.001-.411-.001-.411-.004-.411-.007-.41-.009-.409-.011-.409-.015-.408-.017-.407-.019-.406-.022-.406-.024-.405-.027-.405-.03-.403-.032-.403-.034-.402-.037-.402-.039-.4-.042-.4-.045-.399-.047-.398-.049-.398-.052-.397-.054-.396-.057-.395-.059-.394-.061-.394-.064-.393-.066-.392-.069-.391-.071-.39-.073-.389-.076-.389-.078-.387-.081-.387-.083-.386-.085-.385-.087-.385-.09-.383-.092-.382-.095-.381-.097-.381-.099-.38-.101-.378-.104-.378-.105-.377-.109-.376-.11-.375-.113-.374-.114-.373-.117-.372-.12-.371-.121-.37-.124-.369-.126-.368-.128-.367-.13-.366-.133-.365-.134-.364-.137-.363-.139-.362-.141-.361-.144-.359-.145-.359-.147-.357-.15-.357-.152-.355-.154-.355-.155-.353-.158-.352-.16-.351-.163-.35-.164-.348-.166-.348-.168-.346-.171-.345-.172-.344-.175-.343-.176-.342-.178-.34-.181-.34-.182-.338-.185-.337-.186-.335-.188-.334-.191-.334-.192-.331-.194-.331-.196-.329-.198-.328-.2-.327-.202-.326-.204-.324-.206-.323-.207-.322-.21-.32-.211-.319-.213-.317-.215-.316-.218-.315-.218-.314-.221-.312-.223-.311-.224-.309-.226-.308-.228-.307-.23-.305-.231-.304-.233-.302-.235-.301-.237-.3-.239-.297-.24-.297-.242-.295-.244-.294-.246-.292-.247-.29-.249-.289-.251-.288-.252-.286-.254-.285-.256-.283-.257-.281-.259-.28-.261-.279-.262-.276-.264-.276-.266-.273-.267-.272-.269-.271-.271-.269-.272-.267-.273-.266-.275-.264-.277-.262-.279-.261-.28-.259-.281-.257-.283-.256-.285-.254-.286-.252-.287-.251-.29-.249-.29-.247-.292-.246-.294-.244-.295-.242-.296-.24-.299-.239-.299-.237-.301-.235-.302-.233-.304-.231-.305-.23-.307-.228-.308-.226-.309-.225-.311-.222-.312-.221-.314-.218-.314-.217-.317-.216-.317-.213-.319-.211-.32-.21-.322-.207-.323-.206-.324-.204-.326-.202-.326-.2-.329-.198-.329-.196-.33-.194-.332-.192-.334-.191-.334-.188-.336-.187-.336-.184-.338-.182-.34-.181-.34-.178-.342-.177-.343-.174-.344-.172-.345-.171-.346-.168-.348-.166-.348-.164-.35-.163-.351-.16-.352-.158-.353-.155-.355-.154-.355-.152-.357-.15-.357-.147-.359-.145-.359-.144-.361-.141-.362-.139-.363-.137-.364-.134-.365-.133-.366-.13-.367-.128-.368-.126-.369-.124-.37-.121-.371-.12-.372-.117-.373-.114-.374-.113-.375-.11-.376-.109-.377-.105-.378-.104-.378-.101-.38-.099-.381-.097-.381-.095-.382-.092-.384-.09-.384-.087-.385-.085-.386-.083-.387-.081-.387-.078-.389-.076-.389-.073-.39-.071-.392-.069-.391-.066-.393-.064-.394-.061-.394-.059-.395-.057-.396-.054-.397-.052-.398-.049-.398-.047-.399-.045-.4-.042-.4-.039-.402-.037-.402-.034-.403-.032-.403-.03-.405-.027-.405-.024-.406-.022-.406-.019-.407-.017-.408-.014-.409-.012-.409-.009-.41-.007-.411-.004-.411-.001-.411.001z"
-></path>
-<path
-  fill="#565656"
-  d="M289.817 202.686a7.59 7.59 0 00-7.585-7.586H117.768a7.59 7.59 0 00-7.585 7.586v3.834a7.59 7.59 0 007.585 7.586h164.464a7.59 7.59 0 007.585-7.586v-3.834z"
-></path>
-<path
-  fill="#888"
-  d="M245.908 331.685a5.593 5.593 0 00-5.59-5.59h-80.636a5.593 5.593 0 00-5.59 5.59v2.813a5.593 5.593 0 005.59 5.59h80.636a5.593 5.593 0 005.59-5.59v-2.813zM306.311 281.891a5.686 5.686 0 00-5.684-5.683H99.373a5.686 5.686 0 00-5.684 5.683v2.674a5.687 5.687 0 005.684 5.684h201.254a5.687 5.687 0 005.684-5.684v-2.674zM326.619 259.129a5.686 5.686 0 00-5.683-5.683H72.341a5.686 5.686 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.675zM326.619 236.368a5.687 5.687 0 00-5.683-5.684H72.341a5.687 5.687 0 00-5.683 5.684v2.674a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.674z"
-></path>
-<path
-  fill="#888"
-  fillRule="nonzero"
-  d="M176.939 83.289h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719h-46.122a3.72 3.72 0 01-3.719-3.719V87.009a3.721 3.721 0 013.719-3.72zm43.89 47.609V89.24h-41.658v41.658h41.658zM200 98.167c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976h-29.756a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-></path>
-</svg>;
-icons.infoLeftOverlay = <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 400 400"
->
-<path
-  fill="#fff"
-  stroke="#eee"
-  strokeWidth="5"
-  d="M390.993 131.261c0-7.864-5.144-14.248-11.479-14.248H20.486c-6.335 0-11.479 6.384-11.479 14.248v177.518c0 7.865 5.144 14.249 11.479 14.249h359.028c6.335 0 11.479-6.384 11.479-14.249V131.261z"
-></path>
-<path
-  fill="#fff"
-  d="M151.446 72.879c0-7.396-6.005-13.401-13.401-13.401H42.752c-7.396 0-13.4 6.005-13.4 13.401v88.334c0 7.395 6.004 13.4 13.4 13.4h95.293c7.396 0 13.401-6.005 13.401-13.4V72.879z"
-></path>
-<path
-  fill="#eee"
-  d="M138.219 59.48l.172.003.172.005.172.008.171.01.17.012.17.014.169.016.168.018.168.02.167.023.167.024.166.026.165.029.165.03.164.033.163.034.162.037.162.038.161.041.16.042.159.044.159.046.157.048.157.05.156.052.156.054.154.055.153.057.153.06.152.06.15.063.15.065.149.066.148.068.147.07.146.071.145.073.144.075.143.077.142.078.14.08.14.081.139.084.137.084.137.087.135.088.135.089.133.091.132.093.13.094.13.096.128.097.128.099.126.1.124.102.124.103.122.105.121.106.12.107.118.109.117.11.116.112.114.113.113.115.112.115.11.117.109.119.107.119.106.121.105.122.103.124.102.125.1.126.099.127.097.128.096.13.094.131.093.132.091.133.089.134.088.135.087.137.085.137.083.139.081.14.08.141.078.142.077.142.075.144.073.145.071.146.07.147.068.148.067.149.064.15.063.151.061.151.059.153.057.153.056.155.053.155.052.156.05.157.048.157.046.159.044.159.043.16.04.161.038.162.037.162.034.164.033.163.03.165.029.165.026.166.025.167.022.167.02.168.019.168.016.17.014.169.012.171.009.171.008.171.005.172.004.173.001.173v88.334l-.001.173-.004.172-.005.172-.008.172-.009.171-.012.17-.014.17-.016.169-.019.168-.02.168-.022.168-.025.166-.026.166-.029.165-.03.165-.033.164-.034.163-.037.162-.038.162-.04.161-.043.16-.044.159-.046.159-.048.158-.05.156-.052.156-.053.156-.056.154-.057.153-.059.153-.061.152-.063.15-.064.15-.067.149-.068.148-.07.147-.071.146-.073.145-.075.144-.077.143-.078.142-.08.14-.081.14-.083.139-.085.137-.087.137-.088.135-.089.135-.091.133-.093.132-.094.13-.096.13-.097.128-.099.128-.1.126-.102.124-.103.124-.105.122-.106.121-.107.12-.109.118-.11.117-.112.116-.113.114-.114.113-.116.112-.117.11-.118.109-.12.107-.121.106-.122.105-.124.103-.124.102-.126.1-.128.099-.128.097-.13.096-.13.094-.132.093-.133.091-.135.09-.135.087-.137.087-.137.085-.139.083-.14.081-.14.08-.142.078-.143.077-.144.075-.145.073-.146.072-.147.069-.148.068-.149.067-.15.064-.15.063-.152.061-.153.059-.153.057-.154.056-.156.053-.156.052-.157.05-.157.048-.159.046-.159.044-.16.043-.161.04-.162.038-.162.037-.163.034-.164.033-.165.03-.165.029-.166.026-.167.025-.167.022-.168.02-.168.019-.169.016-.17.014-.17.012-.171.009-.172.008-.172.005-.172.004-.174.001H42.752l-.173-.001-.172-.004-.172-.005-.172-.008-.171-.009-.17-.012-.17-.014-.169-.016-.169-.019-.167-.02-.168-.022-.166-.025-.166-.026-.165-.029-.165-.03-.164-.033-.163-.034-.162-.037-.162-.038-.161-.04-.16-.043-.159-.044-.159-.046-.158-.048-.157-.05-.156-.052-.155-.053-.154-.056-.154-.057-.152-.059-.152-.061-.151-.063-.149-.064-.149-.067-.148-.068-.147-.069-.146-.072-.145-.073-.144-.075-.143-.077-.142-.078-.141-.08-.139-.081-.139-.083-.138-.085-.136-.087-.136-.087-.134-.09-.133-.091-.132-.093-.131-.094-.129-.096-.129-.097-.127-.099-.126-.1-.124-.102-.124-.103-.122-.105-.121-.106-.12-.107-.118-.109-.117-.11-.116-.112-.114-.113-.113-.114-.112-.116-.11-.117-.109-.118-.108-.12-.106-.121-.104-.122-.103-.124-.102-.124-.1-.126-.099-.128-.097-.128-.096-.13-.094-.13-.093-.132-.091-.133-.09-.135-.088-.135-.086-.137-.085-.137-.083-.139-.082-.14-.079-.14-.079-.142-.076-.143-.075-.144-.073-.145-.072-.146-.069-.147-.068-.148-.067-.149-.064-.15-.063-.15-.061-.152-.059-.153-.057-.153-.056-.154-.053-.156-.052-.156-.05-.156-.048-.158-.046-.159-.045-.159-.042-.16-.04-.161-.039-.162-.036-.162-.035-.163-.032-.164-.031-.165-.028-.165-.026-.166-.025-.166-.022-.168-.021-.168-.018-.168-.016-.169-.014-.17-.012-.17-.01-.171-.007-.172-.006-.172-.003-.172-.001-.173V72.879l.001-.173.003-.173.006-.172.007-.171.01-.171.012-.171.014-.169.016-.17.018-.168.021-.168.022-.167.025-.167.026-.166.028-.165.031-.165.032-.163.035-.164.036-.162.039-.162.04-.161.042-.16.045-.159.046-.159.048-.157.05-.157.052-.156.053-.155.056-.155.057-.153.059-.153.061-.151.063-.151.064-.15.067-.149.068-.148.069-.147.072-.146.073-.145.075-.144.076-.142.079-.142.079-.141.082-.14.083-.139.085-.137.086-.137.088-.135.09-.134.091-.133.093-.132.094-.131.096-.13.097-.128.099-.127.1-.126.102-.125.103-.124.104-.122.106-.121.108-.119.109-.119.11-.117.112-.115.113-.115.114-.113.116-.112.117-.11.118-.109.12-.107.121-.106.122-.105.124-.103.124-.102.126-.1.127-.099.129-.097.129-.096.131-.094.132-.093.133-.091.134-.089.136-.088.136-.087.138-.084.139-.084.139-.081.141-.08.142-.078.143-.077.144-.075.145-.073.146-.071.147-.07.148-.068.149-.066.149-.065.151-.063.152-.06.152-.06.154-.057.154-.055.155-.054.156-.052.157-.05.158-.048.159-.046.159-.044.16-.042.161-.041.162-.038.162-.037.163-.034.164-.033.165-.03.165-.029.166-.026.166-.024.168-.023.167-.02.169-.018.169-.016.17-.014.17-.012.171-.01.172-.008.172-.005.172-.003.173-.002h95.293l.174.002zm-95.452 4.498l-.131.001-.115.002-.115.004-.115.005-.114.007-.113.008-.113.009-.113.011-.112.012-.111.013-.111.015-.111.016-.11.018-.109.019-.109.02-.109.021-.108.023-.107.024-.107.026-.107.027-.106.027-.105.03-.105.03-.104.032-.104.033-.103.034-.103.036-.102.036-.101.038-.101.039-.101.041-.1.041-.099.043-.098.044-.098.045-.098.046-.096.048-.097.048-.095.05-.095.051-.094.052-.094.053-.092.054-.093.055-.091.056-.091.058-.09.058-.089.06-.088.06-.088.062-.087.063-.087.063-.085.065-.085.066-.084.067-.083.068-.082.068-.082.07-.08.071-.08.071-.079.073-.078.073-.077.075-.076.075-.076.076-.074.078-.074.078-.072.078-.072.08-.07.081-.07.081-.069.083-.068.083-.066.084-.066.084-.065.086-.064.086-.062.087-.062.088-.061.088-.059.09-.059.09-.057.09-.056.092-.056.092-.054.093-.053.093-.052.094-.051.095-.049.096-.049.096-.047.097-.046.097-.045.098-.044.099-.043.099-.042.1-.04.1-.039.101-.038.102-.037.102-.035.102-.034.104-.033.103-.032.105-.031.104-.029.106-.028.106-.027.106-.025.107-.024.108-.023.108-.022.108-.02.109-.019.11-.017.11-.016.11-.015.111-.014.112-.012.112-.01.112-.01.113-.008.114-.006.114-.005.114-.004.115-.002.116-.001.13v88.305l.001.131.002.115.004.115.005.115.006.114.008.113.01.113.01.113.012.112.014.111.015.111.016.111.017.11.019.109.02.109.022.109.023.108.024.107.025.107.027.106.028.106.029.106.031.105.032.104.033.104.034.103.035.103.037.102.038.101.039.101.04.101.042.099.043.1.044.098.045.098.046.098.047.096.049.096.049.096.051.095.052.094.053.093.054.093.056.092.056.092.057.09.059.091.059.089.061.088.062.088.062.087.064.086.065.086.066.085.066.084.068.083.069.082.07.081.07.081.072.08.072.079.074.078.074.077.076.076.076.075.077.075.078.073.079.073.08.072.08.07.082.07.082.069.083.067.084.067.085.066.085.065.087.064.087.062.087.062.089.06.089.06.09.059.091.057.091.056.093.056.092.054.094.053.094.052.095.05.095.05.096.049.097.047.098.046.098.045.098.044.1.043.099.041.101.041.101.039.101.038.102.036.103.036.103.034.104.033.104.032.105.031.105.029.106.028.107.026.107.026.107.024.108.023.109.021.109.021.109.018.11.018.111.016.111.015.111.013.112.013.113.01.113.01.113.008.114.006.115.005.115.004.115.002.131.001h95.264l.131-.001.115-.002.115-.004.115-.005.113-.006.114-.008.113-.01.113-.01.112-.013.111-.013.111-.015.111-.016.109-.018.11-.018.109-.021.109-.021.108-.023.107-.024.107-.026.107-.026.105-.028.106-.029.104-.031.105-.032.104-.033.103-.034.103-.036.102-.036.101-.038.101-.039.1-.041.1-.041.099-.043.099-.044.098-.045.097-.046.097-.047.096-.049.096-.05.095-.05.094-.052.093-.053.093-.054.092-.056.092-.056.09-.057.09-.059.09-.06.088-.06.088-.062.087-.062.086-.064.086-.065.085-.066.083-.066.084-.068.082-.069.081-.07.081-.07.08-.072.079-.073.078-.073.077-.075.076-.075.075-.076.075-.077.073-.078.073-.079.071-.08.071-.081.07-.081.069-.082.067-.084.067-.083.066-.085.065-.086.063-.086.063-.087.062-.088.06-.088.06-.09.059-.09.057-.09.056-.092.056-.092.054-.092.053-.094.052-.094.05-.095.05-.096.049-.096.047-.096.046-.098.045-.098.044-.098.043-.1.041-.099.041-.101.039-.101.038-.101.036-.102.036-.103.034-.103.033-.104.032-.104.03-.105.03-.106.028-.106.026-.106.026-.107.024-.107.023-.108.021-.109.02-.109.019-.109.018-.11.016-.111.015-.111.013-.111.012-.112.011-.113.01-.113.007-.113.007-.114.005-.115.004-.115.002-.115.001-.131V72.893l-.001-.13-.002-.116-.004-.115-.005-.114-.007-.114-.007-.114-.01-.113-.011-.112-.012-.112-.013-.112-.015-.111-.016-.11-.018-.11-.019-.11-.02-.109-.021-.108-.023-.108-.024-.108-.026-.107-.026-.106-.028-.106-.03-.106-.03-.104-.032-.105-.033-.103-.034-.104-.036-.102-.036-.102-.038-.102-.039-.101-.041-.1-.041-.1-.043-.099-.044-.099-.045-.098-.046-.097-.047-.097-.049-.096-.05-.096-.05-.095-.052-.094-.053-.093-.055-.093-.055-.092-.056-.092-.058-.09-.058-.09-.06-.09-.06-.088-.062-.088-.063-.087-.063-.086-.065-.086-.066-.084-.067-.084-.067-.083-.069-.083-.07-.081-.071-.081-.071-.08-.073-.078-.073-.078-.075-.077-.075-.077-.076-.075-.077-.075-.078-.073-.079-.073-.08-.071-.081-.071-.081-.07-.082-.068-.084-.068-.083-.067-.085-.066-.086-.065-.086-.063-.087-.063-.088-.062-.088-.06-.09-.06-.09-.058-.09-.058-.092-.056-.092-.055-.093-.054-.093-.053-.094-.052-.095-.051-.096-.05-.096-.048-.096-.048-.098-.046-.098-.045-.098-.044-.1-.043-.1-.041-.1-.041-.101-.039-.101-.038-.102-.036-.103-.036-.103-.034-.104-.033-.104-.032-.105-.03-.106-.03-.106-.027-.106-.027-.107-.026-.107-.024-.108-.023-.109-.021-.109-.02-.109-.019-.11-.018-.111-.016-.111-.015-.111-.013-.112-.012-.113-.011-.113-.009-.113-.008-.115-.007-.114-.005-.115-.004-.115-.002-.131-.001H42.767z"
-></path>
-<path
-  fill="#888"
-  fillRule="nonzero"
-  d="M67.338 90.15h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719H67.338a3.72 3.72 0 01-3.719-3.719V93.87a3.721 3.721 0 013.719-3.72zm43.89 47.609V96.101H69.57v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976H75.521a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-></path>
-<path
-  fill="#565656"
-  d="M299.746 200.62a7.59 7.59 0 00-7.585-7.586H35.651a7.589 7.589 0 00-7.585 7.586v3.834a7.588 7.588 0 007.585 7.585h256.51a7.588 7.588 0 007.585-7.585v-3.834z"
-></path>
-<path
-  fill="#888"
-  d="M372.406 264.094a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h333.112a5.687 5.687 0 005.683-5.684v-2.674zM266.41 291.342a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h227.116a5.686 5.686 0 005.683-5.683v-2.675zM370.188 236.219a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h330.894a5.687 5.687 0 005.683-5.684v-2.674z"
-></path>
-</svg>;
-icons.unlocked = <svg
-xmlns="http://www.w3.org/2000/svg"
-width="24"
-height="24"
-viewBox="0 0 24 24"
->
-<path d="M5 12h14c.276 0 .525.111.707.293S20 12.724 20 13v7c0 .276-.111.525-.293.707S19.276 21 19 21H5c-.276 0-.525-.111-.707-.293S4 20.276 4 20v-7c0-.276.111-.525.293-.707S4.724 12 5 12zm3-2V7a3.988 3.988 0 011.169-2.831 3.983 3.983 0 012.821-1.174 3.985 3.985 0 012.652 1 4.052 4.052 0 011.28 2.209 1 1 0 101.958-.408 6.051 6.051 0 00-1.912-3.299A5.963 5.963 0 0011.995.995c-1.657.002-3.157.676-4.241 1.762S5.998 5.344 6 7v3H5a2.997 2.997 0 00-3 3v7a2.997 2.997 0 003 3h14a2.997 2.997 0 003-3v-7a2.997 2.997 0 00-3-3z"></path>
-</svg>;
-icons.locked = <svg
-xmlns="http://www.w3.org/2000/svg"
-width="24"
-height="24"
-viewBox="0 0 24 24"
->
-<path d="M5 12h14c.276 0 .525.111.707.293S20 12.724 20 13v7c0 .276-.111.525-.293.707S19.276 21 19 21H5c-.276 0-.525-.111-.707-.293S4 20.276 4 20v-7c0-.276.111-.525.293-.707S4.724 12 5 12zm13-2V7c0-1.657-.673-3.158-1.757-4.243S13.657 1 12 1s-3.158.673-4.243 1.757S6 5.343 6 7v3H5a2.997 2.997 0 00-3 3v7a2.997 2.997 0 003 3h14a2.997 2.997 0 003-3v-7a2.997 2.997 0 00-3-3zM8 10V7c0-1.105.447-2.103 1.172-2.828S10.895 3 12 3s2.103.447 2.828 1.172S16 5.895 16 7v3z"></path>
-</svg>;
-icons.px = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-  <path d="M2.896 6.603h1.419v.92h.027c.21-.394.504-.677.88-.848a2.926 2.926 0 011.223-.256c.534 0 1.001.094 1.399.283.399.188.73.447.993.775.263.329.46.712.591 1.15.132.438.197.907.197 1.407 0 .455-.059.898-.177 1.327a3.455 3.455 0 01-.539 1.137 2.699 2.699 0 01-.913.789c-.368.197-.802.295-1.302.295-.219 0-.438-.019-.657-.059a2.855 2.855 0 01-.631-.19 2.51 2.51 0 01-.558-.336 1.823 1.823 0 01-.427-.479h-.027v3.391H2.896V6.603zm5.231 3.404c0-.306-.039-.604-.118-.894a2.418 2.418 0 00-.355-.768 1.829 1.829 0 00-.592-.539 1.603 1.603 0 00-.814-.204c-.631 0-1.107.219-1.427.657-.319.438-.479 1.021-.479 1.748 0 .342.041.66.125.953.083.294.208.546.374.756.167.21.366.377.598.499.232.123.502.184.809.184.341 0 .63-.07.867-.21.237-.14.432-.322.585-.545a2.21 2.21 0 00.328-.763 3.86 3.86 0 00.099-.874zM12.714 9.823l-2.353-3.22h1.814l1.42 2.09 1.485-2.09h1.735l-2.313 3.141 2.602 3.654h-1.801l-1.721-2.51-1.67 2.51h-1.761l2.563-3.575z"></path>
-</g>
-</svg>;
-icons.em = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-  <path d="M6.248 9.402a2.401 2.401 0 00-.152-.683 1.704 1.704 0 00-.867-.967 1.56 1.56 0 00-.69-.151c-.263 0-.502.046-.716.138a1.633 1.633 0 00-.552.381 1.886 1.886 0 00-.368.572 2.002 2.002 0 00-.152.71h3.497zm-3.497.986c0 .263.038.517.112.762.075.245.186.46.335.644.149.184.338.331.565.44.228.11.5.165.815.165.438 0 .791-.095 1.058-.283.268-.188.467-.471.598-.848h1.42a2.824 2.824 0 01-1.104 1.716 2.99 2.99 0 01-.914.446c-.337.101-.69.152-1.058.152-.534 0-1.007-.088-1.419-.263a2.95 2.95 0 01-1.045-.736 3.12 3.12 0 01-.644-1.131 4.577 4.577 0 01-.217-1.445c0-.482.077-.94.23-1.374.153-.434.372-.815.657-1.143a3.13 3.13 0 011.032-.782c.403-.193.859-.29 1.367-.29a3.04 3.04 0 011.439.336c.425.223.778.519 1.058.887.281.368.484.791.611 1.268.127.478.16.971.099 1.479H2.751zM8.889 6.602h1.42v.947h.039c.114-.167.237-.32.368-.46a2.022 2.022 0 01.999-.585c.211-.057.451-.086.723-.086.412 0 .795.092 1.15.276.355.184.607.469.756.855a3.01 3.01 0 01.881-.828c.333-.202.749-.303 1.248-.303.719 0 1.277.176 1.676.526.399.351.598.938.598 1.761v4.693h-1.498v-3.97c0-.271-.009-.519-.027-.742a1.512 1.512 0 00-.151-.579.856.856 0 00-.374-.374c-.167-.088-.395-.132-.684-.132-.508 0-.876.158-1.104.473-.228.316-.342.763-.342 1.341v3.983h-1.498V9.034c0-.473-.085-.83-.256-1.071-.171-.241-.484-.362-.94-.362-.193 0-.379.04-.559.119a1.385 1.385 0 00-.473.341 1.703 1.703 0 00-.328.552 2.084 2.084 0 00-.125.75v4.035H8.889V6.602z"></path>
-</g>
-</svg>;
-icons.percent = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path
-fillRule="nonzero"
-d="M5.689 7.831c0 .246.017.476.053.69.035.215.092.401.17.559.079.158.182.283.309.375a.775.775 0 00.467.138.803.803 0 00.473-.138.978.978 0 00.315-.375 2.11 2.11 0 00.178-.559c.039-.214.059-.444.059-.69 0-.219-.015-.433-.046-.644a1.995 1.995 0 00-.164-.565 1.076 1.076 0 00-.316-.401.794.794 0 00-.499-.151.797.797 0 00-.5.151 1.02 1.02 0 00-.308.401 1.992 1.992 0 00-.152.565 5.253 5.253 0 00-.039.644zm1.012 2.616c-.394 0-.732-.07-1.012-.21a1.899 1.899 0 01-.684-.566 2.316 2.316 0 01-.381-.828 4.148 4.148 0 01-.118-1.012c0-.35.042-.685.125-1.005.083-.32.215-.598.394-.835.18-.236.408-.425.684-.565.276-.14.606-.21.992-.21s.716.07.992.21c.276.14.504.329.684.565.179.237.311.515.394.835.083.32.125.655.125 1.005 0 .36-.039.697-.118 1.012a2.3 2.3 0 01-.382.828 1.887 1.887 0 01-.683.566c-.28.14-.618.21-1.012.21zm5.586 1.722c0 .245.017.475.053.69.035.214.092.401.17.558.079.158.182.283.309.375a.775.775 0 00.467.138.803.803 0 00.473-.138.978.978 0 00.315-.375c.079-.157.138-.344.178-.558.039-.215.059-.445.059-.69 0-.219-.015-.434-.046-.644a1.992 1.992 0 00-.164-.566 1.065 1.065 0 00-.316-.4.795.795 0 00-.499-.152.798.798 0 00-.5.152 1.01 1.01 0 00-.308.4 1.99 1.99 0 00-.152.566c-.026.21-.039.425-.039.644zm1.012 2.615c-.394 0-.732-.07-1.012-.21a1.885 1.885 0 01-.683-.565 2.317 2.317 0 01-.382-.828 4.16 4.16 0 01-.118-1.012c0-.351.042-.686.125-1.006.083-.32.215-.598.394-.834.18-.237.408-.425.684-.566.276-.14.606-.21.992-.21s.716.07.992.21c.276.141.504.329.684.566.179.236.311.514.394.834.083.32.125.655.125 1.006 0 .359-.039.696-.118 1.012a2.332 2.332 0 01-.381.828 1.897 1.897 0 01-.684.565c-.28.14-.618.21-1.012.21zm-1.341-9.7h.999l-5.086 9.832H6.846l5.112-9.832z"
-></path>
-</svg>;
-icons.rem = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M.731 7.079H1.94v1.13h.023c.038-.158.111-.313.22-.463.11-.151.241-.289.396-.413.154-.124.326-.224.514-.3.189-.075.381-.113.577-.113.151 0 .254.004.311.012l.175.022v1.244a5.951 5.951 0 00-.277-.04 2.393 2.393 0 00-.277-.017 1.424 1.424 0 00-1.119.514c-.143.17-.256.379-.339.628a2.712 2.712 0 00-.125.859v2.781H.731V7.079zM8.519 9.486a2.057 2.057 0 00-.13-.587 1.534 1.534 0 00-.294-.492 1.433 1.433 0 00-.452-.339 1.327 1.327 0 00-.593-.13c-.226 0-.432.039-.616.118-.185.08-.343.189-.475.328-.132.14-.238.304-.317.492a1.727 1.727 0 00-.13.61h3.007zm-3.007.848c0 .226.032.445.096.656.064.211.161.395.289.554.128.158.29.284.486.378.196.095.429.142.701.142.376 0 .68-.081.91-.243.229-.162.401-.405.514-.73h1.221a2.422 2.422 0 01-.95 1.476 2.597 2.597 0 01-.785.384c-.291.087-.594.13-.91.13-.46 0-.867-.075-1.221-.226a2.535 2.535 0 01-.899-.633 2.69 2.69 0 01-.554-.972 3.964 3.964 0 01-.186-1.244c0-.414.066-.808.198-1.181.131-.373.32-.701.565-.983.245-.283.54-.507.887-.673A2.692 2.692 0 017.05 6.92c.459 0 .872.096 1.237.289.366.192.669.446.91.763.242.316.417.68.526 1.09.109.411.138.835.085 1.272H5.512zM10.791 7.079h1.221v.813h.034c.098-.143.203-.275.317-.395a1.722 1.722 0 01.859-.503c.18-.049.388-.074.621-.074.355 0 .684.079.989.238.306.158.522.403.65.734.219-.301.471-.538.758-.712.286-.173.644-.26 1.074-.26.618 0 1.098.151 1.441.452.343.302.514.807.514 1.515v4.036h-1.288V9.509c0-.234-.008-.447-.023-.639a1.292 1.292 0 00-.13-.497.737.737 0 00-.322-.322c-.143-.076-.339-.113-.588-.113-.437 0-.754.135-.95.407-.195.271-.293.655-.293 1.153v3.425h-1.289V9.17c0-.407-.074-.714-.221-.921-.146-.208-.416-.311-.808-.311a1.192 1.192 0 00-.887.395 1.48 1.48 0 00-.283.475 1.815 1.815 0 00-.107.644v3.471h-1.289V7.079z"></path>
-</g>
-</svg>;
-icons.vh = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M3.271 6.603H4.9l1.722 5.218h.026l1.656-5.218h1.551l-2.431 6.795H5.742L3.271 6.603zM10.762 4.014h1.499v3.483h.026c.184-.307.458-.563.821-.769a2.425 2.425 0 011.216-.309c.745 0 1.332.193 1.761.578.43.386.644.964.644 1.735v4.666h-1.498V9.127c-.017-.535-.131-.923-.342-1.164-.21-.241-.538-.361-.985-.361-.254 0-.482.046-.684.138a1.484 1.484 0 00-.512.381c-.141.162-.25.353-.329.572a2.035 2.035 0 00-.118.696v4.009h-1.499V4.014z"></path>
-</g>
-</svg>;
-icons.vw = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M1.621 6.603h1.63l1.722 5.218h.026l1.656-5.218h1.551l-2.432 6.795H4.092L1.621 6.603zM8.495 6.603h1.59l1.328 5.073h.026l1.275-5.073h1.512l1.222 5.073h.026l1.38-5.073h1.525l-2.129 6.795h-1.538L13.45 8.351h-.026l-1.249 5.047h-1.577L8.495 6.603z"></path>
-</g>
-</svg>;
-icons.none = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M17.401 4.69L15.31 2.599 2.599 15.31l2.091 2.091L17.401 4.69z"></path>
-<path d="M4.69 2.599L2.599 4.69 15.31 17.401l2.091-2.091L4.69 2.599z"></path>
-</svg>;
-icons.solid = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M18.988 11.478V8.522H1.012v2.956h17.976z"></path>
-</svg>;
-icons.dashed = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M12.512 11.478V8.522H7.488v2.956h5.024zM14.004 8.522v2.956h4.984V8.522h-4.984zM1.012 8.522v2.956H6.05V8.522H1.012z"></path>
-</svg>;
-icons.dotted = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<circle cx="2.503" cy="10" r="1.487"></circle>
-<circle cx="17.486" cy="10" r="1.487"></circle>
-<circle cx="12.447" cy="10" r="1.487"></circle>
-<circle cx="7.455" cy="10" r="1.487"></circle>
-</svg>;
-icons.double = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M1.02 6.561v2.957h17.968V6.561H1.02zM1.012 10.586v2.956H18.98v-2.956H1.012z"></path>
-</svg>;
-icons.lowercase = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M9.1 13.181c0 .184.024.315.072.394.048.079.142.118.283.118h.157a.953.953 0 00.211-.026v1.038a3.222 3.222 0 01-.204.06 3.035 3.035 0 01-.743.111c-.306 0-.561-.061-.762-.184-.202-.122-.333-.337-.394-.644-.298.289-.664.5-1.098.631a4.303 4.303 0 01-1.255.197c-.307 0-.6-.041-.881-.125a2.302 2.302 0 01-.742-.368 1.764 1.764 0 01-.513-.617 1.904 1.904 0 01-.19-.874c0-.421.076-.763.23-1.026.153-.262.354-.468.604-.617.25-.149.53-.257.841-.322.311-.066.625-.116.94-.152.272-.052.53-.089.776-.111a3.6 3.6 0 00.65-.112c.189-.053.337-.134.447-.243.11-.11.164-.274.164-.493a.765.765 0 00-.138-.473.931.931 0 00-.341-.283 1.5 1.5 0 00-.454-.131 3.61 3.61 0 00-.473-.033c-.421 0-.767.088-1.038.263-.272.175-.425.447-.46.815H3.29c.027-.438.132-.802.316-1.091.184-.289.418-.522.703-.697.285-.175.607-.298.966-.368s.727-.105 1.104-.105c.333 0 .662.035.986.105.324.07.615.184.874.342.258.158.466.361.624.611.158.25.237.554.237.914v3.496zm-1.499-1.893a1.822 1.822 0 01-.841.27 10.7 10.7 0 00-.999.138 3.18 3.18 0 00-.46.111c-.149.048-.28.114-.394.197a.874.874 0 00-.27.329c-.065.136-.098.3-.098.493 0 .166.048.307.144.42.097.114.213.204.349.27.136.066.285.112.447.138.162.026.309.039.44.039.166 0 .346-.022.539-.065.193-.044.374-.119.545-.224.171-.105.314-.239.427-.401.114-.162.171-.361.171-.598v-1.117zM10.756 5.308h1.498v3.47h.026c.106-.167.233-.316.382-.447.149-.132.313-.243.492-.335a2.62 2.62 0 01.566-.211c.197-.048.392-.072.584-.072.535 0 1.002.094 1.4.283.399.188.73.446.993.775.262.329.46.712.591 1.15.132.438.197.907.197 1.406 0 .456-.059.899-.177 1.328a3.475 3.475 0 01-.539 1.137 2.713 2.713 0 01-.914.789c-.368.197-.801.295-1.301.295-.228 0-.458-.015-.69-.046a2.614 2.614 0 01-.664-.177c-.21-.088-.4-.202-.571-.342a1.645 1.645 0 01-.427-.552h-.027v.933h-1.419V5.308zm5.231 5.993c0-.306-.04-.604-.118-.894a2.44 2.44 0 00-.355-.768 1.818 1.818 0 00-.592-.539 1.604 1.604 0 00-.815-.204c-.631 0-1.106.219-1.426.657-.32.438-.479 1.021-.479 1.748 0 .342.041.66.124.953.084.294.209.546.375.756.166.21.366.377.598.499.232.123.502.184.808.184.342 0 .631-.07.868-.21.236-.14.431-.322.585-.545a2.21 2.21 0 00.328-.763 3.86 3.86 0 00.099-.874z"></path>
-</g>
-</svg>;
-icons.uppercase = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M4.789 5.308h1.748l3.614 9.384H8.39l-.881-2.484H3.763l-.88 2.484H1.187l3.602-9.384zm-.579 5.651h2.866L5.669 6.924H5.63l-1.42 4.035zM12.688 9.238h2.681c.394 0 .723-.112.986-.335.263-.224.394-.546.394-.967 0-.473-.118-.806-.355-.998-.236-.193-.578-.29-1.025-.29h-2.681v2.59zm-1.643-3.93h4.561c.841 0 1.516.193 2.024.578.508.386.762.968.762 1.748 0 .473-.116.879-.348 1.216-.232.337-.563.598-.993.782v.026c.579.123 1.017.397 1.315.822.298.425.447.957.447 1.597 0 .368-.066.712-.197 1.031a2.11 2.11 0 01-.618.828c-.281.233-.64.417-1.078.553-.438.135-.959.203-1.564.203h-4.311V5.308zm1.643 8.044h2.905c.499 0 .887-.13 1.163-.388.276-.259.414-.624.414-1.098 0-.464-.138-.821-.414-1.071-.276-.25-.664-.374-1.163-.374h-2.905v2.931z"></path>
-</g>
-</svg>;
-icons.capitalize = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M5.393 5.216h1.748l3.615 9.384H8.995l-.881-2.484H4.368L3.487 14.6H1.792l3.601-9.384zm-.578 5.651H7.68L6.274 6.832h-.04l-1.419 4.035zM11.479 5.216h1.498v3.47h.026c.105-.167.232-.316.381-.447a2.38 2.38 0 01.493-.335 2.62 2.62 0 01.566-.211c.197-.048.392-.072.584-.072.535 0 1.001.094 1.4.283.399.188.73.446.993.775.262.329.46.712.591 1.15.131.438.197.907.197 1.406 0 .456-.059.899-.177 1.328a3.475 3.475 0 01-.539 1.137 2.713 2.713 0 01-.914.789c-.368.197-.801.295-1.301.295-.228 0-.458-.015-.69-.046a2.614 2.614 0 01-.664-.177 2.278 2.278 0 01-.571-.342 1.66 1.66 0 01-.428-.552h-.026v.933h-1.419V5.216zm5.231 5.993c0-.306-.04-.604-.119-.894a2.416 2.416 0 00-.354-.768 1.818 1.818 0 00-.592-.539 1.604 1.604 0 00-.815-.204c-.631 0-1.106.219-1.426.657-.32.438-.48 1.021-.48 1.748 0 .342.042.66.125.953.084.294.208.546.375.756.166.21.366.377.598.499.232.123.502.184.808.184.342 0 .631-.07.868-.21.236-.14.431-.322.585-.545.153-.224.262-.478.328-.763a3.86 3.86 0 00.099-.874z"></path>
-</g>
-</svg>;
-icons.dynamic = <svg
-xmlns="http://www.w3.org/2000/svg"
-width="24"
-height="24"
-viewBox="0 0 24 24"
->
-<path d="M4 5.002s.003-.095.213-.288c.245-.225.671-.483 1.306-.73C7.018 3.399 9.34 3 12 3s4.982.399 6.482.984c.634.247 1.061.505 1.306.73.205.189.212.281.212.288 0 .003-.007.095-.213.284-.245.225-.671.483-1.306.73C16.982 6.601 14.66 7 12 7s-4.982-.399-6.482-.984c-.634-.247-1.061-.505-1.306-.73C4.004 5.094 4 5.002 4 5.002zm16 9.53v4.471a.779.779 0 01-.217.291c-.245.225-.671.482-1.303.728-1.495.582-3.809.978-6.48.978s-4.985-.396-6.48-.978c-.633-.246-1.058-.503-1.303-.728a.865.865 0 01-.199-.242l-.006-4.514c.248.126.51.242.782.348C6.591 15.585 9.171 16 12 16s5.409-.415 7.206-1.114c.277-.108.543-.225.794-.354zm0-7.005v4.476a.779.779 0 01-.217.291c-.245.225-.671.482-1.303.728-1.495.582-3.809.978-6.48.978s-4.985-.396-6.48-.978c-.633-.246-1.058-.503-1.303-.728a.865.865 0 01-.199-.242 1.109 1.109 0 00-.009-.117l-.005-4.407c.248.128.513.244.788.352C6.593 8.582 9.18 9 12 9s5.407-.418 7.208-1.12A8.69 8.69 0 0020 7.527zM2 5v14c0 .058.002.116.007.174.057.665.425 1.197.857 1.594.498.457 1.175.824 1.93 1.118C6.591 22.585 9.171 23 12 23s5.409-.415 7.206-1.114c.755-.294 1.432-.661 1.93-1.118.432-.397.8-.929.857-1.594.005-.058.007-.116.007-.174V5c0-.056-.002-.112-.007-.168-.055-.664-.422-1.195-.852-1.59-.498-.459-1.177-.827-1.933-1.122C17.407 1.418 14.82 1 12 1s-5.407.418-7.208 1.12c-.756.295-1.435.664-1.933 1.122-.43.395-.797.927-.852 1.59A1.887 1.887 0 002 5z"></path>
-</svg>;
-icons.link = <svg
-xmlns="http://www.w3.org/2000/svg"
-width="24"
-height="24"
-viewBox="0 0 24 24"
->
-<path d="M15 8h3c1.105 0 2.103.447 2.828 1.172S22 10.895 22 12s-.447 2.103-1.172 2.828S19.105 16 18 16h-3a1 1 0 000 2h3c1.657 0 3.158-.673 4.243-1.757S24 13.657 24 12s-.673-3.158-1.757-4.243S19.657 6 18 6h-3a1 1 0 000 2zm-6 8H6c-1.105 0-2.103-.447-2.828-1.172S2 13.105 2 12s.447-2.103 1.172-2.828S4.895 8 6 8h3a1 1 0 000-2H6c-1.657 0-3.158.673-4.243 1.757S0 10.343 0 12s.673 3.158 1.757 4.243S4.343 18 6 18h3a1 1 0 000-2zm-1-3h8a1 1 0 000-2H8a1 1 0 000 2z"></path>
-</svg>;
-export default icons;
