+++ src/packages/components/src/index.js
@@ -21,7 +21,11 @@ export { default as ImageSizeControl } from './common/image-size-controls';
 export { default as VerticalAlignmentIcon } from './common/vertical-align-icon';
 export { default as KadenceImageControl } from './common/image-controls';
 export { default as KadencePanelBody } from './panel-body';
+export { default as KadenceSelectTerms } from './select-terms-control';
+export { default as SmallResponsiveControl } from './small-responsive-control';
+export { default as ResponsiveControl } from './responsive-control';
 
+/* Links */
 export { default as DynamicLinkControl } from './links/dynamic-link-control';
 export { default as fetchSearchResults } from './links/get-post-search-results';
 export { default as KadenceImageURLInputUI } from './links/image-url-input-link-controls';
@@ -30,7 +34,21 @@ export { default as InputSearch } from './links/input-search';
 export { default as URLInputControl } from './links/link-control';
 export { default as TextHighlight } from './links/text-highlight';
 
+/* Icons */
 export { default as GenIcon } from './icons/gen-icon';
 export { default as IconControl } from './icons/icon-control';
 export { default as IconRender } from './icons/icon-render';
 
+/* Typography */
+export { default as TypographyControls } from './typography/typography-control';
+export { default as InlineTypographyControls } from './typography/inline-typography-control';
+
+/* Range */
+export { default as KadenceRange } from './range/range-control';
+export { default as ResponsiveRangeControls } from './range/responsive-range-control';
+export { default as ResponsiveSingleRangeControl } from './range/single-range-control';
+
+/* Measurement */
+export { default as ResponsiveMeasurementControls } from './measurement/responsive-measurement-control';
+export { default as MeasurementControls } from './measurement/measurement-control';
+export { default as MeasurementSingleControl } from './measurement/single-input-control';
